@@ -1,4 +1,18 @@
-## Copyright (C) 2018-2025 Advanced Micro Devices, Inc. All rights reserved. ##
+## Copyright (C) 2018-2026 Advanced Micro Devices, Inc. All rights reserved. ##
+
+# -*- coding: utf-8 -*-
+#
+# GPUPerfAPI documentation build configuration file, created by
+# sphinx-quickstart on Wed Jan 10 14:37:31 2018.
+#
+# This file is execfile()d with the current directory set to its
+# containing dir.
+#
+# Note that not all possible configuration values are present in this
+# autogenerated file.
+#
+# All configuration values have a default; values that are commented out
+# serve to show the default.
 
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
@@ -44,9 +58,9 @@ author = u'AMD Developer Tools'
 # built documents.
 #
 # The short X.Y version.
-version = u'4.2'
+version = u'4.3'
 # The full version, including alpha/beta/rc tags.
-release = u'4.2'
+release = u'4.3'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
