@@ -1211,444 +1211,446 @@ namespace counter_dx11_gfx11
         GpaCounterGroupDesc{424, "CPG", 0, 92, 2, 4},
         GpaCounterGroupDesc{425, "CPC", 0, 56, 2, 4},
         GpaCounterGroupDesc{426, "GCVML2", 0, 91, 8, 8},
-        GpaCounterGroupDesc{427, "RPB", 0, 64, 4, 0},
-        GpaCounterGroupDesc{428, "GCEA0", 0, 98, 2, 4},
-        GpaCounterGroupDesc{429, "GCEA1", 1, 98, 2, 4},
-        GpaCounterGroupDesc{430, "GCEA2", 2, 98, 2, 4},
-        GpaCounterGroupDesc{431, "GCEA3", 3, 98, 2, 4},
-        GpaCounterGroupDesc{432, "GCEA4", 4, 98, 2, 4},
-        GpaCounterGroupDesc{433, "GCEA5", 5, 98, 2, 4},
-        GpaCounterGroupDesc{434, "GCEA6", 6, 98, 2, 4},
-        GpaCounterGroupDesc{435, "GCEA7", 7, 98, 2, 4},
-        GpaCounterGroupDesc{436, "GCEA8", 8, 98, 2, 4},
-        GpaCounterGroupDesc{437, "GCEA9", 9, 98, 2, 4},
-        GpaCounterGroupDesc{438, "GCEA10", 10, 98, 2, 4},
-        GpaCounterGroupDesc{439, "GCEA11", 11, 98, 2, 4},
-        GpaCounterGroupDesc{440, "GCEA12", 12, 98, 2, 4},
-        GpaCounterGroupDesc{441, "GCEA13", 13, 98, 2, 4},
-        GpaCounterGroupDesc{442, "GCEA14", 14, 98, 2, 4},
-        GpaCounterGroupDesc{443, "GCEA15", 15, 98, 2, 4},
-        GpaCounterGroupDesc{444, "GCEA16", 16, 98, 2, 4},
-        GpaCounterGroupDesc{445, "GCEA17", 17, 98, 2, 4},
-        GpaCounterGroupDesc{446, "GCEA18", 18, 98, 2, 4},
-        GpaCounterGroupDesc{447, "GCEA19", 19, 98, 2, 4},
-        GpaCounterGroupDesc{448, "GCEA20", 20, 98, 2, 4},
-        GpaCounterGroupDesc{449, "GCEA21", 21, 98, 2, 4},
-        GpaCounterGroupDesc{450, "GCEA22", 22, 98, 2, 4},
-        GpaCounterGroupDesc{451, "GCEA23", 23, 98, 2, 4},
-        GpaCounterGroupDesc{452, "RMI0", 0, 139, 2, 4},
-        GpaCounterGroupDesc{453, "RMI1", 1, 139, 2, 4},
-        GpaCounterGroupDesc{454, "RMI2", 2, 139, 2, 4},
-        GpaCounterGroupDesc{455, "RMI3", 3, 139, 2, 4},
-        GpaCounterGroupDesc{456, "RMI4", 4, 139, 2, 4},
-        GpaCounterGroupDesc{457, "RMI5", 5, 139, 2, 4},
-        GpaCounterGroupDesc{458, "RMI6", 6, 139, 2, 4},
-        GpaCounterGroupDesc{459, "RMI7", 7, 139, 2, 4},
-        GpaCounterGroupDesc{460, "RMI8", 8, 139, 2, 4},
-        GpaCounterGroupDesc{461, "RMI9", 9, 139, 2, 4},
-        GpaCounterGroupDesc{462, "RMI10", 10, 139, 2, 4},
-        GpaCounterGroupDesc{463, "RMI11", 11, 139, 2, 4},
-        GpaCounterGroupDesc{464, "RMI12", 12, 139, 2, 4},
-        GpaCounterGroupDesc{465, "RMI13", 13, 139, 2, 4},
-        GpaCounterGroupDesc{466, "RMI14", 14, 139, 2, 4},
-        GpaCounterGroupDesc{467, "RMI15", 15, 139, 2, 4},
-        GpaCounterGroupDesc{468, "RMI16", 16, 139, 2, 4},
-        GpaCounterGroupDesc{469, "RMI17", 17, 139, 2, 4},
-        GpaCounterGroupDesc{470, "RMI18", 18, 139, 2, 4},
-        GpaCounterGroupDesc{471, "RMI19", 19, 139, 2, 4},
-        GpaCounterGroupDesc{472, "RMI20", 20, 139, 2, 4},
-        GpaCounterGroupDesc{473, "RMI21", 21, 139, 2, 4},
-        GpaCounterGroupDesc{474, "RMI22", 22, 139, 2, 4},
-        GpaCounterGroupDesc{475, "RMI23", 23, 139, 2, 4},
-        GpaCounterGroupDesc{476, "RMI24", 24, 139, 2, 4},
-        GpaCounterGroupDesc{477, "RMI25", 25, 139, 2, 4},
-        GpaCounterGroupDesc{478, "RMI26", 26, 139, 2, 4},
-        GpaCounterGroupDesc{479, "RMI27", 27, 139, 2, 4},
-        GpaCounterGroupDesc{480, "RMI28", 28, 139, 2, 4},
-        GpaCounterGroupDesc{481, "RMI29", 29, 139, 2, 4},
-        GpaCounterGroupDesc{482, "RMI30", 30, 139, 2, 4},
-        GpaCounterGroupDesc{483, "RMI31", 31, 139, 2, 4},
-        GpaCounterGroupDesc{484, "RMI32", 32, 139, 2, 4},
-        GpaCounterGroupDesc{485, "RMI33", 33, 139, 2, 4},
-        GpaCounterGroupDesc{486, "RMI34", 34, 139, 2, 4},
-        GpaCounterGroupDesc{487, "RMI35", 35, 139, 2, 4},
-        GpaCounterGroupDesc{488, "RMI36", 36, 139, 2, 4},
-        GpaCounterGroupDesc{489, "RMI37", 37, 139, 2, 4},
-        GpaCounterGroupDesc{490, "RMI38", 38, 139, 2, 4},
-        GpaCounterGroupDesc{491, "RMI39", 39, 139, 2, 4},
-        GpaCounterGroupDesc{492, "RMI40", 40, 139, 2, 4},
-        GpaCounterGroupDesc{493, "RMI41", 41, 139, 2, 4},
-        GpaCounterGroupDesc{494, "RMI42", 42, 139, 2, 4},
-        GpaCounterGroupDesc{495, "RMI43", 43, 139, 2, 4},
-        GpaCounterGroupDesc{496, "RMI44", 44, 139, 2, 4},
-        GpaCounterGroupDesc{497, "RMI45", 45, 139, 2, 4},
-        GpaCounterGroupDesc{498, "RMI46", 46, 139, 2, 4},
-        GpaCounterGroupDesc{499, "RMI47", 47, 139, 2, 4},
-        GpaCounterGroupDesc{500, "CHA", 0, 40, 4, 4},
-        GpaCounterGroupDesc{501, "CHC0", 0, 44, 4, 4},
-        GpaCounterGroupDesc{502, "CHC1", 1, 44, 4, 4},
-        GpaCounterGroupDesc{503, "CHC2", 2, 44, 4, 4},
-        GpaCounterGroupDesc{504, "CHC3", 3, 44, 4, 4},
-        GpaCounterGroupDesc{505, "CHCG", 0, 44, 4, 4},
-        GpaCounterGroupDesc{506, "GCR", 0, 155, 2, 4},
-        GpaCounterGroupDesc{507, "GE", 0, 40, 4, 16},
-        GpaCounterGroupDesc{508, "GL1A0", 0, 24, 4, 4},
-        GpaCounterGroupDesc{509, "GL1A1", 1, 24, 4, 4},
-        GpaCounterGroupDesc{510, "GL1A2", 2, 24, 4, 4},
-        GpaCounterGroupDesc{511, "GL1A3", 3, 24, 4, 4},
-        GpaCounterGroupDesc{512, "GL1A4", 4, 24, 4, 4},
-        GpaCounterGroupDesc{513, "GL1A5", 5, 24, 4, 4},
-        GpaCounterGroupDesc{514, "GL1A6", 6, 24, 4, 4},
-        GpaCounterGroupDesc{515, "GL1A7", 7, 24, 4, 4},
-        GpaCounterGroupDesc{516, "GL1A8", 8, 24, 4, 4},
-        GpaCounterGroupDesc{517, "GL1A9", 9, 24, 4, 4},
-        GpaCounterGroupDesc{518, "GL1A10", 10, 24, 4, 4},
-        GpaCounterGroupDesc{519, "GL1A11", 11, 24, 4, 4},
-        GpaCounterGroupDesc{520, "GL1C0", 0, 84, 4, 4},
-        GpaCounterGroupDesc{521, "GL1C1", 1, 84, 4, 4},
-        GpaCounterGroupDesc{522, "GL1C2", 2, 84, 4, 4},
-        GpaCounterGroupDesc{523, "GL1C3", 3, 84, 4, 4},
-        GpaCounterGroupDesc{524, "GL1C4", 4, 84, 4, 4},
-        GpaCounterGroupDesc{525, "GL1C5", 5, 84, 4, 4},
-        GpaCounterGroupDesc{526, "GL1C6", 6, 84, 4, 4},
-        GpaCounterGroupDesc{527, "GL1C7", 7, 84, 4, 4},
-        GpaCounterGroupDesc{528, "GL1C8", 8, 84, 4, 4},
-        GpaCounterGroupDesc{529, "GL1C9", 9, 84, 4, 4},
-        GpaCounterGroupDesc{530, "GL1C10", 10, 84, 4, 4},
-        GpaCounterGroupDesc{531, "GL1C11", 11, 84, 4, 4},
-        GpaCounterGroupDesc{532, "GL1C12", 12, 84, 4, 4},
-        GpaCounterGroupDesc{533, "GL1C13", 13, 84, 4, 4},
-        GpaCounterGroupDesc{534, "GL1C14", 14, 84, 4, 4},
-        GpaCounterGroupDesc{535, "GL1C15", 15, 84, 4, 4},
-        GpaCounterGroupDesc{536, "GL1C16", 16, 84, 4, 4},
-        GpaCounterGroupDesc{537, "GL1C17", 17, 84, 4, 4},
-        GpaCounterGroupDesc{538, "GL1C18", 18, 84, 4, 4},
-        GpaCounterGroupDesc{539, "GL1C19", 19, 84, 4, 4},
-        GpaCounterGroupDesc{540, "GL1C20", 20, 84, 4, 4},
-        GpaCounterGroupDesc{541, "GL1C21", 21, 84, 4, 4},
-        GpaCounterGroupDesc{542, "GL1C22", 22, 84, 4, 4},
-        GpaCounterGroupDesc{543, "GL1C23", 23, 84, 4, 4},
-        GpaCounterGroupDesc{544, "GL1C24", 24, 84, 4, 4},
-        GpaCounterGroupDesc{545, "GL1C25", 25, 84, 4, 4},
-        GpaCounterGroupDesc{546, "GL1C26", 26, 84, 4, 4},
-        GpaCounterGroupDesc{547, "GL1C27", 27, 84, 4, 4},
-        GpaCounterGroupDesc{548, "GL1C28", 28, 84, 4, 4},
-        GpaCounterGroupDesc{549, "GL1C29", 29, 84, 4, 4},
-        GpaCounterGroupDesc{550, "GL1C30", 30, 84, 4, 4},
-        GpaCounterGroupDesc{551, "GL1C31", 31, 84, 4, 4},
-        GpaCounterGroupDesc{552, "GL1C32", 32, 84, 4, 4},
-        GpaCounterGroupDesc{553, "GL1C33", 33, 84, 4, 4},
-        GpaCounterGroupDesc{554, "GL1C34", 34, 84, 4, 4},
-        GpaCounterGroupDesc{555, "GL1C35", 35, 84, 4, 4},
-        GpaCounterGroupDesc{556, "GL1C36", 36, 84, 4, 4},
-        GpaCounterGroupDesc{557, "GL1C37", 37, 84, 4, 4},
-        GpaCounterGroupDesc{558, "GL1C38", 38, 84, 4, 4},
-        GpaCounterGroupDesc{559, "GL1C39", 39, 84, 4, 4},
-        GpaCounterGroupDesc{560, "GL1C40", 40, 84, 4, 4},
-        GpaCounterGroupDesc{561, "GL1C41", 41, 84, 4, 4},
-        GpaCounterGroupDesc{562, "GL1C42", 42, 84, 4, 4},
-        GpaCounterGroupDesc{563, "GL1C43", 43, 84, 4, 4},
-        GpaCounterGroupDesc{564, "GL1C44", 44, 84, 4, 4},
-        GpaCounterGroupDesc{565, "GL1C45", 45, 84, 4, 4},
-        GpaCounterGroupDesc{566, "GL1C46", 46, 84, 4, 4},
-        GpaCounterGroupDesc{567, "GL1C47", 47, 84, 4, 4},
-        GpaCounterGroupDesc{568, "GL2A0", 0, 108, 4, 8},
-        GpaCounterGroupDesc{569, "GL2A1", 1, 108, 4, 8},
-        GpaCounterGroupDesc{570, "GL2A2", 2, 108, 4, 8},
-        GpaCounterGroupDesc{571, "GL2A3", 3, 108, 4, 8},
-        GpaCounterGroupDesc{572, "GL2C0", 0, 259, 4, 8},
-        GpaCounterGroupDesc{573, "GL2C1", 1, 259, 4, 8},
-        GpaCounterGroupDesc{574, "GL2C2", 2, 259, 4, 8},
-        GpaCounterGroupDesc{575, "GL2C3", 3, 259, 4, 8},
-        GpaCounterGroupDesc{576, "GL2C4", 4, 259, 4, 8},
-        GpaCounterGroupDesc{577, "GL2C5", 5, 259, 4, 8},
-        GpaCounterGroupDesc{578, "GL2C6", 6, 259, 4, 8},
-        GpaCounterGroupDesc{579, "GL2C7", 7, 259, 4, 8},
-        GpaCounterGroupDesc{580, "GL2C8", 8, 259, 4, 8},
-        GpaCounterGroupDesc{581, "GL2C9", 9, 259, 4, 8},
-        GpaCounterGroupDesc{582, "GL2C10", 10, 259, 4, 8},
-        GpaCounterGroupDesc{583, "GL2C11", 11, 259, 4, 8},
-        GpaCounterGroupDesc{584, "GL2C12", 12, 259, 4, 8},
-        GpaCounterGroupDesc{585, "GL2C13", 13, 259, 4, 8},
-        GpaCounterGroupDesc{586, "GL2C14", 14, 259, 4, 8},
-        GpaCounterGroupDesc{587, "GL2C15", 15, 259, 4, 8},
-        GpaCounterGroupDesc{588, "GL2C16", 16, 259, 4, 8},
-        GpaCounterGroupDesc{589, "GL2C17", 17, 259, 4, 8},
-        GpaCounterGroupDesc{590, "GL2C18", 18, 259, 4, 8},
-        GpaCounterGroupDesc{591, "GL2C19", 19, 259, 4, 8},
-        GpaCounterGroupDesc{592, "GL2C20", 20, 259, 4, 8},
-        GpaCounterGroupDesc{593, "GL2C21", 21, 259, 4, 8},
-        GpaCounterGroupDesc{594, "GL2C22", 22, 259, 4, 8},
-        GpaCounterGroupDesc{595, "GL2C23", 23, 259, 4, 8},
-        GpaCounterGroupDesc{596, "GUS", 0, 176, 2, 4},
-        GpaCounterGroupDesc{597, "PA_PH", 0, 1024, 8, 16},
-        GpaCounterGroupDesc{598, "UTCL10", 0, 66, 2, 0},
-        GpaCounterGroupDesc{599, "UTCL11", 1, 66, 2, 0},
-        GpaCounterGroupDesc{600, "UTCL12", 2, 66, 2, 0},
-        GpaCounterGroupDesc{601, "UTCL13", 3, 66, 2, 0},
-        GpaCounterGroupDesc{602, "UTCL14", 4, 66, 2, 0},
-        GpaCounterGroupDesc{603, "UTCL15", 5, 66, 2, 0},
-        GpaCounterGroupDesc{604, "UTCL16", 6, 66, 2, 0},
-        GpaCounterGroupDesc{605, "UTCL17", 7, 66, 2, 0},
-        GpaCounterGroupDesc{606, "UTCL18", 8, 66, 2, 0},
-        GpaCounterGroupDesc{607, "UTCL19", 9, 66, 2, 0},
-        GpaCounterGroupDesc{608, "UTCL110", 10, 66, 2, 0},
-        GpaCounterGroupDesc{609, "UTCL111", 11, 66, 2, 0},
-        GpaCounterGroupDesc{610, "GE2_DIST", 0, 101, 4, 16},
-        GpaCounterGroupDesc{611, "GE_SE0", 0, 88, 4, 16},
-        GpaCounterGroupDesc{612, "GE_SE1", 1, 88, 4, 16},
-        GpaCounterGroupDesc{613, "GE_SE2", 2, 88, 4, 16},
-        GpaCounterGroupDesc{614, "GE_SE3", 3, 88, 4, 16},
-        GpaCounterGroupDesc{615, "GE_SE4", 4, 88, 4, 16},
-        GpaCounterGroupDesc{616, "GE_SE5", 5, 88, 4, 16},
-        GpaCounterGroupDesc{617, "SQWGP0", 0, 512, 8, 16},
-        GpaCounterGroupDesc{618, "SQWGP1", 1, 512, 8, 16},
-        GpaCounterGroupDesc{619, "SQWGP2", 2, 512, 8, 16},
-        GpaCounterGroupDesc{620, "SQWGP3", 3, 512, 8, 16},
-        GpaCounterGroupDesc{621, "SQWGP4", 4, 512, 8, 16},
-        GpaCounterGroupDesc{622, "SQWGP5", 5, 512, 8, 16},
-        GpaCounterGroupDesc{623, "SQWGP6", 6, 512, 8, 16},
-        GpaCounterGroupDesc{624, "SQWGP7", 7, 512, 8, 16},
-        GpaCounterGroupDesc{625, "SQWGP8", 8, 512, 8, 16},
-        GpaCounterGroupDesc{626, "SQWGP9", 9, 512, 8, 16},
-        GpaCounterGroupDesc{627, "SQWGP10", 10, 512, 8, 16},
-        GpaCounterGroupDesc{628, "SQWGP11", 11, 512, 8, 16},
-        GpaCounterGroupDesc{629, "SQWGP12", 12, 512, 8, 16},
-        GpaCounterGroupDesc{630, "SQWGP13", 13, 512, 8, 16},
-        GpaCounterGroupDesc{631, "SQWGP14", 14, 512, 8, 16},
-        GpaCounterGroupDesc{632, "SQWGP15", 15, 512, 8, 16},
-        GpaCounterGroupDesc{633, "SQWGP16", 16, 512, 8, 16},
-        GpaCounterGroupDesc{634, "SQWGP17", 17, 512, 8, 16},
-        GpaCounterGroupDesc{635, "SQWGP18", 18, 512, 8, 16},
-        GpaCounterGroupDesc{636, "SQWGP19", 19, 512, 8, 16},
-        GpaCounterGroupDesc{637, "SQWGP20", 20, 512, 8, 16},
-        GpaCounterGroupDesc{638, "SQWGP21", 21, 512, 8, 16},
-        GpaCounterGroupDesc{639, "SQWGP22", 22, 512, 8, 16},
-        GpaCounterGroupDesc{640, "SQWGP23", 23, 512, 8, 16},
-        GpaCounterGroupDesc{641, "SQWGP24", 24, 512, 8, 16},
-        GpaCounterGroupDesc{642, "SQWGP25", 25, 512, 8, 16},
-        GpaCounterGroupDesc{643, "SQWGP26", 26, 512, 8, 16},
-        GpaCounterGroupDesc{644, "SQWGP27", 27, 512, 8, 16},
-        GpaCounterGroupDesc{645, "SQWGP28", 28, 512, 8, 16},
-        GpaCounterGroupDesc{646, "SQWGP29", 29, 512, 8, 16},
-        GpaCounterGroupDesc{647, "SQWGP30", 30, 512, 8, 16},
-        GpaCounterGroupDesc{648, "SQWGP31", 31, 512, 8, 16},
-        GpaCounterGroupDesc{649, "SQWGP32", 32, 512, 8, 16},
-        GpaCounterGroupDesc{650, "SQWGP33", 33, 512, 8, 16},
-        GpaCounterGroupDesc{651, "SQWGP34", 34, 512, 8, 16},
-        GpaCounterGroupDesc{652, "SQWGP35", 35, 512, 8, 16},
-        GpaCounterGroupDesc{653, "SQWGP36", 36, 512, 8, 16},
-        GpaCounterGroupDesc{654, "SQWGP37", 37, 512, 8, 16},
-        GpaCounterGroupDesc{655, "SQWGP38", 38, 512, 8, 16},
-        GpaCounterGroupDesc{656, "SQWGP39", 39, 512, 8, 16},
-        GpaCounterGroupDesc{657, "SQWGP40", 40, 512, 8, 16},
-        GpaCounterGroupDesc{658, "SQWGP41", 41, 512, 8, 16},
-        GpaCounterGroupDesc{659, "SQWGP42", 42, 512, 8, 16},
-        GpaCounterGroupDesc{660, "SQWGP43", 43, 512, 8, 16},
-        GpaCounterGroupDesc{661, "SQWGP44", 44, 512, 8, 16},
-        GpaCounterGroupDesc{662, "SQWGP45", 45, 512, 8, 16},
-        GpaCounterGroupDesc{663, "SQWGP46", 46, 512, 8, 16},
-        GpaCounterGroupDesc{664, "SQWGP47", 47, 512, 8, 16},
-        GpaCounterGroupDesc{665, "SQWGP_GS0", 0, 512, 8, 16},
-        GpaCounterGroupDesc{666, "SQWGP_GS1", 1, 512, 8, 16},
-        GpaCounterGroupDesc{667, "SQWGP_GS2", 2, 512, 8, 16},
-        GpaCounterGroupDesc{668, "SQWGP_GS3", 3, 512, 8, 16},
-        GpaCounterGroupDesc{669, "SQWGP_GS4", 4, 512, 8, 16},
-        GpaCounterGroupDesc{670, "SQWGP_GS5", 5, 512, 8, 16},
-        GpaCounterGroupDesc{671, "SQWGP_GS6", 6, 512, 8, 16},
-        GpaCounterGroupDesc{672, "SQWGP_GS7", 7, 512, 8, 16},
-        GpaCounterGroupDesc{673, "SQWGP_GS8", 8, 512, 8, 16},
-        GpaCounterGroupDesc{674, "SQWGP_GS9", 9, 512, 8, 16},
-        GpaCounterGroupDesc{675, "SQWGP_GS10", 10, 512, 8, 16},
-        GpaCounterGroupDesc{676, "SQWGP_GS11", 11, 512, 8, 16},
-        GpaCounterGroupDesc{677, "SQWGP_GS12", 12, 512, 8, 16},
-        GpaCounterGroupDesc{678, "SQWGP_GS13", 13, 512, 8, 16},
-        GpaCounterGroupDesc{679, "SQWGP_GS14", 14, 512, 8, 16},
-        GpaCounterGroupDesc{680, "SQWGP_GS15", 15, 512, 8, 16},
-        GpaCounterGroupDesc{681, "SQWGP_GS16", 16, 512, 8, 16},
-        GpaCounterGroupDesc{682, "SQWGP_GS17", 17, 512, 8, 16},
-        GpaCounterGroupDesc{683, "SQWGP_GS18", 18, 512, 8, 16},
-        GpaCounterGroupDesc{684, "SQWGP_GS19", 19, 512, 8, 16},
-        GpaCounterGroupDesc{685, "SQWGP_GS20", 20, 512, 8, 16},
-        GpaCounterGroupDesc{686, "SQWGP_GS21", 21, 512, 8, 16},
-        GpaCounterGroupDesc{687, "SQWGP_GS22", 22, 512, 8, 16},
-        GpaCounterGroupDesc{688, "SQWGP_GS23", 23, 512, 8, 16},
-        GpaCounterGroupDesc{689, "SQWGP_GS24", 24, 512, 8, 16},
-        GpaCounterGroupDesc{690, "SQWGP_GS25", 25, 512, 8, 16},
-        GpaCounterGroupDesc{691, "SQWGP_GS26", 26, 512, 8, 16},
-        GpaCounterGroupDesc{692, "SQWGP_GS27", 27, 512, 8, 16},
-        GpaCounterGroupDesc{693, "SQWGP_GS28", 28, 512, 8, 16},
-        GpaCounterGroupDesc{694, "SQWGP_GS29", 29, 512, 8, 16},
-        GpaCounterGroupDesc{695, "SQWGP_GS30", 30, 512, 8, 16},
-        GpaCounterGroupDesc{696, "SQWGP_GS31", 31, 512, 8, 16},
-        GpaCounterGroupDesc{697, "SQWGP_GS32", 32, 512, 8, 16},
-        GpaCounterGroupDesc{698, "SQWGP_GS33", 33, 512, 8, 16},
-        GpaCounterGroupDesc{699, "SQWGP_GS34", 34, 512, 8, 16},
-        GpaCounterGroupDesc{700, "SQWGP_GS35", 35, 512, 8, 16},
-        GpaCounterGroupDesc{701, "SQWGP_GS36", 36, 512, 8, 16},
-        GpaCounterGroupDesc{702, "SQWGP_GS37", 37, 512, 8, 16},
-        GpaCounterGroupDesc{703, "SQWGP_GS38", 38, 512, 8, 16},
-        GpaCounterGroupDesc{704, "SQWGP_GS39", 39, 512, 8, 16},
-        GpaCounterGroupDesc{705, "SQWGP_GS40", 40, 512, 8, 16},
-        GpaCounterGroupDesc{706, "SQWGP_GS41", 41, 512, 8, 16},
-        GpaCounterGroupDesc{707, "SQWGP_GS42", 42, 512, 8, 16},
-        GpaCounterGroupDesc{708, "SQWGP_GS43", 43, 512, 8, 16},
-        GpaCounterGroupDesc{709, "SQWGP_GS44", 44, 512, 8, 16},
-        GpaCounterGroupDesc{710, "SQWGP_GS45", 45, 512, 8, 16},
-        GpaCounterGroupDesc{711, "SQWGP_GS46", 46, 512, 8, 16},
-        GpaCounterGroupDesc{712, "SQWGP_GS47", 47, 512, 8, 16},
-        GpaCounterGroupDesc{713, "SQWGP_PS0", 0, 512, 8, 16},
-        GpaCounterGroupDesc{714, "SQWGP_PS1", 1, 512, 8, 16},
-        GpaCounterGroupDesc{715, "SQWGP_PS2", 2, 512, 8, 16},
-        GpaCounterGroupDesc{716, "SQWGP_PS3", 3, 512, 8, 16},
-        GpaCounterGroupDesc{717, "SQWGP_PS4", 4, 512, 8, 16},
-        GpaCounterGroupDesc{718, "SQWGP_PS5", 5, 512, 8, 16},
-        GpaCounterGroupDesc{719, "SQWGP_PS6", 6, 512, 8, 16},
-        GpaCounterGroupDesc{720, "SQWGP_PS7", 7, 512, 8, 16},
-        GpaCounterGroupDesc{721, "SQWGP_PS8", 8, 512, 8, 16},
-        GpaCounterGroupDesc{722, "SQWGP_PS9", 9, 512, 8, 16},
-        GpaCounterGroupDesc{723, "SQWGP_PS10", 10, 512, 8, 16},
-        GpaCounterGroupDesc{724, "SQWGP_PS11", 11, 512, 8, 16},
-        GpaCounterGroupDesc{725, "SQWGP_PS12", 12, 512, 8, 16},
-        GpaCounterGroupDesc{726, "SQWGP_PS13", 13, 512, 8, 16},
-        GpaCounterGroupDesc{727, "SQWGP_PS14", 14, 512, 8, 16},
-        GpaCounterGroupDesc{728, "SQWGP_PS15", 15, 512, 8, 16},
-        GpaCounterGroupDesc{729, "SQWGP_PS16", 16, 512, 8, 16},
-        GpaCounterGroupDesc{730, "SQWGP_PS17", 17, 512, 8, 16},
-        GpaCounterGroupDesc{731, "SQWGP_PS18", 18, 512, 8, 16},
-        GpaCounterGroupDesc{732, "SQWGP_PS19", 19, 512, 8, 16},
-        GpaCounterGroupDesc{733, "SQWGP_PS20", 20, 512, 8, 16},
-        GpaCounterGroupDesc{734, "SQWGP_PS21", 21, 512, 8, 16},
-        GpaCounterGroupDesc{735, "SQWGP_PS22", 22, 512, 8, 16},
-        GpaCounterGroupDesc{736, "SQWGP_PS23", 23, 512, 8, 16},
-        GpaCounterGroupDesc{737, "SQWGP_PS24", 24, 512, 8, 16},
-        GpaCounterGroupDesc{738, "SQWGP_PS25", 25, 512, 8, 16},
-        GpaCounterGroupDesc{739, "SQWGP_PS26", 26, 512, 8, 16},
-        GpaCounterGroupDesc{740, "SQWGP_PS27", 27, 512, 8, 16},
-        GpaCounterGroupDesc{741, "SQWGP_PS28", 28, 512, 8, 16},
-        GpaCounterGroupDesc{742, "SQWGP_PS29", 29, 512, 8, 16},
-        GpaCounterGroupDesc{743, "SQWGP_PS30", 30, 512, 8, 16},
-        GpaCounterGroupDesc{744, "SQWGP_PS31", 31, 512, 8, 16},
-        GpaCounterGroupDesc{745, "SQWGP_PS32", 32, 512, 8, 16},
-        GpaCounterGroupDesc{746, "SQWGP_PS33", 33, 512, 8, 16},
-        GpaCounterGroupDesc{747, "SQWGP_PS34", 34, 512, 8, 16},
-        GpaCounterGroupDesc{748, "SQWGP_PS35", 35, 512, 8, 16},
-        GpaCounterGroupDesc{749, "SQWGP_PS36", 36, 512, 8, 16},
-        GpaCounterGroupDesc{750, "SQWGP_PS37", 37, 512, 8, 16},
-        GpaCounterGroupDesc{751, "SQWGP_PS38", 38, 512, 8, 16},
-        GpaCounterGroupDesc{752, "SQWGP_PS39", 39, 512, 8, 16},
-        GpaCounterGroupDesc{753, "SQWGP_PS40", 40, 512, 8, 16},
-        GpaCounterGroupDesc{754, "SQWGP_PS41", 41, 512, 8, 16},
-        GpaCounterGroupDesc{755, "SQWGP_PS42", 42, 512, 8, 16},
-        GpaCounterGroupDesc{756, "SQWGP_PS43", 43, 512, 8, 16},
-        GpaCounterGroupDesc{757, "SQWGP_PS44", 44, 512, 8, 16},
-        GpaCounterGroupDesc{758, "SQWGP_PS45", 45, 512, 8, 16},
-        GpaCounterGroupDesc{759, "SQWGP_PS46", 46, 512, 8, 16},
-        GpaCounterGroupDesc{760, "SQWGP_PS47", 47, 512, 8, 16},
-        GpaCounterGroupDesc{761, "SQWGP_HS0", 0, 512, 8, 16},
-        GpaCounterGroupDesc{762, "SQWGP_HS1", 1, 512, 8, 16},
-        GpaCounterGroupDesc{763, "SQWGP_HS2", 2, 512, 8, 16},
-        GpaCounterGroupDesc{764, "SQWGP_HS3", 3, 512, 8, 16},
-        GpaCounterGroupDesc{765, "SQWGP_HS4", 4, 512, 8, 16},
-        GpaCounterGroupDesc{766, "SQWGP_HS5", 5, 512, 8, 16},
-        GpaCounterGroupDesc{767, "SQWGP_HS6", 6, 512, 8, 16},
-        GpaCounterGroupDesc{768, "SQWGP_HS7", 7, 512, 8, 16},
-        GpaCounterGroupDesc{769, "SQWGP_HS8", 8, 512, 8, 16},
-        GpaCounterGroupDesc{770, "SQWGP_HS9", 9, 512, 8, 16},
-        GpaCounterGroupDesc{771, "SQWGP_HS10", 10, 512, 8, 16},
-        GpaCounterGroupDesc{772, "SQWGP_HS11", 11, 512, 8, 16},
-        GpaCounterGroupDesc{773, "SQWGP_HS12", 12, 512, 8, 16},
-        GpaCounterGroupDesc{774, "SQWGP_HS13", 13, 512, 8, 16},
-        GpaCounterGroupDesc{775, "SQWGP_HS14", 14, 512, 8, 16},
-        GpaCounterGroupDesc{776, "SQWGP_HS15", 15, 512, 8, 16},
-        GpaCounterGroupDesc{777, "SQWGP_HS16", 16, 512, 8, 16},
-        GpaCounterGroupDesc{778, "SQWGP_HS17", 17, 512, 8, 16},
-        GpaCounterGroupDesc{779, "SQWGP_HS18", 18, 512, 8, 16},
-        GpaCounterGroupDesc{780, "SQWGP_HS19", 19, 512, 8, 16},
-        GpaCounterGroupDesc{781, "SQWGP_HS20", 20, 512, 8, 16},
-        GpaCounterGroupDesc{782, "SQWGP_HS21", 21, 512, 8, 16},
-        GpaCounterGroupDesc{783, "SQWGP_HS22", 22, 512, 8, 16},
-        GpaCounterGroupDesc{784, "SQWGP_HS23", 23, 512, 8, 16},
-        GpaCounterGroupDesc{785, "SQWGP_HS24", 24, 512, 8, 16},
-        GpaCounterGroupDesc{786, "SQWGP_HS25", 25, 512, 8, 16},
-        GpaCounterGroupDesc{787, "SQWGP_HS26", 26, 512, 8, 16},
-        GpaCounterGroupDesc{788, "SQWGP_HS27", 27, 512, 8, 16},
-        GpaCounterGroupDesc{789, "SQWGP_HS28", 28, 512, 8, 16},
-        GpaCounterGroupDesc{790, "SQWGP_HS29", 29, 512, 8, 16},
-        GpaCounterGroupDesc{791, "SQWGP_HS30", 30, 512, 8, 16},
-        GpaCounterGroupDesc{792, "SQWGP_HS31", 31, 512, 8, 16},
-        GpaCounterGroupDesc{793, "SQWGP_HS32", 32, 512, 8, 16},
-        GpaCounterGroupDesc{794, "SQWGP_HS33", 33, 512, 8, 16},
-        GpaCounterGroupDesc{795, "SQWGP_HS34", 34, 512, 8, 16},
-        GpaCounterGroupDesc{796, "SQWGP_HS35", 35, 512, 8, 16},
-        GpaCounterGroupDesc{797, "SQWGP_HS36", 36, 512, 8, 16},
-        GpaCounterGroupDesc{798, "SQWGP_HS37", 37, 512, 8, 16},
-        GpaCounterGroupDesc{799, "SQWGP_HS38", 38, 512, 8, 16},
-        GpaCounterGroupDesc{800, "SQWGP_HS39", 39, 512, 8, 16},
-        GpaCounterGroupDesc{801, "SQWGP_HS40", 40, 512, 8, 16},
-        GpaCounterGroupDesc{802, "SQWGP_HS41", 41, 512, 8, 16},
-        GpaCounterGroupDesc{803, "SQWGP_HS42", 42, 512, 8, 16},
-        GpaCounterGroupDesc{804, "SQWGP_HS43", 43, 512, 8, 16},
-        GpaCounterGroupDesc{805, "SQWGP_HS44", 44, 512, 8, 16},
-        GpaCounterGroupDesc{806, "SQWGP_HS45", 45, 512, 8, 16},
-        GpaCounterGroupDesc{807, "SQWGP_HS46", 46, 512, 8, 16},
-        GpaCounterGroupDesc{808, "SQWGP_HS47", 47, 512, 8, 16},
-        GpaCounterGroupDesc{809, "SQWGP_CS0", 0, 512, 8, 16},
-        GpaCounterGroupDesc{810, "SQWGP_CS1", 1, 512, 8, 16},
-        GpaCounterGroupDesc{811, "SQWGP_CS2", 2, 512, 8, 16},
-        GpaCounterGroupDesc{812, "SQWGP_CS3", 3, 512, 8, 16},
-        GpaCounterGroupDesc{813, "SQWGP_CS4", 4, 512, 8, 16},
-        GpaCounterGroupDesc{814, "SQWGP_CS5", 5, 512, 8, 16},
-        GpaCounterGroupDesc{815, "SQWGP_CS6", 6, 512, 8, 16},
-        GpaCounterGroupDesc{816, "SQWGP_CS7", 7, 512, 8, 16},
-        GpaCounterGroupDesc{817, "SQWGP_CS8", 8, 512, 8, 16},
-        GpaCounterGroupDesc{818, "SQWGP_CS9", 9, 512, 8, 16},
-        GpaCounterGroupDesc{819, "SQWGP_CS10", 10, 512, 8, 16},
-        GpaCounterGroupDesc{820, "SQWGP_CS11", 11, 512, 8, 16},
-        GpaCounterGroupDesc{821, "SQWGP_CS12", 12, 512, 8, 16},
-        GpaCounterGroupDesc{822, "SQWGP_CS13", 13, 512, 8, 16},
-        GpaCounterGroupDesc{823, "SQWGP_CS14", 14, 512, 8, 16},
-        GpaCounterGroupDesc{824, "SQWGP_CS15", 15, 512, 8, 16},
-        GpaCounterGroupDesc{825, "SQWGP_CS16", 16, 512, 8, 16},
-        GpaCounterGroupDesc{826, "SQWGP_CS17", 17, 512, 8, 16},
-        GpaCounterGroupDesc{827, "SQWGP_CS18", 18, 512, 8, 16},
-        GpaCounterGroupDesc{828, "SQWGP_CS19", 19, 512, 8, 16},
-        GpaCounterGroupDesc{829, "SQWGP_CS20", 20, 512, 8, 16},
-        GpaCounterGroupDesc{830, "SQWGP_CS21", 21, 512, 8, 16},
-        GpaCounterGroupDesc{831, "SQWGP_CS22", 22, 512, 8, 16},
-        GpaCounterGroupDesc{832, "SQWGP_CS23", 23, 512, 8, 16},
-        GpaCounterGroupDesc{833, "SQWGP_CS24", 24, 512, 8, 16},
-        GpaCounterGroupDesc{834, "SQWGP_CS25", 25, 512, 8, 16},
-        GpaCounterGroupDesc{835, "SQWGP_CS26", 26, 512, 8, 16},
-        GpaCounterGroupDesc{836, "SQWGP_CS27", 27, 512, 8, 16},
-        GpaCounterGroupDesc{837, "SQWGP_CS28", 28, 512, 8, 16},
-        GpaCounterGroupDesc{838, "SQWGP_CS29", 29, 512, 8, 16},
-        GpaCounterGroupDesc{839, "SQWGP_CS30", 30, 512, 8, 16},
-        GpaCounterGroupDesc{840, "SQWGP_CS31", 31, 512, 8, 16},
-        GpaCounterGroupDesc{841, "SQWGP_CS32", 32, 512, 8, 16},
-        GpaCounterGroupDesc{842, "SQWGP_CS33", 33, 512, 8, 16},
-        GpaCounterGroupDesc{843, "SQWGP_CS34", 34, 512, 8, 16},
-        GpaCounterGroupDesc{844, "SQWGP_CS35", 35, 512, 8, 16},
-        GpaCounterGroupDesc{845, "SQWGP_CS36", 36, 512, 8, 16},
-        GpaCounterGroupDesc{846, "SQWGP_CS37", 37, 512, 8, 16},
-        GpaCounterGroupDesc{847, "SQWGP_CS38", 38, 512, 8, 16},
-        GpaCounterGroupDesc{848, "SQWGP_CS39", 39, 512, 8, 16},
-        GpaCounterGroupDesc{849, "SQWGP_CS40", 40, 512, 8, 16},
-        GpaCounterGroupDesc{850, "SQWGP_CS41", 41, 512, 8, 16},
-        GpaCounterGroupDesc{851, "SQWGP_CS42", 42, 512, 8, 16},
-        GpaCounterGroupDesc{852, "SQWGP_CS43", 43, 512, 8, 16},
-        GpaCounterGroupDesc{853, "SQWGP_CS44", 44, 512, 8, 16},
-        GpaCounterGroupDesc{854, "SQWGP_CS45", 45, 512, 8, 16},
-        GpaCounterGroupDesc{855, "SQWGP_CS46", 46, 512, 8, 16},
-        GpaCounterGroupDesc{856, "SQWGP_CS47", 47, 512, 8, 16},
-        GpaCounterGroupDesc{857, "PC0", 0, 47, 4, 0},
-        GpaCounterGroupDesc{858, "PC1", 1, 47, 4, 0},
-        GpaCounterGroupDesc{859, "PC2", 2, 47, 4, 0},
-        GpaCounterGroupDesc{860, "PC3", 3, 47, 4, 0},
-        GpaCounterGroupDesc{861, "PC4", 4, 47, 4, 0},
-        GpaCounterGroupDesc{862, "PC5", 5, 47, 4, 0},
-        GpaCounterGroupDesc{863, "GPUTimeStamp", 0, 1, 1, 0},
-        GpaCounterGroupDesc{864, "GPUTime", 0, 6, 3, 0},
+        GpaCounterGroupDesc{427, "ATC", 0, 24, 0, 0},
+        GpaCounterGroupDesc{428, "ATCL2", 0, 18, 0, 0},
+        GpaCounterGroupDesc{429, "RPB", 0, 64, 4, 0},
+        GpaCounterGroupDesc{430, "GCEA0", 0, 98, 2, 4},
+        GpaCounterGroupDesc{431, "GCEA1", 1, 98, 2, 4},
+        GpaCounterGroupDesc{432, "GCEA2", 2, 98, 2, 4},
+        GpaCounterGroupDesc{433, "GCEA3", 3, 98, 2, 4},
+        GpaCounterGroupDesc{434, "GCEA4", 4, 98, 2, 4},
+        GpaCounterGroupDesc{435, "GCEA5", 5, 98, 2, 4},
+        GpaCounterGroupDesc{436, "GCEA6", 6, 98, 2, 4},
+        GpaCounterGroupDesc{437, "GCEA7", 7, 98, 2, 4},
+        GpaCounterGroupDesc{438, "GCEA8", 8, 98, 2, 4},
+        GpaCounterGroupDesc{439, "GCEA9", 9, 98, 2, 4},
+        GpaCounterGroupDesc{440, "GCEA10", 10, 98, 2, 4},
+        GpaCounterGroupDesc{441, "GCEA11", 11, 98, 2, 4},
+        GpaCounterGroupDesc{442, "GCEA12", 12, 98, 2, 4},
+        GpaCounterGroupDesc{443, "GCEA13", 13, 98, 2, 4},
+        GpaCounterGroupDesc{444, "GCEA14", 14, 98, 2, 4},
+        GpaCounterGroupDesc{445, "GCEA15", 15, 98, 2, 4},
+        GpaCounterGroupDesc{446, "GCEA16", 16, 98, 2, 4},
+        GpaCounterGroupDesc{447, "GCEA17", 17, 98, 2, 4},
+        GpaCounterGroupDesc{448, "GCEA18", 18, 98, 2, 4},
+        GpaCounterGroupDesc{449, "GCEA19", 19, 98, 2, 4},
+        GpaCounterGroupDesc{450, "GCEA20", 20, 98, 2, 4},
+        GpaCounterGroupDesc{451, "GCEA21", 21, 98, 2, 4},
+        GpaCounterGroupDesc{452, "GCEA22", 22, 98, 2, 4},
+        GpaCounterGroupDesc{453, "GCEA23", 23, 98, 2, 4},
+        GpaCounterGroupDesc{454, "RMI0", 0, 139, 2, 4},
+        GpaCounterGroupDesc{455, "RMI1", 1, 139, 2, 4},
+        GpaCounterGroupDesc{456, "RMI2", 2, 139, 2, 4},
+        GpaCounterGroupDesc{457, "RMI3", 3, 139, 2, 4},
+        GpaCounterGroupDesc{458, "RMI4", 4, 139, 2, 4},
+        GpaCounterGroupDesc{459, "RMI5", 5, 139, 2, 4},
+        GpaCounterGroupDesc{460, "RMI6", 6, 139, 2, 4},
+        GpaCounterGroupDesc{461, "RMI7", 7, 139, 2, 4},
+        GpaCounterGroupDesc{462, "RMI8", 8, 139, 2, 4},
+        GpaCounterGroupDesc{463, "RMI9", 9, 139, 2, 4},
+        GpaCounterGroupDesc{464, "RMI10", 10, 139, 2, 4},
+        GpaCounterGroupDesc{465, "RMI11", 11, 139, 2, 4},
+        GpaCounterGroupDesc{466, "RMI12", 12, 139, 2, 4},
+        GpaCounterGroupDesc{467, "RMI13", 13, 139, 2, 4},
+        GpaCounterGroupDesc{468, "RMI14", 14, 139, 2, 4},
+        GpaCounterGroupDesc{469, "RMI15", 15, 139, 2, 4},
+        GpaCounterGroupDesc{470, "RMI16", 16, 139, 2, 4},
+        GpaCounterGroupDesc{471, "RMI17", 17, 139, 2, 4},
+        GpaCounterGroupDesc{472, "RMI18", 18, 139, 2, 4},
+        GpaCounterGroupDesc{473, "RMI19", 19, 139, 2, 4},
+        GpaCounterGroupDesc{474, "RMI20", 20, 139, 2, 4},
+        GpaCounterGroupDesc{475, "RMI21", 21, 139, 2, 4},
+        GpaCounterGroupDesc{476, "RMI22", 22, 139, 2, 4},
+        GpaCounterGroupDesc{477, "RMI23", 23, 139, 2, 4},
+        GpaCounterGroupDesc{478, "RMI24", 24, 139, 2, 4},
+        GpaCounterGroupDesc{479, "RMI25", 25, 139, 2, 4},
+        GpaCounterGroupDesc{480, "RMI26", 26, 139, 2, 4},
+        GpaCounterGroupDesc{481, "RMI27", 27, 139, 2, 4},
+        GpaCounterGroupDesc{482, "RMI28", 28, 139, 2, 4},
+        GpaCounterGroupDesc{483, "RMI29", 29, 139, 2, 4},
+        GpaCounterGroupDesc{484, "RMI30", 30, 139, 2, 4},
+        GpaCounterGroupDesc{485, "RMI31", 31, 139, 2, 4},
+        GpaCounterGroupDesc{486, "RMI32", 32, 139, 2, 4},
+        GpaCounterGroupDesc{487, "RMI33", 33, 139, 2, 4},
+        GpaCounterGroupDesc{488, "RMI34", 34, 139, 2, 4},
+        GpaCounterGroupDesc{489, "RMI35", 35, 139, 2, 4},
+        GpaCounterGroupDesc{490, "RMI36", 36, 139, 2, 4},
+        GpaCounterGroupDesc{491, "RMI37", 37, 139, 2, 4},
+        GpaCounterGroupDesc{492, "RMI38", 38, 139, 2, 4},
+        GpaCounterGroupDesc{493, "RMI39", 39, 139, 2, 4},
+        GpaCounterGroupDesc{494, "RMI40", 40, 139, 2, 4},
+        GpaCounterGroupDesc{495, "RMI41", 41, 139, 2, 4},
+        GpaCounterGroupDesc{496, "RMI42", 42, 139, 2, 4},
+        GpaCounterGroupDesc{497, "RMI43", 43, 139, 2, 4},
+        GpaCounterGroupDesc{498, "RMI44", 44, 139, 2, 4},
+        GpaCounterGroupDesc{499, "RMI45", 45, 139, 2, 4},
+        GpaCounterGroupDesc{500, "RMI46", 46, 139, 2, 4},
+        GpaCounterGroupDesc{501, "RMI47", 47, 139, 2, 4},
+        GpaCounterGroupDesc{502, "CHA", 0, 40, 4, 4},
+        GpaCounterGroupDesc{503, "CHC0", 0, 44, 4, 4},
+        GpaCounterGroupDesc{504, "CHC1", 1, 44, 4, 4},
+        GpaCounterGroupDesc{505, "CHC2", 2, 44, 4, 4},
+        GpaCounterGroupDesc{506, "CHC3", 3, 44, 4, 4},
+        GpaCounterGroupDesc{507, "CHCG", 0, 44, 4, 4},
+        GpaCounterGroupDesc{508, "GCR", 0, 155, 2, 4},
+        GpaCounterGroupDesc{509, "GE", 0, 40, 4, 16},
+        GpaCounterGroupDesc{510, "GL1A0", 0, 24, 4, 4},
+        GpaCounterGroupDesc{511, "GL1A1", 1, 24, 4, 4},
+        GpaCounterGroupDesc{512, "GL1A2", 2, 24, 4, 4},
+        GpaCounterGroupDesc{513, "GL1A3", 3, 24, 4, 4},
+        GpaCounterGroupDesc{514, "GL1A4", 4, 24, 4, 4},
+        GpaCounterGroupDesc{515, "GL1A5", 5, 24, 4, 4},
+        GpaCounterGroupDesc{516, "GL1A6", 6, 24, 4, 4},
+        GpaCounterGroupDesc{517, "GL1A7", 7, 24, 4, 4},
+        GpaCounterGroupDesc{518, "GL1A8", 8, 24, 4, 4},
+        GpaCounterGroupDesc{519, "GL1A9", 9, 24, 4, 4},
+        GpaCounterGroupDesc{520, "GL1A10", 10, 24, 4, 4},
+        GpaCounterGroupDesc{521, "GL1A11", 11, 24, 4, 4},
+        GpaCounterGroupDesc{522, "GL1C0", 0, 84, 4, 4},
+        GpaCounterGroupDesc{523, "GL1C1", 1, 84, 4, 4},
+        GpaCounterGroupDesc{524, "GL1C2", 2, 84, 4, 4},
+        GpaCounterGroupDesc{525, "GL1C3", 3, 84, 4, 4},
+        GpaCounterGroupDesc{526, "GL1C4", 4, 84, 4, 4},
+        GpaCounterGroupDesc{527, "GL1C5", 5, 84, 4, 4},
+        GpaCounterGroupDesc{528, "GL1C6", 6, 84, 4, 4},
+        GpaCounterGroupDesc{529, "GL1C7", 7, 84, 4, 4},
+        GpaCounterGroupDesc{530, "GL1C8", 8, 84, 4, 4},
+        GpaCounterGroupDesc{531, "GL1C9", 9, 84, 4, 4},
+        GpaCounterGroupDesc{532, "GL1C10", 10, 84, 4, 4},
+        GpaCounterGroupDesc{533, "GL1C11", 11, 84, 4, 4},
+        GpaCounterGroupDesc{534, "GL1C12", 12, 84, 4, 4},
+        GpaCounterGroupDesc{535, "GL1C13", 13, 84, 4, 4},
+        GpaCounterGroupDesc{536, "GL1C14", 14, 84, 4, 4},
+        GpaCounterGroupDesc{537, "GL1C15", 15, 84, 4, 4},
+        GpaCounterGroupDesc{538, "GL1C16", 16, 84, 4, 4},
+        GpaCounterGroupDesc{539, "GL1C17", 17, 84, 4, 4},
+        GpaCounterGroupDesc{540, "GL1C18", 18, 84, 4, 4},
+        GpaCounterGroupDesc{541, "GL1C19", 19, 84, 4, 4},
+        GpaCounterGroupDesc{542, "GL1C20", 20, 84, 4, 4},
+        GpaCounterGroupDesc{543, "GL1C21", 21, 84, 4, 4},
+        GpaCounterGroupDesc{544, "GL1C22", 22, 84, 4, 4},
+        GpaCounterGroupDesc{545, "GL1C23", 23, 84, 4, 4},
+        GpaCounterGroupDesc{546, "GL1C24", 24, 84, 4, 4},
+        GpaCounterGroupDesc{547, "GL1C25", 25, 84, 4, 4},
+        GpaCounterGroupDesc{548, "GL1C26", 26, 84, 4, 4},
+        GpaCounterGroupDesc{549, "GL1C27", 27, 84, 4, 4},
+        GpaCounterGroupDesc{550, "GL1C28", 28, 84, 4, 4},
+        GpaCounterGroupDesc{551, "GL1C29", 29, 84, 4, 4},
+        GpaCounterGroupDesc{552, "GL1C30", 30, 84, 4, 4},
+        GpaCounterGroupDesc{553, "GL1C31", 31, 84, 4, 4},
+        GpaCounterGroupDesc{554, "GL1C32", 32, 84, 4, 4},
+        GpaCounterGroupDesc{555, "GL1C33", 33, 84, 4, 4},
+        GpaCounterGroupDesc{556, "GL1C34", 34, 84, 4, 4},
+        GpaCounterGroupDesc{557, "GL1C35", 35, 84, 4, 4},
+        GpaCounterGroupDesc{558, "GL1C36", 36, 84, 4, 4},
+        GpaCounterGroupDesc{559, "GL1C37", 37, 84, 4, 4},
+        GpaCounterGroupDesc{560, "GL1C38", 38, 84, 4, 4},
+        GpaCounterGroupDesc{561, "GL1C39", 39, 84, 4, 4},
+        GpaCounterGroupDesc{562, "GL1C40", 40, 84, 4, 4},
+        GpaCounterGroupDesc{563, "GL1C41", 41, 84, 4, 4},
+        GpaCounterGroupDesc{564, "GL1C42", 42, 84, 4, 4},
+        GpaCounterGroupDesc{565, "GL1C43", 43, 84, 4, 4},
+        GpaCounterGroupDesc{566, "GL1C44", 44, 84, 4, 4},
+        GpaCounterGroupDesc{567, "GL1C45", 45, 84, 4, 4},
+        GpaCounterGroupDesc{568, "GL1C46", 46, 84, 4, 4},
+        GpaCounterGroupDesc{569, "GL1C47", 47, 84, 4, 4},
+        GpaCounterGroupDesc{570, "GL2A0", 0, 108, 4, 8},
+        GpaCounterGroupDesc{571, "GL2A1", 1, 108, 4, 8},
+        GpaCounterGroupDesc{572, "GL2A2", 2, 108, 4, 8},
+        GpaCounterGroupDesc{573, "GL2A3", 3, 108, 4, 8},
+        GpaCounterGroupDesc{574, "GL2C0", 0, 259, 4, 8},
+        GpaCounterGroupDesc{575, "GL2C1", 1, 259, 4, 8},
+        GpaCounterGroupDesc{576, "GL2C2", 2, 259, 4, 8},
+        GpaCounterGroupDesc{577, "GL2C3", 3, 259, 4, 8},
+        GpaCounterGroupDesc{578, "GL2C4", 4, 259, 4, 8},
+        GpaCounterGroupDesc{579, "GL2C5", 5, 259, 4, 8},
+        GpaCounterGroupDesc{580, "GL2C6", 6, 259, 4, 8},
+        GpaCounterGroupDesc{581, "GL2C7", 7, 259, 4, 8},
+        GpaCounterGroupDesc{582, "GL2C8", 8, 259, 4, 8},
+        GpaCounterGroupDesc{583, "GL2C9", 9, 259, 4, 8},
+        GpaCounterGroupDesc{584, "GL2C10", 10, 259, 4, 8},
+        GpaCounterGroupDesc{585, "GL2C11", 11, 259, 4, 8},
+        GpaCounterGroupDesc{586, "GL2C12", 12, 259, 4, 8},
+        GpaCounterGroupDesc{587, "GL2C13", 13, 259, 4, 8},
+        GpaCounterGroupDesc{588, "GL2C14", 14, 259, 4, 8},
+        GpaCounterGroupDesc{589, "GL2C15", 15, 259, 4, 8},
+        GpaCounterGroupDesc{590, "GL2C16", 16, 259, 4, 8},
+        GpaCounterGroupDesc{591, "GL2C17", 17, 259, 4, 8},
+        GpaCounterGroupDesc{592, "GL2C18", 18, 259, 4, 8},
+        GpaCounterGroupDesc{593, "GL2C19", 19, 259, 4, 8},
+        GpaCounterGroupDesc{594, "GL2C20", 20, 259, 4, 8},
+        GpaCounterGroupDesc{595, "GL2C21", 21, 259, 4, 8},
+        GpaCounterGroupDesc{596, "GL2C22", 22, 259, 4, 8},
+        GpaCounterGroupDesc{597, "GL2C23", 23, 259, 4, 8},
+        GpaCounterGroupDesc{598, "GUS", 0, 176, 2, 4},
+        GpaCounterGroupDesc{599, "PA_PH", 0, 1024, 8, 16},
+        GpaCounterGroupDesc{600, "UTCL10", 0, 66, 2, 0},
+        GpaCounterGroupDesc{601, "UTCL11", 1, 66, 2, 0},
+        GpaCounterGroupDesc{602, "UTCL12", 2, 66, 2, 0},
+        GpaCounterGroupDesc{603, "UTCL13", 3, 66, 2, 0},
+        GpaCounterGroupDesc{604, "UTCL14", 4, 66, 2, 0},
+        GpaCounterGroupDesc{605, "UTCL15", 5, 66, 2, 0},
+        GpaCounterGroupDesc{606, "UTCL16", 6, 66, 2, 0},
+        GpaCounterGroupDesc{607, "UTCL17", 7, 66, 2, 0},
+        GpaCounterGroupDesc{608, "UTCL18", 8, 66, 2, 0},
+        GpaCounterGroupDesc{609, "UTCL19", 9, 66, 2, 0},
+        GpaCounterGroupDesc{610, "UTCL110", 10, 66, 2, 0},
+        GpaCounterGroupDesc{611, "UTCL111", 11, 66, 2, 0},
+        GpaCounterGroupDesc{612, "GE2_DIST", 0, 101, 4, 16},
+        GpaCounterGroupDesc{613, "GE_SE0", 0, 88, 4, 16},
+        GpaCounterGroupDesc{614, "GE_SE1", 1, 88, 4, 16},
+        GpaCounterGroupDesc{615, "GE_SE2", 2, 88, 4, 16},
+        GpaCounterGroupDesc{616, "GE_SE3", 3, 88, 4, 16},
+        GpaCounterGroupDesc{617, "GE_SE4", 4, 88, 4, 16},
+        GpaCounterGroupDesc{618, "GE_SE5", 5, 88, 4, 16},
+        GpaCounterGroupDesc{619, "SQWGP0", 0, 512, 8, 16},
+        GpaCounterGroupDesc{620, "SQWGP1", 1, 512, 8, 16},
+        GpaCounterGroupDesc{621, "SQWGP2", 2, 512, 8, 16},
+        GpaCounterGroupDesc{622, "SQWGP3", 3, 512, 8, 16},
+        GpaCounterGroupDesc{623, "SQWGP4", 4, 512, 8, 16},
+        GpaCounterGroupDesc{624, "SQWGP5", 5, 512, 8, 16},
+        GpaCounterGroupDesc{625, "SQWGP6", 6, 512, 8, 16},
+        GpaCounterGroupDesc{626, "SQWGP7", 7, 512, 8, 16},
+        GpaCounterGroupDesc{627, "SQWGP8", 8, 512, 8, 16},
+        GpaCounterGroupDesc{628, "SQWGP9", 9, 512, 8, 16},
+        GpaCounterGroupDesc{629, "SQWGP10", 10, 512, 8, 16},
+        GpaCounterGroupDesc{630, "SQWGP11", 11, 512, 8, 16},
+        GpaCounterGroupDesc{631, "SQWGP12", 12, 512, 8, 16},
+        GpaCounterGroupDesc{632, "SQWGP13", 13, 512, 8, 16},
+        GpaCounterGroupDesc{633, "SQWGP14", 14, 512, 8, 16},
+        GpaCounterGroupDesc{634, "SQWGP15", 15, 512, 8, 16},
+        GpaCounterGroupDesc{635, "SQWGP16", 16, 512, 8, 16},
+        GpaCounterGroupDesc{636, "SQWGP17", 17, 512, 8, 16},
+        GpaCounterGroupDesc{637, "SQWGP18", 18, 512, 8, 16},
+        GpaCounterGroupDesc{638, "SQWGP19", 19, 512, 8, 16},
+        GpaCounterGroupDesc{639, "SQWGP20", 20, 512, 8, 16},
+        GpaCounterGroupDesc{640, "SQWGP21", 21, 512, 8, 16},
+        GpaCounterGroupDesc{641, "SQWGP22", 22, 512, 8, 16},
+        GpaCounterGroupDesc{642, "SQWGP23", 23, 512, 8, 16},
+        GpaCounterGroupDesc{643, "SQWGP24", 24, 512, 8, 16},
+        GpaCounterGroupDesc{644, "SQWGP25", 25, 512, 8, 16},
+        GpaCounterGroupDesc{645, "SQWGP26", 26, 512, 8, 16},
+        GpaCounterGroupDesc{646, "SQWGP27", 27, 512, 8, 16},
+        GpaCounterGroupDesc{647, "SQWGP28", 28, 512, 8, 16},
+        GpaCounterGroupDesc{648, "SQWGP29", 29, 512, 8, 16},
+        GpaCounterGroupDesc{649, "SQWGP30", 30, 512, 8, 16},
+        GpaCounterGroupDesc{650, "SQWGP31", 31, 512, 8, 16},
+        GpaCounterGroupDesc{651, "SQWGP32", 32, 512, 8, 16},
+        GpaCounterGroupDesc{652, "SQWGP33", 33, 512, 8, 16},
+        GpaCounterGroupDesc{653, "SQWGP34", 34, 512, 8, 16},
+        GpaCounterGroupDesc{654, "SQWGP35", 35, 512, 8, 16},
+        GpaCounterGroupDesc{655, "SQWGP36", 36, 512, 8, 16},
+        GpaCounterGroupDesc{656, "SQWGP37", 37, 512, 8, 16},
+        GpaCounterGroupDesc{657, "SQWGP38", 38, 512, 8, 16},
+        GpaCounterGroupDesc{658, "SQWGP39", 39, 512, 8, 16},
+        GpaCounterGroupDesc{659, "SQWGP40", 40, 512, 8, 16},
+        GpaCounterGroupDesc{660, "SQWGP41", 41, 512, 8, 16},
+        GpaCounterGroupDesc{661, "SQWGP42", 42, 512, 8, 16},
+        GpaCounterGroupDesc{662, "SQWGP43", 43, 512, 8, 16},
+        GpaCounterGroupDesc{663, "SQWGP44", 44, 512, 8, 16},
+        GpaCounterGroupDesc{664, "SQWGP45", 45, 512, 8, 16},
+        GpaCounterGroupDesc{665, "SQWGP46", 46, 512, 8, 16},
+        GpaCounterGroupDesc{666, "SQWGP47", 47, 512, 8, 16},
+        GpaCounterGroupDesc{667, "SQWGP_GS0", 0, 512, 8, 16},
+        GpaCounterGroupDesc{668, "SQWGP_GS1", 1, 512, 8, 16},
+        GpaCounterGroupDesc{669, "SQWGP_GS2", 2, 512, 8, 16},
+        GpaCounterGroupDesc{670, "SQWGP_GS3", 3, 512, 8, 16},
+        GpaCounterGroupDesc{671, "SQWGP_GS4", 4, 512, 8, 16},
+        GpaCounterGroupDesc{672, "SQWGP_GS5", 5, 512, 8, 16},
+        GpaCounterGroupDesc{673, "SQWGP_GS6", 6, 512, 8, 16},
+        GpaCounterGroupDesc{674, "SQWGP_GS7", 7, 512, 8, 16},
+        GpaCounterGroupDesc{675, "SQWGP_GS8", 8, 512, 8, 16},
+        GpaCounterGroupDesc{676, "SQWGP_GS9", 9, 512, 8, 16},
+        GpaCounterGroupDesc{677, "SQWGP_GS10", 10, 512, 8, 16},
+        GpaCounterGroupDesc{678, "SQWGP_GS11", 11, 512, 8, 16},
+        GpaCounterGroupDesc{679, "SQWGP_GS12", 12, 512, 8, 16},
+        GpaCounterGroupDesc{680, "SQWGP_GS13", 13, 512, 8, 16},
+        GpaCounterGroupDesc{681, "SQWGP_GS14", 14, 512, 8, 16},
+        GpaCounterGroupDesc{682, "SQWGP_GS15", 15, 512, 8, 16},
+        GpaCounterGroupDesc{683, "SQWGP_GS16", 16, 512, 8, 16},
+        GpaCounterGroupDesc{684, "SQWGP_GS17", 17, 512, 8, 16},
+        GpaCounterGroupDesc{685, "SQWGP_GS18", 18, 512, 8, 16},
+        GpaCounterGroupDesc{686, "SQWGP_GS19", 19, 512, 8, 16},
+        GpaCounterGroupDesc{687, "SQWGP_GS20", 20, 512, 8, 16},
+        GpaCounterGroupDesc{688, "SQWGP_GS21", 21, 512, 8, 16},
+        GpaCounterGroupDesc{689, "SQWGP_GS22", 22, 512, 8, 16},
+        GpaCounterGroupDesc{690, "SQWGP_GS23", 23, 512, 8, 16},
+        GpaCounterGroupDesc{691, "SQWGP_GS24", 24, 512, 8, 16},
+        GpaCounterGroupDesc{692, "SQWGP_GS25", 25, 512, 8, 16},
+        GpaCounterGroupDesc{693, "SQWGP_GS26", 26, 512, 8, 16},
+        GpaCounterGroupDesc{694, "SQWGP_GS27", 27, 512, 8, 16},
+        GpaCounterGroupDesc{695, "SQWGP_GS28", 28, 512, 8, 16},
+        GpaCounterGroupDesc{696, "SQWGP_GS29", 29, 512, 8, 16},
+        GpaCounterGroupDesc{697, "SQWGP_GS30", 30, 512, 8, 16},
+        GpaCounterGroupDesc{698, "SQWGP_GS31", 31, 512, 8, 16},
+        GpaCounterGroupDesc{699, "SQWGP_GS32", 32, 512, 8, 16},
+        GpaCounterGroupDesc{700, "SQWGP_GS33", 33, 512, 8, 16},
+        GpaCounterGroupDesc{701, "SQWGP_GS34", 34, 512, 8, 16},
+        GpaCounterGroupDesc{702, "SQWGP_GS35", 35, 512, 8, 16},
+        GpaCounterGroupDesc{703, "SQWGP_GS36", 36, 512, 8, 16},
+        GpaCounterGroupDesc{704, "SQWGP_GS37", 37, 512, 8, 16},
+        GpaCounterGroupDesc{705, "SQWGP_GS38", 38, 512, 8, 16},
+        GpaCounterGroupDesc{706, "SQWGP_GS39", 39, 512, 8, 16},
+        GpaCounterGroupDesc{707, "SQWGP_GS40", 40, 512, 8, 16},
+        GpaCounterGroupDesc{708, "SQWGP_GS41", 41, 512, 8, 16},
+        GpaCounterGroupDesc{709, "SQWGP_GS42", 42, 512, 8, 16},
+        GpaCounterGroupDesc{710, "SQWGP_GS43", 43, 512, 8, 16},
+        GpaCounterGroupDesc{711, "SQWGP_GS44", 44, 512, 8, 16},
+        GpaCounterGroupDesc{712, "SQWGP_GS45", 45, 512, 8, 16},
+        GpaCounterGroupDesc{713, "SQWGP_GS46", 46, 512, 8, 16},
+        GpaCounterGroupDesc{714, "SQWGP_GS47", 47, 512, 8, 16},
+        GpaCounterGroupDesc{715, "SQWGP_PS0", 0, 512, 8, 16},
+        GpaCounterGroupDesc{716, "SQWGP_PS1", 1, 512, 8, 16},
+        GpaCounterGroupDesc{717, "SQWGP_PS2", 2, 512, 8, 16},
+        GpaCounterGroupDesc{718, "SQWGP_PS3", 3, 512, 8, 16},
+        GpaCounterGroupDesc{719, "SQWGP_PS4", 4, 512, 8, 16},
+        GpaCounterGroupDesc{720, "SQWGP_PS5", 5, 512, 8, 16},
+        GpaCounterGroupDesc{721, "SQWGP_PS6", 6, 512, 8, 16},
+        GpaCounterGroupDesc{722, "SQWGP_PS7", 7, 512, 8, 16},
+        GpaCounterGroupDesc{723, "SQWGP_PS8", 8, 512, 8, 16},
+        GpaCounterGroupDesc{724, "SQWGP_PS9", 9, 512, 8, 16},
+        GpaCounterGroupDesc{725, "SQWGP_PS10", 10, 512, 8, 16},
+        GpaCounterGroupDesc{726, "SQWGP_PS11", 11, 512, 8, 16},
+        GpaCounterGroupDesc{727, "SQWGP_PS12", 12, 512, 8, 16},
+        GpaCounterGroupDesc{728, "SQWGP_PS13", 13, 512, 8, 16},
+        GpaCounterGroupDesc{729, "SQWGP_PS14", 14, 512, 8, 16},
+        GpaCounterGroupDesc{730, "SQWGP_PS15", 15, 512, 8, 16},
+        GpaCounterGroupDesc{731, "SQWGP_PS16", 16, 512, 8, 16},
+        GpaCounterGroupDesc{732, "SQWGP_PS17", 17, 512, 8, 16},
+        GpaCounterGroupDesc{733, "SQWGP_PS18", 18, 512, 8, 16},
+        GpaCounterGroupDesc{734, "SQWGP_PS19", 19, 512, 8, 16},
+        GpaCounterGroupDesc{735, "SQWGP_PS20", 20, 512, 8, 16},
+        GpaCounterGroupDesc{736, "SQWGP_PS21", 21, 512, 8, 16},
+        GpaCounterGroupDesc{737, "SQWGP_PS22", 22, 512, 8, 16},
+        GpaCounterGroupDesc{738, "SQWGP_PS23", 23, 512, 8, 16},
+        GpaCounterGroupDesc{739, "SQWGP_PS24", 24, 512, 8, 16},
+        GpaCounterGroupDesc{740, "SQWGP_PS25", 25, 512, 8, 16},
+        GpaCounterGroupDesc{741, "SQWGP_PS26", 26, 512, 8, 16},
+        GpaCounterGroupDesc{742, "SQWGP_PS27", 27, 512, 8, 16},
+        GpaCounterGroupDesc{743, "SQWGP_PS28", 28, 512, 8, 16},
+        GpaCounterGroupDesc{744, "SQWGP_PS29", 29, 512, 8, 16},
+        GpaCounterGroupDesc{745, "SQWGP_PS30", 30, 512, 8, 16},
+        GpaCounterGroupDesc{746, "SQWGP_PS31", 31, 512, 8, 16},
+        GpaCounterGroupDesc{747, "SQWGP_PS32", 32, 512, 8, 16},
+        GpaCounterGroupDesc{748, "SQWGP_PS33", 33, 512, 8, 16},
+        GpaCounterGroupDesc{749, "SQWGP_PS34", 34, 512, 8, 16},
+        GpaCounterGroupDesc{750, "SQWGP_PS35", 35, 512, 8, 16},
+        GpaCounterGroupDesc{751, "SQWGP_PS36", 36, 512, 8, 16},
+        GpaCounterGroupDesc{752, "SQWGP_PS37", 37, 512, 8, 16},
+        GpaCounterGroupDesc{753, "SQWGP_PS38", 38, 512, 8, 16},
+        GpaCounterGroupDesc{754, "SQWGP_PS39", 39, 512, 8, 16},
+        GpaCounterGroupDesc{755, "SQWGP_PS40", 40, 512, 8, 16},
+        GpaCounterGroupDesc{756, "SQWGP_PS41", 41, 512, 8, 16},
+        GpaCounterGroupDesc{757, "SQWGP_PS42", 42, 512, 8, 16},
+        GpaCounterGroupDesc{758, "SQWGP_PS43", 43, 512, 8, 16},
+        GpaCounterGroupDesc{759, "SQWGP_PS44", 44, 512, 8, 16},
+        GpaCounterGroupDesc{760, "SQWGP_PS45", 45, 512, 8, 16},
+        GpaCounterGroupDesc{761, "SQWGP_PS46", 46, 512, 8, 16},
+        GpaCounterGroupDesc{762, "SQWGP_PS47", 47, 512, 8, 16},
+        GpaCounterGroupDesc{763, "SQWGP_HS0", 0, 512, 8, 16},
+        GpaCounterGroupDesc{764, "SQWGP_HS1", 1, 512, 8, 16},
+        GpaCounterGroupDesc{765, "SQWGP_HS2", 2, 512, 8, 16},
+        GpaCounterGroupDesc{766, "SQWGP_HS3", 3, 512, 8, 16},
+        GpaCounterGroupDesc{767, "SQWGP_HS4", 4, 512, 8, 16},
+        GpaCounterGroupDesc{768, "SQWGP_HS5", 5, 512, 8, 16},
+        GpaCounterGroupDesc{769, "SQWGP_HS6", 6, 512, 8, 16},
+        GpaCounterGroupDesc{770, "SQWGP_HS7", 7, 512, 8, 16},
+        GpaCounterGroupDesc{771, "SQWGP_HS8", 8, 512, 8, 16},
+        GpaCounterGroupDesc{772, "SQWGP_HS9", 9, 512, 8, 16},
+        GpaCounterGroupDesc{773, "SQWGP_HS10", 10, 512, 8, 16},
+        GpaCounterGroupDesc{774, "SQWGP_HS11", 11, 512, 8, 16},
+        GpaCounterGroupDesc{775, "SQWGP_HS12", 12, 512, 8, 16},
+        GpaCounterGroupDesc{776, "SQWGP_HS13", 13, 512, 8, 16},
+        GpaCounterGroupDesc{777, "SQWGP_HS14", 14, 512, 8, 16},
+        GpaCounterGroupDesc{778, "SQWGP_HS15", 15, 512, 8, 16},
+        GpaCounterGroupDesc{779, "SQWGP_HS16", 16, 512, 8, 16},
+        GpaCounterGroupDesc{780, "SQWGP_HS17", 17, 512, 8, 16},
+        GpaCounterGroupDesc{781, "SQWGP_HS18", 18, 512, 8, 16},
+        GpaCounterGroupDesc{782, "SQWGP_HS19", 19, 512, 8, 16},
+        GpaCounterGroupDesc{783, "SQWGP_HS20", 20, 512, 8, 16},
+        GpaCounterGroupDesc{784, "SQWGP_HS21", 21, 512, 8, 16},
+        GpaCounterGroupDesc{785, "SQWGP_HS22", 22, 512, 8, 16},
+        GpaCounterGroupDesc{786, "SQWGP_HS23", 23, 512, 8, 16},
+        GpaCounterGroupDesc{787, "SQWGP_HS24", 24, 512, 8, 16},
+        GpaCounterGroupDesc{788, "SQWGP_HS25", 25, 512, 8, 16},
+        GpaCounterGroupDesc{789, "SQWGP_HS26", 26, 512, 8, 16},
+        GpaCounterGroupDesc{790, "SQWGP_HS27", 27, 512, 8, 16},
+        GpaCounterGroupDesc{791, "SQWGP_HS28", 28, 512, 8, 16},
+        GpaCounterGroupDesc{792, "SQWGP_HS29", 29, 512, 8, 16},
+        GpaCounterGroupDesc{793, "SQWGP_HS30", 30, 512, 8, 16},
+        GpaCounterGroupDesc{794, "SQWGP_HS31", 31, 512, 8, 16},
+        GpaCounterGroupDesc{795, "SQWGP_HS32", 32, 512, 8, 16},
+        GpaCounterGroupDesc{796, "SQWGP_HS33", 33, 512, 8, 16},
+        GpaCounterGroupDesc{797, "SQWGP_HS34", 34, 512, 8, 16},
+        GpaCounterGroupDesc{798, "SQWGP_HS35", 35, 512, 8, 16},
+        GpaCounterGroupDesc{799, "SQWGP_HS36", 36, 512, 8, 16},
+        GpaCounterGroupDesc{800, "SQWGP_HS37", 37, 512, 8, 16},
+        GpaCounterGroupDesc{801, "SQWGP_HS38", 38, 512, 8, 16},
+        GpaCounterGroupDesc{802, "SQWGP_HS39", 39, 512, 8, 16},
+        GpaCounterGroupDesc{803, "SQWGP_HS40", 40, 512, 8, 16},
+        GpaCounterGroupDesc{804, "SQWGP_HS41", 41, 512, 8, 16},
+        GpaCounterGroupDesc{805, "SQWGP_HS42", 42, 512, 8, 16},
+        GpaCounterGroupDesc{806, "SQWGP_HS43", 43, 512, 8, 16},
+        GpaCounterGroupDesc{807, "SQWGP_HS44", 44, 512, 8, 16},
+        GpaCounterGroupDesc{808, "SQWGP_HS45", 45, 512, 8, 16},
+        GpaCounterGroupDesc{809, "SQWGP_HS46", 46, 512, 8, 16},
+        GpaCounterGroupDesc{810, "SQWGP_HS47", 47, 512, 8, 16},
+        GpaCounterGroupDesc{811, "SQWGP_CS0", 0, 512, 8, 16},
+        GpaCounterGroupDesc{812, "SQWGP_CS1", 1, 512, 8, 16},
+        GpaCounterGroupDesc{813, "SQWGP_CS2", 2, 512, 8, 16},
+        GpaCounterGroupDesc{814, "SQWGP_CS3", 3, 512, 8, 16},
+        GpaCounterGroupDesc{815, "SQWGP_CS4", 4, 512, 8, 16},
+        GpaCounterGroupDesc{816, "SQWGP_CS5", 5, 512, 8, 16},
+        GpaCounterGroupDesc{817, "SQWGP_CS6", 6, 512, 8, 16},
+        GpaCounterGroupDesc{818, "SQWGP_CS7", 7, 512, 8, 16},
+        GpaCounterGroupDesc{819, "SQWGP_CS8", 8, 512, 8, 16},
+        GpaCounterGroupDesc{820, "SQWGP_CS9", 9, 512, 8, 16},
+        GpaCounterGroupDesc{821, "SQWGP_CS10", 10, 512, 8, 16},
+        GpaCounterGroupDesc{822, "SQWGP_CS11", 11, 512, 8, 16},
+        GpaCounterGroupDesc{823, "SQWGP_CS12", 12, 512, 8, 16},
+        GpaCounterGroupDesc{824, "SQWGP_CS13", 13, 512, 8, 16},
+        GpaCounterGroupDesc{825, "SQWGP_CS14", 14, 512, 8, 16},
+        GpaCounterGroupDesc{826, "SQWGP_CS15", 15, 512, 8, 16},
+        GpaCounterGroupDesc{827, "SQWGP_CS16", 16, 512, 8, 16},
+        GpaCounterGroupDesc{828, "SQWGP_CS17", 17, 512, 8, 16},
+        GpaCounterGroupDesc{829, "SQWGP_CS18", 18, 512, 8, 16},
+        GpaCounterGroupDesc{830, "SQWGP_CS19", 19, 512, 8, 16},
+        GpaCounterGroupDesc{831, "SQWGP_CS20", 20, 512, 8, 16},
+        GpaCounterGroupDesc{832, "SQWGP_CS21", 21, 512, 8, 16},
+        GpaCounterGroupDesc{833, "SQWGP_CS22", 22, 512, 8, 16},
+        GpaCounterGroupDesc{834, "SQWGP_CS23", 23, 512, 8, 16},
+        GpaCounterGroupDesc{835, "SQWGP_CS24", 24, 512, 8, 16},
+        GpaCounterGroupDesc{836, "SQWGP_CS25", 25, 512, 8, 16},
+        GpaCounterGroupDesc{837, "SQWGP_CS26", 26, 512, 8, 16},
+        GpaCounterGroupDesc{838, "SQWGP_CS27", 27, 512, 8, 16},
+        GpaCounterGroupDesc{839, "SQWGP_CS28", 28, 512, 8, 16},
+        GpaCounterGroupDesc{840, "SQWGP_CS29", 29, 512, 8, 16},
+        GpaCounterGroupDesc{841, "SQWGP_CS30", 30, 512, 8, 16},
+        GpaCounterGroupDesc{842, "SQWGP_CS31", 31, 512, 8, 16},
+        GpaCounterGroupDesc{843, "SQWGP_CS32", 32, 512, 8, 16},
+        GpaCounterGroupDesc{844, "SQWGP_CS33", 33, 512, 8, 16},
+        GpaCounterGroupDesc{845, "SQWGP_CS34", 34, 512, 8, 16},
+        GpaCounterGroupDesc{846, "SQWGP_CS35", 35, 512, 8, 16},
+        GpaCounterGroupDesc{847, "SQWGP_CS36", 36, 512, 8, 16},
+        GpaCounterGroupDesc{848, "SQWGP_CS37", 37, 512, 8, 16},
+        GpaCounterGroupDesc{849, "SQWGP_CS38", 38, 512, 8, 16},
+        GpaCounterGroupDesc{850, "SQWGP_CS39", 39, 512, 8, 16},
+        GpaCounterGroupDesc{851, "SQWGP_CS40", 40, 512, 8, 16},
+        GpaCounterGroupDesc{852, "SQWGP_CS41", 41, 512, 8, 16},
+        GpaCounterGroupDesc{853, "SQWGP_CS42", 42, 512, 8, 16},
+        GpaCounterGroupDesc{854, "SQWGP_CS43", 43, 512, 8, 16},
+        GpaCounterGroupDesc{855, "SQWGP_CS44", 44, 512, 8, 16},
+        GpaCounterGroupDesc{856, "SQWGP_CS45", 45, 512, 8, 16},
+        GpaCounterGroupDesc{857, "SQWGP_CS46", 46, 512, 8, 16},
+        GpaCounterGroupDesc{858, "SQWGP_CS47", 47, 512, 8, 16},
+        GpaCounterGroupDesc{859, "PC0", 0, 47, 4, 0},
+        GpaCounterGroupDesc{860, "PC1", 1, 47, 4, 0},
+        GpaCounterGroupDesc{861, "PC2", 2, 47, 4, 0},
+        GpaCounterGroupDesc{862, "PC3", 3, 47, 4, 0},
+        GpaCounterGroupDesc{863, "PC4", 4, 47, 4, 0},
+        GpaCounterGroupDesc{864, "PC5", 5, 47, 4, 0},
+        GpaCounterGroupDesc{865, "GPUTimeStamp", 0, 1, 1, 0},
+        GpaCounterGroupDesc{866, "GPUTime", 0, 6, 3, 0},
     };
     const gpa_array_view<GpaCounterGroupDesc> kHwDx11GroupsGfx11 = kHwDx11GroupsGfx11Data;
 
@@ -2069,350 +2071,350 @@ namespace counter_dx11_gfx11
         {420, 89030},
         {421, 89109},
         {422, 89188},
-        {428, 89718},
-        {429, 89816},
-        {430, 89914},
-        {431, 90012},
-        {432, 90110},
-        {433, 90208},
-        {434, 90306},
-        {435, 90404},
-        {436, 90502},
-        {437, 90600},
-        {438, 90698},
-        {439, 90796},
-        {440, 90894},
-        {441, 90992},
-        {442, 91090},
-        {443, 91188},
-        {444, 91286},
-        {445, 91384},
-        {446, 91482},
-        {447, 91580},
-        {448, 91678},
-        {449, 91776},
-        {450, 91874},
-        {451, 91972},
-        {520, 99485},
-        {521, 99569},
-        {522, 99653},
-        {523, 99737},
-        {524, 99821},
-        {525, 99905},
-        {526, 99989},
-        {527, 100073},
-        {528, 100157},
-        {529, 100241},
-        {530, 100325},
-        {531, 100409},
-        {532, 100493},
-        {533, 100577},
-        {534, 100661},
-        {535, 100745},
-        {536, 100829},
-        {537, 100913},
-        {538, 100997},
-        {539, 101081},
-        {540, 101165},
-        {541, 101249},
-        {542, 101333},
-        {543, 101417},
-        {544, 101501},
-        {545, 101585},
-        {546, 101669},
-        {547, 101753},
-        {548, 101837},
-        {549, 101921},
-        {550, 102005},
-        {551, 102089},
-        {552, 102173},
-        {553, 102257},
-        {554, 102341},
-        {555, 102425},
-        {556, 102509},
-        {557, 102593},
-        {558, 102677},
-        {559, 102761},
-        {560, 102845},
-        {561, 102929},
-        {562, 103013},
-        {563, 103097},
-        {564, 103181},
-        {565, 103265},
-        {566, 103349},
-        {567, 103433},
-        {572, 103949},
-        {573, 104208},
-        {574, 104467},
-        {575, 104726},
-        {576, 104985},
-        {577, 105244},
-        {578, 105503},
-        {579, 105762},
-        {580, 106021},
-        {581, 106280},
-        {582, 106539},
-        {583, 106798},
-        {584, 107057},
-        {585, 107316},
-        {586, 107575},
-        {587, 107834},
-        {588, 108093},
-        {589, 108352},
-        {590, 108611},
-        {591, 108870},
-        {592, 109129},
-        {593, 109388},
-        {594, 109647},
-        {595, 109906},
-        {610, 112157},
-        {611, 112258},
-        {612, 112346},
-        {613, 112434},
-        {614, 112522},
-        {615, 112610},
-        {616, 112698},
-        {617, 112786},
-        {618, 113298},
-        {619, 113810},
-        {620, 114322},
-        {621, 114834},
-        {622, 115346},
-        {623, 115858},
-        {624, 116370},
-        {625, 116882},
-        {626, 117394},
-        {627, 117906},
-        {628, 118418},
-        {629, 118930},
-        {630, 119442},
-        {631, 119954},
-        {632, 120466},
-        {633, 120978},
-        {634, 121490},
-        {635, 122002},
-        {636, 122514},
-        {637, 123026},
-        {638, 123538},
-        {639, 124050},
-        {640, 124562},
-        {641, 125074},
-        {642, 125586},
-        {643, 126098},
-        {644, 126610},
-        {645, 127122},
-        {646, 127634},
-        {647, 128146},
-        {648, 128658},
-        {649, 129170},
-        {650, 129682},
-        {651, 130194},
-        {652, 130706},
-        {653, 131218},
-        {654, 131730},
-        {655, 132242},
-        {656, 132754},
-        {657, 133266},
-        {658, 133778},
-        {659, 134290},
-        {660, 134802},
-        {661, 135314},
-        {662, 135826},
-        {663, 136338},
-        {664, 136850},
-        {665, 137362},
-        {666, 137874},
-        {667, 138386},
-        {668, 138898},
-        {669, 139410},
-        {670, 139922},
-        {671, 140434},
-        {672, 140946},
-        {673, 141458},
-        {674, 141970},
-        {675, 142482},
-        {676, 142994},
-        {677, 143506},
-        {678, 144018},
-        {679, 144530},
-        {680, 145042},
-        {681, 145554},
-        {682, 146066},
-        {683, 146578},
-        {684, 147090},
-        {685, 147602},
-        {686, 148114},
-        {687, 148626},
-        {688, 149138},
-        {689, 149650},
-        {690, 150162},
-        {691, 150674},
-        {692, 151186},
-        {693, 151698},
-        {694, 152210},
-        {695, 152722},
-        {696, 153234},
-        {697, 153746},
-        {698, 154258},
-        {699, 154770},
-        {700, 155282},
-        {701, 155794},
-        {702, 156306},
-        {703, 156818},
-        {704, 157330},
-        {705, 157842},
-        {706, 158354},
-        {707, 158866},
-        {708, 159378},
-        {709, 159890},
-        {710, 160402},
-        {711, 160914},
-        {712, 161426},
-        {713, 161938},
-        {714, 162450},
-        {715, 162962},
-        {716, 163474},
-        {717, 163986},
-        {718, 164498},
-        {719, 165010},
-        {720, 165522},
-        {721, 166034},
-        {722, 166546},
-        {723, 167058},
-        {724, 167570},
-        {725, 168082},
-        {726, 168594},
-        {727, 169106},
-        {728, 169618},
-        {729, 170130},
-        {730, 170642},
-        {731, 171154},
-        {732, 171666},
-        {733, 172178},
-        {734, 172690},
-        {735, 173202},
-        {736, 173714},
-        {737, 174226},
-        {738, 174738},
-        {739, 175250},
-        {740, 175762},
-        {741, 176274},
-        {742, 176786},
-        {743, 177298},
-        {744, 177810},
-        {745, 178322},
-        {746, 178834},
-        {747, 179346},
-        {748, 179858},
-        {749, 180370},
-        {750, 180882},
-        {751, 181394},
-        {752, 181906},
-        {753, 182418},
-        {754, 182930},
-        {755, 183442},
-        {756, 183954},
-        {757, 184466},
-        {758, 184978},
-        {759, 185490},
-        {760, 186002},
-        {761, 186514},
-        {762, 187026},
-        {763, 187538},
-        {764, 188050},
-        {765, 188562},
-        {766, 189074},
-        {767, 189586},
-        {768, 190098},
-        {769, 190610},
-        {770, 191122},
-        {771, 191634},
-        {772, 192146},
-        {773, 192658},
-        {774, 193170},
-        {775, 193682},
-        {776, 194194},
-        {777, 194706},
-        {778, 195218},
-        {779, 195730},
-        {780, 196242},
-        {781, 196754},
-        {782, 197266},
-        {783, 197778},
-        {784, 198290},
-        {785, 198802},
-        {786, 199314},
-        {787, 199826},
-        {788, 200338},
-        {789, 200850},
-        {790, 201362},
-        {791, 201874},
-        {792, 202386},
-        {793, 202898},
-        {794, 203410},
-        {795, 203922},
-        {796, 204434},
-        {797, 204946},
-        {798, 205458},
-        {799, 205970},
-        {800, 206482},
-        {801, 206994},
-        {802, 207506},
-        {803, 208018},
-        {804, 208530},
-        {805, 209042},
-        {806, 209554},
-        {807, 210066},
-        {808, 210578},
-        {809, 211090},
-        {810, 211602},
-        {811, 212114},
-        {812, 212626},
-        {813, 213138},
-        {814, 213650},
-        {815, 214162},
-        {816, 214674},
-        {817, 215186},
-        {818, 215698},
-        {819, 216210},
-        {820, 216722},
-        {821, 217234},
-        {822, 217746},
-        {823, 218258},
-        {824, 218770},
-        {825, 219282},
-        {826, 219794},
-        {827, 220306},
-        {828, 220818},
-        {829, 221330},
-        {830, 221842},
-        {831, 222354},
-        {832, 222866},
-        {833, 223378},
-        {834, 223890},
-        {835, 224402},
-        {836, 224914},
-        {837, 225426},
-        {838, 225938},
-        {839, 226450},
-        {840, 226962},
-        {841, 227474},
-        {842, 227986},
-        {843, 228498},
-        {844, 229010},
-        {845, 229522},
-        {846, 230034},
-        {847, 230546},
-        {848, 231058},
-        {849, 231570},
-        {850, 232082},
-        {851, 232594},
-        {852, 233106},
-        {853, 233618},
-        {854, 234130},
-        {855, 234642},
-        {856, 235154},
-        {864, 235949},
+        {430, 89760},
+        {431, 89858},
+        {432, 89956},
+        {433, 90054},
+        {434, 90152},
+        {435, 90250},
+        {436, 90348},
+        {437, 90446},
+        {438, 90544},
+        {439, 90642},
+        {440, 90740},
+        {441, 90838},
+        {442, 90936},
+        {443, 91034},
+        {444, 91132},
+        {445, 91230},
+        {446, 91328},
+        {447, 91426},
+        {448, 91524},
+        {449, 91622},
+        {450, 91720},
+        {451, 91818},
+        {452, 91916},
+        {453, 92014},
+        {522, 99527},
+        {523, 99611},
+        {524, 99695},
+        {525, 99779},
+        {526, 99863},
+        {527, 99947},
+        {528, 100031},
+        {529, 100115},
+        {530, 100199},
+        {531, 100283},
+        {532, 100367},
+        {533, 100451},
+        {534, 100535},
+        {535, 100619},
+        {536, 100703},
+        {537, 100787},
+        {538, 100871},
+        {539, 100955},
+        {540, 101039},
+        {541, 101123},
+        {542, 101207},
+        {543, 101291},
+        {544, 101375},
+        {545, 101459},
+        {546, 101543},
+        {547, 101627},
+        {548, 101711},
+        {549, 101795},
+        {550, 101879},
+        {551, 101963},
+        {552, 102047},
+        {553, 102131},
+        {554, 102215},
+        {555, 102299},
+        {556, 102383},
+        {557, 102467},
+        {558, 102551},
+        {559, 102635},
+        {560, 102719},
+        {561, 102803},
+        {562, 102887},
+        {563, 102971},
+        {564, 103055},
+        {565, 103139},
+        {566, 103223},
+        {567, 103307},
+        {568, 103391},
+        {569, 103475},
+        {574, 103991},
+        {575, 104250},
+        {576, 104509},
+        {577, 104768},
+        {578, 105027},
+        {579, 105286},
+        {580, 105545},
+        {581, 105804},
+        {582, 106063},
+        {583, 106322},
+        {584, 106581},
+        {585, 106840},
+        {586, 107099},
+        {587, 107358},
+        {588, 107617},
+        {589, 107876},
+        {590, 108135},
+        {591, 108394},
+        {592, 108653},
+        {593, 108912},
+        {594, 109171},
+        {595, 109430},
+        {596, 109689},
+        {597, 109948},
+        {612, 112199},
+        {613, 112300},
+        {614, 112388},
+        {615, 112476},
+        {616, 112564},
+        {617, 112652},
+        {618, 112740},
+        {619, 112828},
+        {620, 113340},
+        {621, 113852},
+        {622, 114364},
+        {623, 114876},
+        {624, 115388},
+        {625, 115900},
+        {626, 116412},
+        {627, 116924},
+        {628, 117436},
+        {629, 117948},
+        {630, 118460},
+        {631, 118972},
+        {632, 119484},
+        {633, 119996},
+        {634, 120508},
+        {635, 121020},
+        {636, 121532},
+        {637, 122044},
+        {638, 122556},
+        {639, 123068},
+        {640, 123580},
+        {641, 124092},
+        {642, 124604},
+        {643, 125116},
+        {644, 125628},
+        {645, 126140},
+        {646, 126652},
+        {647, 127164},
+        {648, 127676},
+        {649, 128188},
+        {650, 128700},
+        {651, 129212},
+        {652, 129724},
+        {653, 130236},
+        {654, 130748},
+        {655, 131260},
+        {656, 131772},
+        {657, 132284},
+        {658, 132796},
+        {659, 133308},
+        {660, 133820},
+        {661, 134332},
+        {662, 134844},
+        {663, 135356},
+        {664, 135868},
+        {665, 136380},
+        {666, 136892},
+        {667, 137404},
+        {668, 137916},
+        {669, 138428},
+        {670, 138940},
+        {671, 139452},
+        {672, 139964},
+        {673, 140476},
+        {674, 140988},
+        {675, 141500},
+        {676, 142012},
+        {677, 142524},
+        {678, 143036},
+        {679, 143548},
+        {680, 144060},
+        {681, 144572},
+        {682, 145084},
+        {683, 145596},
+        {684, 146108},
+        {685, 146620},
+        {686, 147132},
+        {687, 147644},
+        {688, 148156},
+        {689, 148668},
+        {690, 149180},
+        {691, 149692},
+        {692, 150204},
+        {693, 150716},
+        {694, 151228},
+        {695, 151740},
+        {696, 152252},
+        {697, 152764},
+        {698, 153276},
+        {699, 153788},
+        {700, 154300},
+        {701, 154812},
+        {702, 155324},
+        {703, 155836},
+        {704, 156348},
+        {705, 156860},
+        {706, 157372},
+        {707, 157884},
+        {708, 158396},
+        {709, 158908},
+        {710, 159420},
+        {711, 159932},
+        {712, 160444},
+        {713, 160956},
+        {714, 161468},
+        {715, 161980},
+        {716, 162492},
+        {717, 163004},
+        {718, 163516},
+        {719, 164028},
+        {720, 164540},
+        {721, 165052},
+        {722, 165564},
+        {723, 166076},
+        {724, 166588},
+        {725, 167100},
+        {726, 167612},
+        {727, 168124},
+        {728, 168636},
+        {729, 169148},
+        {730, 169660},
+        {731, 170172},
+        {732, 170684},
+        {733, 171196},
+        {734, 171708},
+        {735, 172220},
+        {736, 172732},
+        {737, 173244},
+        {738, 173756},
+        {739, 174268},
+        {740, 174780},
+        {741, 175292},
+        {742, 175804},
+        {743, 176316},
+        {744, 176828},
+        {745, 177340},
+        {746, 177852},
+        {747, 178364},
+        {748, 178876},
+        {749, 179388},
+        {750, 179900},
+        {751, 180412},
+        {752, 180924},
+        {753, 181436},
+        {754, 181948},
+        {755, 182460},
+        {756, 182972},
+        {757, 183484},
+        {758, 183996},
+        {759, 184508},
+        {760, 185020},
+        {761, 185532},
+        {762, 186044},
+        {763, 186556},
+        {764, 187068},
+        {765, 187580},
+        {766, 188092},
+        {767, 188604},
+        {768, 189116},
+        {769, 189628},
+        {770, 190140},
+        {771, 190652},
+        {772, 191164},
+        {773, 191676},
+        {774, 192188},
+        {775, 192700},
+        {776, 193212},
+        {777, 193724},
+        {778, 194236},
+        {779, 194748},
+        {780, 195260},
+        {781, 195772},
+        {782, 196284},
+        {783, 196796},
+        {784, 197308},
+        {785, 197820},
+        {786, 198332},
+        {787, 198844},
+        {788, 199356},
+        {789, 199868},
+        {790, 200380},
+        {791, 200892},
+        {792, 201404},
+        {793, 201916},
+        {794, 202428},
+        {795, 202940},
+        {796, 203452},
+        {797, 203964},
+        {798, 204476},
+        {799, 204988},
+        {800, 205500},
+        {801, 206012},
+        {802, 206524},
+        {803, 207036},
+        {804, 207548},
+        {805, 208060},
+        {806, 208572},
+        {807, 209084},
+        {808, 209596},
+        {809, 210108},
+        {810, 210620},
+        {811, 211132},
+        {812, 211644},
+        {813, 212156},
+        {814, 212668},
+        {815, 213180},
+        {816, 213692},
+        {817, 214204},
+        {818, 214716},
+        {819, 215228},
+        {820, 215740},
+        {821, 216252},
+        {822, 216764},
+        {823, 217276},
+        {824, 217788},
+        {825, 218300},
+        {826, 218812},
+        {827, 219324},
+        {828, 219836},
+        {829, 220348},
+        {830, 220860},
+        {831, 221372},
+        {832, 221884},
+        {833, 222396},
+        {834, 222908},
+        {835, 223420},
+        {836, 223932},
+        {837, 224444},
+        {838, 224956},
+        {839, 225468},
+        {840, 225980},
+        {841, 226492},
+        {842, 227004},
+        {843, 227516},
+        {844, 228028},
+        {845, 228540},
+        {846, 229052},
+        {847, 229564},
+        {848, 230076},
+        {849, 230588},
+        {850, 231100},
+        {851, 231612},
+        {852, 232124},
+        {853, 232636},
+        {854, 233148},
+        {855, 233660},
+        {856, 234172},
+        {857, 234684},
+        {858, 235196},
+        {866, 235991},
     };
 
     GpaPaddedCounterDesc kDx11PaddedCounterByGroupGfx11[] = {
@@ -2656,12 +2658,14 @@ namespace counter_dx11_gfx11
         {420, 89030, {78}},
         {421, 89109, {78}},
         {422, 89188, {78}},
-        {857, 235666, {45, 46}},
-        {858, 235713, {45, 46}},
-        {859, 235760, {45, 46}},
-        {860, 235807, {45, 46}},
-        {861, 235854, {45, 46}},
-        {862, 235901, {45, 46}},
+        {427, 89654, {0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23}},
+        {428, 89678, {0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17}},
+        {859, 235708, {45, 46}},
+        {860, 235755, {45, 46}},
+        {861, 235802, {45, 46}},
+        {862, 235849, {45, 46}},
+        {863, 235896, {45, 46}},
+        {864, 235943, {45, 46}},
     };
 
     GpaSqCounterGroupDesc kHwDx11SqGroupsGfx11[] = {
@@ -2695,246 +2699,246 @@ namespace counter_dx11_gfx11
         {132, 3, kSqCs},
         {133, 4, kSqCs},
         {134, 5, kSqCs},
-        {617, 0, kSqAll},
-        {618, 1, kSqAll},
-        {619, 2, kSqAll},
-        {620, 3, kSqAll},
-        {621, 4, kSqAll},
-        {622, 5, kSqAll},
-        {623, 6, kSqAll},
-        {624, 7, kSqAll},
-        {625, 8, kSqAll},
-        {626, 9, kSqAll},
-        {627, 10, kSqAll},
-        {628, 11, kSqAll},
-        {629, 12, kSqAll},
-        {630, 13, kSqAll},
-        {631, 14, kSqAll},
-        {632, 15, kSqAll},
-        {633, 16, kSqAll},
-        {634, 17, kSqAll},
-        {635, 18, kSqAll},
-        {636, 19, kSqAll},
-        {637, 20, kSqAll},
-        {638, 21, kSqAll},
-        {639, 22, kSqAll},
-        {640, 23, kSqAll},
-        {641, 24, kSqAll},
-        {642, 25, kSqAll},
-        {643, 26, kSqAll},
-        {644, 27, kSqAll},
-        {645, 28, kSqAll},
-        {646, 29, kSqAll},
-        {647, 30, kSqAll},
-        {648, 31, kSqAll},
-        {649, 32, kSqAll},
-        {650, 33, kSqAll},
-        {651, 34, kSqAll},
-        {652, 35, kSqAll},
-        {653, 36, kSqAll},
-        {654, 37, kSqAll},
-        {655, 38, kSqAll},
-        {656, 39, kSqAll},
-        {657, 40, kSqAll},
-        {658, 41, kSqAll},
-        {659, 42, kSqAll},
-        {660, 43, kSqAll},
-        {661, 44, kSqAll},
-        {662, 45, kSqAll},
-        {663, 46, kSqAll},
-        {664, 47, kSqAll},
-        {665, 0, kSqGs},
-        {666, 1, kSqGs},
-        {667, 2, kSqGs},
-        {668, 3, kSqGs},
-        {669, 4, kSqGs},
-        {670, 5, kSqGs},
-        {671, 6, kSqGs},
-        {672, 7, kSqGs},
-        {673, 8, kSqGs},
-        {674, 9, kSqGs},
-        {675, 10, kSqGs},
-        {676, 11, kSqGs},
-        {677, 12, kSqGs},
-        {678, 13, kSqGs},
-        {679, 14, kSqGs},
-        {680, 15, kSqGs},
-        {681, 16, kSqGs},
-        {682, 17, kSqGs},
-        {683, 18, kSqGs},
-        {684, 19, kSqGs},
-        {685, 20, kSqGs},
-        {686, 21, kSqGs},
-        {687, 22, kSqGs},
-        {688, 23, kSqGs},
-        {689, 24, kSqGs},
-        {690, 25, kSqGs},
-        {691, 26, kSqGs},
-        {692, 27, kSqGs},
-        {693, 28, kSqGs},
-        {694, 29, kSqGs},
-        {695, 30, kSqGs},
-        {696, 31, kSqGs},
-        {697, 32, kSqGs},
-        {698, 33, kSqGs},
-        {699, 34, kSqGs},
-        {700, 35, kSqGs},
-        {701, 36, kSqGs},
-        {702, 37, kSqGs},
-        {703, 38, kSqGs},
-        {704, 39, kSqGs},
-        {705, 40, kSqGs},
-        {706, 41, kSqGs},
-        {707, 42, kSqGs},
-        {708, 43, kSqGs},
-        {709, 44, kSqGs},
-        {710, 45, kSqGs},
-        {711, 46, kSqGs},
-        {712, 47, kSqGs},
-        {713, 0, kSqPs},
-        {714, 1, kSqPs},
-        {715, 2, kSqPs},
-        {716, 3, kSqPs},
-        {717, 4, kSqPs},
-        {718, 5, kSqPs},
-        {719, 6, kSqPs},
-        {720, 7, kSqPs},
-        {721, 8, kSqPs},
-        {722, 9, kSqPs},
-        {723, 10, kSqPs},
-        {724, 11, kSqPs},
-        {725, 12, kSqPs},
-        {726, 13, kSqPs},
-        {727, 14, kSqPs},
-        {728, 15, kSqPs},
-        {729, 16, kSqPs},
-        {730, 17, kSqPs},
-        {731, 18, kSqPs},
-        {732, 19, kSqPs},
-        {733, 20, kSqPs},
-        {734, 21, kSqPs},
-        {735, 22, kSqPs},
-        {736, 23, kSqPs},
-        {737, 24, kSqPs},
-        {738, 25, kSqPs},
-        {739, 26, kSqPs},
-        {740, 27, kSqPs},
-        {741, 28, kSqPs},
-        {742, 29, kSqPs},
-        {743, 30, kSqPs},
-        {744, 31, kSqPs},
-        {745, 32, kSqPs},
-        {746, 33, kSqPs},
-        {747, 34, kSqPs},
-        {748, 35, kSqPs},
-        {749, 36, kSqPs},
-        {750, 37, kSqPs},
-        {751, 38, kSqPs},
-        {752, 39, kSqPs},
-        {753, 40, kSqPs},
-        {754, 41, kSqPs},
-        {755, 42, kSqPs},
-        {756, 43, kSqPs},
-        {757, 44, kSqPs},
-        {758, 45, kSqPs},
-        {759, 46, kSqPs},
-        {760, 47, kSqPs},
-        {761, 0, kSqHs},
-        {762, 1, kSqHs},
-        {763, 2, kSqHs},
-        {764, 3, kSqHs},
-        {765, 4, kSqHs},
-        {766, 5, kSqHs},
-        {767, 6, kSqHs},
-        {768, 7, kSqHs},
-        {769, 8, kSqHs},
-        {770, 9, kSqHs},
-        {771, 10, kSqHs},
-        {772, 11, kSqHs},
-        {773, 12, kSqHs},
-        {774, 13, kSqHs},
-        {775, 14, kSqHs},
-        {776, 15, kSqHs},
-        {777, 16, kSqHs},
-        {778, 17, kSqHs},
-        {779, 18, kSqHs},
-        {780, 19, kSqHs},
-        {781, 20, kSqHs},
-        {782, 21, kSqHs},
-        {783, 22, kSqHs},
-        {784, 23, kSqHs},
-        {785, 24, kSqHs},
-        {786, 25, kSqHs},
-        {787, 26, kSqHs},
-        {788, 27, kSqHs},
-        {789, 28, kSqHs},
-        {790, 29, kSqHs},
-        {791, 30, kSqHs},
-        {792, 31, kSqHs},
-        {793, 32, kSqHs},
-        {794, 33, kSqHs},
-        {795, 34, kSqHs},
-        {796, 35, kSqHs},
-        {797, 36, kSqHs},
-        {798, 37, kSqHs},
-        {799, 38, kSqHs},
-        {800, 39, kSqHs},
-        {801, 40, kSqHs},
-        {802, 41, kSqHs},
-        {803, 42, kSqHs},
-        {804, 43, kSqHs},
-        {805, 44, kSqHs},
-        {806, 45, kSqHs},
-        {807, 46, kSqHs},
-        {808, 47, kSqHs},
-        {809, 0, kSqCs},
-        {810, 1, kSqCs},
-        {811, 2, kSqCs},
-        {812, 3, kSqCs},
-        {813, 4, kSqCs},
-        {814, 5, kSqCs},
-        {815, 6, kSqCs},
-        {816, 7, kSqCs},
-        {817, 8, kSqCs},
-        {818, 9, kSqCs},
-        {819, 10, kSqCs},
-        {820, 11, kSqCs},
-        {821, 12, kSqCs},
-        {822, 13, kSqCs},
-        {823, 14, kSqCs},
-        {824, 15, kSqCs},
-        {825, 16, kSqCs},
-        {826, 17, kSqCs},
-        {827, 18, kSqCs},
-        {828, 19, kSqCs},
-        {829, 20, kSqCs},
-        {830, 21, kSqCs},
-        {831, 22, kSqCs},
-        {832, 23, kSqCs},
-        {833, 24, kSqCs},
-        {834, 25, kSqCs},
-        {835, 26, kSqCs},
-        {836, 27, kSqCs},
-        {837, 28, kSqCs},
-        {838, 29, kSqCs},
-        {839, 30, kSqCs},
-        {840, 31, kSqCs},
-        {841, 32, kSqCs},
-        {842, 33, kSqCs},
-        {843, 34, kSqCs},
-        {844, 35, kSqCs},
-        {845, 36, kSqCs},
-        {846, 37, kSqCs},
-        {847, 38, kSqCs},
-        {848, 39, kSqCs},
-        {849, 40, kSqCs},
-        {850, 41, kSqCs},
-        {851, 42, kSqCs},
-        {852, 43, kSqCs},
-        {853, 44, kSqCs},
-        {854, 45, kSqCs},
-        {855, 46, kSqCs},
-        {856, 47, kSqCs},
+        {619, 0, kSqAll},
+        {620, 1, kSqAll},
+        {621, 2, kSqAll},
+        {622, 3, kSqAll},
+        {623, 4, kSqAll},
+        {624, 5, kSqAll},
+        {625, 6, kSqAll},
+        {626, 7, kSqAll},
+        {627, 8, kSqAll},
+        {628, 9, kSqAll},
+        {629, 10, kSqAll},
+        {630, 11, kSqAll},
+        {631, 12, kSqAll},
+        {632, 13, kSqAll},
+        {633, 14, kSqAll},
+        {634, 15, kSqAll},
+        {635, 16, kSqAll},
+        {636, 17, kSqAll},
+        {637, 18, kSqAll},
+        {638, 19, kSqAll},
+        {639, 20, kSqAll},
+        {640, 21, kSqAll},
+        {641, 22, kSqAll},
+        {642, 23, kSqAll},
+        {643, 24, kSqAll},
+        {644, 25, kSqAll},
+        {645, 26, kSqAll},
+        {646, 27, kSqAll},
+        {647, 28, kSqAll},
+        {648, 29, kSqAll},
+        {649, 30, kSqAll},
+        {650, 31, kSqAll},
+        {651, 32, kSqAll},
+        {652, 33, kSqAll},
+        {653, 34, kSqAll},
+        {654, 35, kSqAll},
+        {655, 36, kSqAll},
+        {656, 37, kSqAll},
+        {657, 38, kSqAll},
+        {658, 39, kSqAll},
+        {659, 40, kSqAll},
+        {660, 41, kSqAll},
+        {661, 42, kSqAll},
+        {662, 43, kSqAll},
+        {663, 44, kSqAll},
+        {664, 45, kSqAll},
+        {665, 46, kSqAll},
+        {666, 47, kSqAll},
+        {667, 0, kSqGs},
+        {668, 1, kSqGs},
+        {669, 2, kSqGs},
+        {670, 3, kSqGs},
+        {671, 4, kSqGs},
+        {672, 5, kSqGs},
+        {673, 6, kSqGs},
+        {674, 7, kSqGs},
+        {675, 8, kSqGs},
+        {676, 9, kSqGs},
+        {677, 10, kSqGs},
+        {678, 11, kSqGs},
+        {679, 12, kSqGs},
+        {680, 13, kSqGs},
+        {681, 14, kSqGs},
+        {682, 15, kSqGs},
+        {683, 16, kSqGs},
+        {684, 17, kSqGs},
+        {685, 18, kSqGs},
+        {686, 19, kSqGs},
+        {687, 20, kSqGs},
+        {688, 21, kSqGs},
+        {689, 22, kSqGs},
+        {690, 23, kSqGs},
+        {691, 24, kSqGs},
+        {692, 25, kSqGs},
+        {693, 26, kSqGs},
+        {694, 27, kSqGs},
+        {695, 28, kSqGs},
+        {696, 29, kSqGs},
+        {697, 30, kSqGs},
+        {698, 31, kSqGs},
+        {699, 32, kSqGs},
+        {700, 33, kSqGs},
+        {701, 34, kSqGs},
+        {702, 35, kSqGs},
+        {703, 36, kSqGs},
+        {704, 37, kSqGs},
+        {705, 38, kSqGs},
+        {706, 39, kSqGs},
+        {707, 40, kSqGs},
+        {708, 41, kSqGs},
+        {709, 42, kSqGs},
+        {710, 43, kSqGs},
+        {711, 44, kSqGs},
+        {712, 45, kSqGs},
+        {713, 46, kSqGs},
+        {714, 47, kSqGs},
+        {715, 0, kSqPs},
+        {716, 1, kSqPs},
+        {717, 2, kSqPs},
+        {718, 3, kSqPs},
+        {719, 4, kSqPs},
+        {720, 5, kSqPs},
+        {721, 6, kSqPs},
+        {722, 7, kSqPs},
+        {723, 8, kSqPs},
+        {724, 9, kSqPs},
+        {725, 10, kSqPs},
+        {726, 11, kSqPs},
+        {727, 12, kSqPs},
+        {728, 13, kSqPs},
+        {729, 14, kSqPs},
+        {730, 15, kSqPs},
+        {731, 16, kSqPs},
+        {732, 17, kSqPs},
+        {733, 18, kSqPs},
+        {734, 19, kSqPs},
+        {735, 20, kSqPs},
+        {736, 21, kSqPs},
+        {737, 22, kSqPs},
+        {738, 23, kSqPs},
+        {739, 24, kSqPs},
+        {740, 25, kSqPs},
+        {741, 26, kSqPs},
+        {742, 27, kSqPs},
+        {743, 28, kSqPs},
+        {744, 29, kSqPs},
+        {745, 30, kSqPs},
+        {746, 31, kSqPs},
+        {747, 32, kSqPs},
+        {748, 33, kSqPs},
+        {749, 34, kSqPs},
+        {750, 35, kSqPs},
+        {751, 36, kSqPs},
+        {752, 37, kSqPs},
+        {753, 38, kSqPs},
+        {754, 39, kSqPs},
+        {755, 40, kSqPs},
+        {756, 41, kSqPs},
+        {757, 42, kSqPs},
+        {758, 43, kSqPs},
+        {759, 44, kSqPs},
+        {760, 45, kSqPs},
+        {761, 46, kSqPs},
+        {762, 47, kSqPs},
+        {763, 0, kSqHs},
+        {764, 1, kSqHs},
+        {765, 2, kSqHs},
+        {766, 3, kSqHs},
+        {767, 4, kSqHs},
+        {768, 5, kSqHs},
+        {769, 6, kSqHs},
+        {770, 7, kSqHs},
+        {771, 8, kSqHs},
+        {772, 9, kSqHs},
+        {773, 10, kSqHs},
+        {774, 11, kSqHs},
+        {775, 12, kSqHs},
+        {776, 13, kSqHs},
+        {777, 14, kSqHs},
+        {778, 15, kSqHs},
+        {779, 16, kSqHs},
+        {780, 17, kSqHs},
+        {781, 18, kSqHs},
+        {782, 19, kSqHs},
+        {783, 20, kSqHs},
+        {784, 21, kSqHs},
+        {785, 22, kSqHs},
+        {786, 23, kSqHs},
+        {787, 24, kSqHs},
+        {788, 25, kSqHs},
+        {789, 26, kSqHs},
+        {790, 27, kSqHs},
+        {791, 28, kSqHs},
+        {792, 29, kSqHs},
+        {793, 30, kSqHs},
+        {794, 31, kSqHs},
+        {795, 32, kSqHs},
+        {796, 33, kSqHs},
+        {797, 34, kSqHs},
+        {798, 35, kSqHs},
+        {799, 36, kSqHs},
+        {800, 37, kSqHs},
+        {801, 38, kSqHs},
+        {802, 39, kSqHs},
+        {803, 40, kSqHs},
+        {804, 41, kSqHs},
+        {805, 42, kSqHs},
+        {806, 43, kSqHs},
+        {807, 44, kSqHs},
+        {808, 45, kSqHs},
+        {809, 46, kSqHs},
+        {810, 47, kSqHs},
+        {811, 0, kSqCs},
+        {812, 1, kSqCs},
+        {813, 2, kSqCs},
+        {814, 3, kSqCs},
+        {815, 4, kSqCs},
+        {816, 5, kSqCs},
+        {817, 6, kSqCs},
+        {818, 7, kSqCs},
+        {819, 8, kSqCs},
+        {820, 9, kSqCs},
+        {821, 10, kSqCs},
+        {822, 11, kSqCs},
+        {823, 12, kSqCs},
+        {824, 13, kSqCs},
+        {825, 14, kSqCs},
+        {826, 15, kSqCs},
+        {827, 16, kSqCs},
+        {828, 17, kSqCs},
+        {829, 18, kSqCs},
+        {830, 19, kSqCs},
+        {831, 20, kSqCs},
+        {832, 21, kSqCs},
+        {833, 22, kSqCs},
+        {834, 23, kSqCs},
+        {835, 24, kSqCs},
+        {836, 25, kSqCs},
+        {837, 26, kSqCs},
+        {838, 27, kSqCs},
+        {839, 28, kSqCs},
+        {840, 29, kSqCs},
+        {841, 30, kSqCs},
+        {842, 31, kSqCs},
+        {843, 32, kSqCs},
+        {844, 33, kSqCs},
+        {845, 34, kSqCs},
+        {846, 35, kSqCs},
+        {847, 36, kSqCs},
+        {848, 37, kSqCs},
+        {849, 38, kSqCs},
+        {850, 39, kSqCs},
+        {851, 40, kSqCs},
+        {852, 41, kSqCs},
+        {853, 42, kSqCs},
+        {854, 43, kSqCs},
+        {855, 44, kSqCs},
+        {856, 45, kSqCs},
+        {857, 46, kSqCs},
+        {858, 47, kSqCs},
     };
 
     unsigned int kHwDx11SqIsolatedGroupsGfx11[] = {
@@ -3656,6 +3660,8 @@ namespace counter_dx11_gfx11
     PE_BLOCK_CPG,
     PE_BLOCK_CPC,
     PE_BLOCK_MCVML2,
+    PE_BLOCK_ATC,
+    PE_BLOCK_ATCL2,
     PE_BLOCK_RPB,
     PE_BLOCK_EA,
     PE_BLOCK_EA,
@@ -4097,16 +4103,16 @@ namespace counter_dx11_gfx11
     };
 
     // DX11: Only supports GPU time duration counters. Unavailable counters indices are assigned -1.
-    const unsigned int kHwDx11GpuTimeBottomToBottomDurationIndexGfx11 = 235949; ///< Index of delta between the previous command reaching bottom of pipe and the current command reaching bottom of pipe, will not include latency of first data to travel through pipeline, best for large data sets. for Dx11 for Gfx11 family
+    const unsigned int kHwDx11GpuTimeBottomToBottomDurationIndexGfx11 = 235991; ///< Index of delta between the previous command reaching bottom of pipe and the current command reaching bottom of pipe, will not include latency of first data to travel through pipeline, best for large data sets. for Dx11 for Gfx11 family
     const unsigned int kHwDx11GpuTimeBottomToBottomStartIndexGfx11 = static_cast<unsigned int>(-1); ///< Index of time of the previous command reaching bottom of pipe for Dx11 for Gfx11 family
     const unsigned int kHwDx11GpuTimeBottomToBottomEndIndexGfx11 = static_cast<unsigned int>(-1); ///< Index of time of the current command reaching bottom of pipe for Dx11 for Gfx11 family
-    const unsigned int kHwDx11GpuTimeTopToBottomDurationIndexGfx11 = 235952; ///< Index of execution duration of the current command from top of pipe to bottom of pipe, may include overhead of time in queue for Dx11 for Gfx11 family
+    const unsigned int kHwDx11GpuTimeTopToBottomDurationIndexGfx11 = 235994; ///< Index of execution duration of the current command from top of pipe to bottom of pipe, may include overhead of time in queue for Dx11 for Gfx11 family
     const unsigned int kHwDx11GpuTimeTopToBottomStartIndexGfx11 = static_cast<unsigned int>(-1); ///< Index of time that the current command reaches the top of pipe for Dx11 for Gfx11 family
     const unsigned int kHwDx11GpuTimeTopToBottomEndIndexGfx11 = static_cast<unsigned int>(-1); ///< Index of time that the current command reaches the bottom of pipe for Dx11 for Gfx11 family
-    const std::set<unsigned int> kHwDx11TimestampBlockIdsGfx11 =       {864, 863}; ///< Timestamp block id's for DX11 for gfx11 family
-    const std::set<unsigned int> kHwDx11TimeCounterIndicesGfx11 =      {235949, 235950, 235951, 235952, 235953, 235954, 235948}; ///< Timestamp counter indices for DX11 for gfx11 family
+    const std::set<unsigned int> kHwDx11TimestampBlockIdsGfx11 =       {866, 865}; ///< Timestamp block id's for DX11 for gfx11 family
+    const std::set<unsigned int> kHwDx11TimeCounterIndicesGfx11 =      {235991, 235992, 235993, 235994, 235995, 235996, 235990}; ///< Timestamp counter indices for DX11 for gfx11 family
     const unsigned int           kHwDx11GroupCountGfx11                = static_cast<unsigned int>(kHwDx11GroupsGfx11.size());
-    const unsigned int           kDx11PaddedCounterGroupCountGfx11 = 246;
+    const unsigned int           kDx11PaddedCounterGroupCountGfx11 = 248;
     const unsigned int           kHwDx11SqGroupCountGfx11              = sizeof(kHwDx11SqGroupsGfx11) / sizeof(GpaSqCounterGroupDesc);
     const unsigned int           kHwDx11SqIsolatedGroupCountGfx11      = sizeof(kHwDx11SqIsolatedGroupsGfx11) / sizeof(unsigned int);
 } //  namespace counter_dx11_gfx11
