@@ -1216,468 +1216,470 @@ namespace counter_vk_gfx11
         GpaCounterGroupDesc{429, "SDMA5", 5, 101, 2, 8},
         GpaCounterGroupDesc{430, "CPG", 0, 92, 2, 4},
         GpaCounterGroupDesc{431, "CPC", 0, 56, 2, 4},
-        GpaCounterGroupDesc{432, "GCVML2", 0, 91, 8, 8},
-        GpaCounterGroupDesc{433, "GCEA0", 0, 98, 2, 4},
-        GpaCounterGroupDesc{434, "GCEA1", 1, 98, 2, 4},
-        GpaCounterGroupDesc{435, "GCEA2", 2, 98, 2, 4},
-        GpaCounterGroupDesc{436, "GCEA3", 3, 98, 2, 4},
-        GpaCounterGroupDesc{437, "GCEA4", 4, 98, 2, 4},
-        GpaCounterGroupDesc{438, "GCEA5", 5, 98, 2, 4},
-        GpaCounterGroupDesc{439, "GCEA6", 6, 98, 2, 4},
-        GpaCounterGroupDesc{440, "GCEA7", 7, 98, 2, 4},
-        GpaCounterGroupDesc{441, "GCEA8", 8, 98, 2, 4},
-        GpaCounterGroupDesc{442, "GCEA9", 9, 98, 2, 4},
-        GpaCounterGroupDesc{443, "GCEA10", 10, 98, 2, 4},
-        GpaCounterGroupDesc{444, "GCEA11", 11, 98, 2, 4},
-        GpaCounterGroupDesc{445, "GCEA12", 12, 98, 2, 4},
-        GpaCounterGroupDesc{446, "GCEA13", 13, 98, 2, 4},
-        GpaCounterGroupDesc{447, "GCEA14", 14, 98, 2, 4},
-        GpaCounterGroupDesc{448, "GCEA15", 15, 98, 2, 4},
-        GpaCounterGroupDesc{449, "GCEA16", 16, 98, 2, 4},
-        GpaCounterGroupDesc{450, "GCEA17", 17, 98, 2, 4},
-        GpaCounterGroupDesc{451, "GCEA18", 18, 98, 2, 4},
-        GpaCounterGroupDesc{452, "GCEA19", 19, 98, 2, 4},
-        GpaCounterGroupDesc{453, "GCEA20", 20, 98, 2, 4},
-        GpaCounterGroupDesc{454, "GCEA21", 21, 98, 2, 4},
-        GpaCounterGroupDesc{455, "GCEA22", 22, 98, 2, 4},
-        GpaCounterGroupDesc{456, "GCEA23", 23, 98, 2, 4},
-        GpaCounterGroupDesc{457, "RPB", 0, 64, 4, 0},
-        GpaCounterGroupDesc{458, "RMI0", 0, 139, 2, 4},
-        GpaCounterGroupDesc{459, "RMI1", 1, 139, 2, 4},
-        GpaCounterGroupDesc{460, "RMI2", 2, 139, 2, 4},
-        GpaCounterGroupDesc{461, "RMI3", 3, 139, 2, 4},
-        GpaCounterGroupDesc{462, "RMI4", 4, 139, 2, 4},
-        GpaCounterGroupDesc{463, "RMI5", 5, 139, 2, 4},
-        GpaCounterGroupDesc{464, "RMI6", 6, 139, 2, 4},
-        GpaCounterGroupDesc{465, "RMI7", 7, 139, 2, 4},
-        GpaCounterGroupDesc{466, "RMI8", 8, 139, 2, 4},
-        GpaCounterGroupDesc{467, "RMI9", 9, 139, 2, 4},
-        GpaCounterGroupDesc{468, "RMI10", 10, 139, 2, 4},
-        GpaCounterGroupDesc{469, "RMI11", 11, 139, 2, 4},
-        GpaCounterGroupDesc{470, "RMI12", 12, 139, 2, 4},
-        GpaCounterGroupDesc{471, "RMI13", 13, 139, 2, 4},
-        GpaCounterGroupDesc{472, "RMI14", 14, 139, 2, 4},
-        GpaCounterGroupDesc{473, "RMI15", 15, 139, 2, 4},
-        GpaCounterGroupDesc{474, "RMI16", 16, 139, 2, 4},
-        GpaCounterGroupDesc{475, "RMI17", 17, 139, 2, 4},
-        GpaCounterGroupDesc{476, "RMI18", 18, 139, 2, 4},
-        GpaCounterGroupDesc{477, "RMI19", 19, 139, 2, 4},
-        GpaCounterGroupDesc{478, "RMI20", 20, 139, 2, 4},
-        GpaCounterGroupDesc{479, "RMI21", 21, 139, 2, 4},
-        GpaCounterGroupDesc{480, "RMI22", 22, 139, 2, 4},
-        GpaCounterGroupDesc{481, "RMI23", 23, 139, 2, 4},
-        GpaCounterGroupDesc{482, "RMI24", 24, 139, 2, 4},
-        GpaCounterGroupDesc{483, "RMI25", 25, 139, 2, 4},
-        GpaCounterGroupDesc{484, "RMI26", 26, 139, 2, 4},
-        GpaCounterGroupDesc{485, "RMI27", 27, 139, 2, 4},
-        GpaCounterGroupDesc{486, "RMI28", 28, 139, 2, 4},
-        GpaCounterGroupDesc{487, "RMI29", 29, 139, 2, 4},
-        GpaCounterGroupDesc{488, "RMI30", 30, 139, 2, 4},
-        GpaCounterGroupDesc{489, "RMI31", 31, 139, 2, 4},
-        GpaCounterGroupDesc{490, "RMI32", 32, 139, 2, 4},
-        GpaCounterGroupDesc{491, "RMI33", 33, 139, 2, 4},
-        GpaCounterGroupDesc{492, "RMI34", 34, 139, 2, 4},
-        GpaCounterGroupDesc{493, "RMI35", 35, 139, 2, 4},
-        GpaCounterGroupDesc{494, "RMI36", 36, 139, 2, 4},
-        GpaCounterGroupDesc{495, "RMI37", 37, 139, 2, 4},
-        GpaCounterGroupDesc{496, "RMI38", 38, 139, 2, 4},
-        GpaCounterGroupDesc{497, "RMI39", 39, 139, 2, 4},
-        GpaCounterGroupDesc{498, "RMI40", 40, 139, 2, 4},
-        GpaCounterGroupDesc{499, "RMI41", 41, 139, 2, 4},
-        GpaCounterGroupDesc{500, "RMI42", 42, 139, 2, 4},
-        GpaCounterGroupDesc{501, "RMI43", 43, 139, 2, 4},
-        GpaCounterGroupDesc{502, "RMI44", 44, 139, 2, 4},
-        GpaCounterGroupDesc{503, "RMI45", 45, 139, 2, 4},
-        GpaCounterGroupDesc{504, "RMI46", 46, 139, 2, 4},
-        GpaCounterGroupDesc{505, "RMI47", 47, 139, 2, 4},
-        GpaCounterGroupDesc{506, "UMC0", 0, 40, 5, 0},
-        GpaCounterGroupDesc{507, "UMC1", 1, 40, 5, 0},
-        GpaCounterGroupDesc{508, "UMC2", 2, 40, 5, 0},
-        GpaCounterGroupDesc{509, "UMC3", 3, 40, 5, 0},
-        GpaCounterGroupDesc{510, "UMC4", 4, 40, 5, 0},
-        GpaCounterGroupDesc{511, "UMC5", 5, 40, 5, 0},
-        GpaCounterGroupDesc{512, "UMC6", 6, 40, 5, 0},
-        GpaCounterGroupDesc{513, "UMC7", 7, 40, 5, 0},
-        GpaCounterGroupDesc{514, "UMC8", 8, 40, 5, 0},
-        GpaCounterGroupDesc{515, "UMC9", 9, 40, 5, 0},
-        GpaCounterGroupDesc{516, "UMC10", 10, 40, 5, 0},
-        GpaCounterGroupDesc{517, "UMC11", 11, 40, 5, 0},
-        GpaCounterGroupDesc{518, "UMC12", 12, 40, 5, 0},
-        GpaCounterGroupDesc{519, "UMC13", 13, 40, 5, 0},
-        GpaCounterGroupDesc{520, "UMC14", 14, 40, 5, 0},
-        GpaCounterGroupDesc{521, "UMC15", 15, 40, 5, 0},
-        GpaCounterGroupDesc{522, "UMC16", 16, 40, 5, 0},
-        GpaCounterGroupDesc{523, "UMC17", 17, 40, 5, 0},
-        GpaCounterGroupDesc{524, "UMC18", 18, 40, 5, 0},
-        GpaCounterGroupDesc{525, "UMC19", 19, 40, 5, 0},
-        GpaCounterGroupDesc{526, "UMC20", 20, 40, 5, 0},
-        GpaCounterGroupDesc{527, "UMC21", 21, 40, 5, 0},
-        GpaCounterGroupDesc{528, "UMC22", 22, 40, 5, 0},
-        GpaCounterGroupDesc{529, "UMC23", 23, 40, 5, 0},
-        GpaCounterGroupDesc{530, "GE", 0, 40, 4, 16},
-        GpaCounterGroupDesc{531, "GL1A0", 0, 24, 4, 4},
-        GpaCounterGroupDesc{532, "GL1A1", 1, 24, 4, 4},
-        GpaCounterGroupDesc{533, "GL1A2", 2, 24, 4, 4},
-        GpaCounterGroupDesc{534, "GL1A3", 3, 24, 4, 4},
-        GpaCounterGroupDesc{535, "GL1A4", 4, 24, 4, 4},
-        GpaCounterGroupDesc{536, "GL1A5", 5, 24, 4, 4},
-        GpaCounterGroupDesc{537, "GL1A6", 6, 24, 4, 4},
-        GpaCounterGroupDesc{538, "GL1A7", 7, 24, 4, 4},
-        GpaCounterGroupDesc{539, "GL1A8", 8, 24, 4, 4},
-        GpaCounterGroupDesc{540, "GL1A9", 9, 24, 4, 4},
-        GpaCounterGroupDesc{541, "GL1A10", 10, 24, 4, 4},
-        GpaCounterGroupDesc{542, "GL1A11", 11, 24, 4, 4},
-        GpaCounterGroupDesc{543, "GL1C0", 0, 84, 4, 4},
-        GpaCounterGroupDesc{544, "GL1C1", 1, 84, 4, 4},
-        GpaCounterGroupDesc{545, "GL1C2", 2, 84, 4, 4},
-        GpaCounterGroupDesc{546, "GL1C3", 3, 84, 4, 4},
-        GpaCounterGroupDesc{547, "GL1C4", 4, 84, 4, 4},
-        GpaCounterGroupDesc{548, "GL1C5", 5, 84, 4, 4},
-        GpaCounterGroupDesc{549, "GL1C6", 6, 84, 4, 4},
-        GpaCounterGroupDesc{550, "GL1C7", 7, 84, 4, 4},
-        GpaCounterGroupDesc{551, "GL1C8", 8, 84, 4, 4},
-        GpaCounterGroupDesc{552, "GL1C9", 9, 84, 4, 4},
-        GpaCounterGroupDesc{553, "GL1C10", 10, 84, 4, 4},
-        GpaCounterGroupDesc{554, "GL1C11", 11, 84, 4, 4},
-        GpaCounterGroupDesc{555, "GL1C12", 12, 84, 4, 4},
-        GpaCounterGroupDesc{556, "GL1C13", 13, 84, 4, 4},
-        GpaCounterGroupDesc{557, "GL1C14", 14, 84, 4, 4},
-        GpaCounterGroupDesc{558, "GL1C15", 15, 84, 4, 4},
-        GpaCounterGroupDesc{559, "GL1C16", 16, 84, 4, 4},
-        GpaCounterGroupDesc{560, "GL1C17", 17, 84, 4, 4},
-        GpaCounterGroupDesc{561, "GL1C18", 18, 84, 4, 4},
-        GpaCounterGroupDesc{562, "GL1C19", 19, 84, 4, 4},
-        GpaCounterGroupDesc{563, "GL1C20", 20, 84, 4, 4},
-        GpaCounterGroupDesc{564, "GL1C21", 21, 84, 4, 4},
-        GpaCounterGroupDesc{565, "GL1C22", 22, 84, 4, 4},
-        GpaCounterGroupDesc{566, "GL1C23", 23, 84, 4, 4},
-        GpaCounterGroupDesc{567, "GL1C24", 24, 84, 4, 4},
-        GpaCounterGroupDesc{568, "GL1C25", 25, 84, 4, 4},
-        GpaCounterGroupDesc{569, "GL1C26", 26, 84, 4, 4},
-        GpaCounterGroupDesc{570, "GL1C27", 27, 84, 4, 4},
-        GpaCounterGroupDesc{571, "GL1C28", 28, 84, 4, 4},
-        GpaCounterGroupDesc{572, "GL1C29", 29, 84, 4, 4},
-        GpaCounterGroupDesc{573, "GL1C30", 30, 84, 4, 4},
-        GpaCounterGroupDesc{574, "GL1C31", 31, 84, 4, 4},
-        GpaCounterGroupDesc{575, "GL1C32", 32, 84, 4, 4},
-        GpaCounterGroupDesc{576, "GL1C33", 33, 84, 4, 4},
-        GpaCounterGroupDesc{577, "GL1C34", 34, 84, 4, 4},
-        GpaCounterGroupDesc{578, "GL1C35", 35, 84, 4, 4},
-        GpaCounterGroupDesc{579, "GL1C36", 36, 84, 4, 4},
-        GpaCounterGroupDesc{580, "GL1C37", 37, 84, 4, 4},
-        GpaCounterGroupDesc{581, "GL1C38", 38, 84, 4, 4},
-        GpaCounterGroupDesc{582, "GL1C39", 39, 84, 4, 4},
-        GpaCounterGroupDesc{583, "GL1C40", 40, 84, 4, 4},
-        GpaCounterGroupDesc{584, "GL1C41", 41, 84, 4, 4},
-        GpaCounterGroupDesc{585, "GL1C42", 42, 84, 4, 4},
-        GpaCounterGroupDesc{586, "GL1C43", 43, 84, 4, 4},
-        GpaCounterGroupDesc{587, "GL1C44", 44, 84, 4, 4},
-        GpaCounterGroupDesc{588, "GL1C45", 45, 84, 4, 4},
-        GpaCounterGroupDesc{589, "GL1C46", 46, 84, 4, 4},
-        GpaCounterGroupDesc{590, "GL1C47", 47, 84, 4, 4},
-        GpaCounterGroupDesc{591, "GL2A0", 0, 108, 4, 8},
-        GpaCounterGroupDesc{592, "GL2A1", 1, 108, 4, 8},
-        GpaCounterGroupDesc{593, "GL2A2", 2, 108, 4, 8},
-        GpaCounterGroupDesc{594, "GL2A3", 3, 108, 4, 8},
-        GpaCounterGroupDesc{595, "GL2C0", 0, 259, 4, 8},
-        GpaCounterGroupDesc{596, "GL2C1", 1, 259, 4, 8},
-        GpaCounterGroupDesc{597, "GL2C2", 2, 259, 4, 8},
-        GpaCounterGroupDesc{598, "GL2C3", 3, 259, 4, 8},
-        GpaCounterGroupDesc{599, "GL2C4", 4, 259, 4, 8},
-        GpaCounterGroupDesc{600, "GL2C5", 5, 259, 4, 8},
-        GpaCounterGroupDesc{601, "GL2C6", 6, 259, 4, 8},
-        GpaCounterGroupDesc{602, "GL2C7", 7, 259, 4, 8},
-        GpaCounterGroupDesc{603, "GL2C8", 8, 259, 4, 8},
-        GpaCounterGroupDesc{604, "GL2C9", 9, 259, 4, 8},
-        GpaCounterGroupDesc{605, "GL2C10", 10, 259, 4, 8},
-        GpaCounterGroupDesc{606, "GL2C11", 11, 259, 4, 8},
-        GpaCounterGroupDesc{607, "GL2C12", 12, 259, 4, 8},
-        GpaCounterGroupDesc{608, "GL2C13", 13, 259, 4, 8},
-        GpaCounterGroupDesc{609, "GL2C14", 14, 259, 4, 8},
-        GpaCounterGroupDesc{610, "GL2C15", 15, 259, 4, 8},
-        GpaCounterGroupDesc{611, "GL2C16", 16, 259, 4, 8},
-        GpaCounterGroupDesc{612, "GL2C17", 17, 259, 4, 8},
-        GpaCounterGroupDesc{613, "GL2C18", 18, 259, 4, 8},
-        GpaCounterGroupDesc{614, "GL2C19", 19, 259, 4, 8},
-        GpaCounterGroupDesc{615, "GL2C20", 20, 259, 4, 8},
-        GpaCounterGroupDesc{616, "GL2C21", 21, 259, 4, 8},
-        GpaCounterGroupDesc{617, "GL2C22", 22, 259, 4, 8},
-        GpaCounterGroupDesc{618, "GL2C23", 23, 259, 4, 8},
-        GpaCounterGroupDesc{619, "CHA", 0, 40, 4, 4},
-        GpaCounterGroupDesc{620, "CHC0", 0, 44, 4, 4},
-        GpaCounterGroupDesc{621, "CHC1", 1, 44, 4, 4},
-        GpaCounterGroupDesc{622, "CHC2", 2, 44, 4, 4},
-        GpaCounterGroupDesc{623, "CHC3", 3, 44, 4, 4},
-        GpaCounterGroupDesc{624, "CHCG", 0, 44, 4, 4},
-        GpaCounterGroupDesc{625, "GUS", 0, 176, 2, 4},
-        GpaCounterGroupDesc{626, "GCR", 0, 155, 2, 4},
-        GpaCounterGroupDesc{627, "PA_PH", 0, 1024, 8, 16},
-        GpaCounterGroupDesc{628, "UTCL10", 0, 66, 2, 0},
-        GpaCounterGroupDesc{629, "UTCL11", 1, 66, 2, 0},
-        GpaCounterGroupDesc{630, "UTCL12", 2, 66, 2, 0},
-        GpaCounterGroupDesc{631, "UTCL13", 3, 66, 2, 0},
-        GpaCounterGroupDesc{632, "UTCL14", 4, 66, 2, 0},
-        GpaCounterGroupDesc{633, "UTCL15", 5, 66, 2, 0},
-        GpaCounterGroupDesc{634, "UTCL16", 6, 66, 2, 0},
-        GpaCounterGroupDesc{635, "UTCL17", 7, 66, 2, 0},
-        GpaCounterGroupDesc{636, "UTCL18", 8, 66, 2, 0},
-        GpaCounterGroupDesc{637, "UTCL19", 9, 66, 2, 0},
-        GpaCounterGroupDesc{638, "UTCL110", 10, 66, 2, 0},
-        GpaCounterGroupDesc{639, "UTCL111", 11, 66, 2, 0},
-        GpaCounterGroupDesc{640, "GE2_DIST", 0, 101, 4, 16},
-        GpaCounterGroupDesc{641, "GE_SE0", 0, 88, 4, 16},
-        GpaCounterGroupDesc{642, "GE_SE1", 1, 88, 4, 16},
-        GpaCounterGroupDesc{643, "GE_SE2", 2, 88, 4, 16},
-        GpaCounterGroupDesc{644, "GE_SE3", 3, 88, 4, 16},
-        GpaCounterGroupDesc{645, "GE_SE4", 4, 88, 4, 16},
-        GpaCounterGroupDesc{646, "GE_SE5", 5, 88, 4, 16},
-        GpaCounterGroupDesc{647, "SQWGP0", 0, 512, 8, 16},
-        GpaCounterGroupDesc{648, "SQWGP1", 1, 512, 8, 16},
-        GpaCounterGroupDesc{649, "SQWGP2", 2, 512, 8, 16},
-        GpaCounterGroupDesc{650, "SQWGP3", 3, 512, 8, 16},
-        GpaCounterGroupDesc{651, "SQWGP4", 4, 512, 8, 16},
-        GpaCounterGroupDesc{652, "SQWGP5", 5, 512, 8, 16},
-        GpaCounterGroupDesc{653, "SQWGP6", 6, 512, 8, 16},
-        GpaCounterGroupDesc{654, "SQWGP7", 7, 512, 8, 16},
-        GpaCounterGroupDesc{655, "SQWGP8", 8, 512, 8, 16},
-        GpaCounterGroupDesc{656, "SQWGP9", 9, 512, 8, 16},
-        GpaCounterGroupDesc{657, "SQWGP10", 10, 512, 8, 16},
-        GpaCounterGroupDesc{658, "SQWGP11", 11, 512, 8, 16},
-        GpaCounterGroupDesc{659, "SQWGP12", 12, 512, 8, 16},
-        GpaCounterGroupDesc{660, "SQWGP13", 13, 512, 8, 16},
-        GpaCounterGroupDesc{661, "SQWGP14", 14, 512, 8, 16},
-        GpaCounterGroupDesc{662, "SQWGP15", 15, 512, 8, 16},
-        GpaCounterGroupDesc{663, "SQWGP16", 16, 512, 8, 16},
-        GpaCounterGroupDesc{664, "SQWGP17", 17, 512, 8, 16},
-        GpaCounterGroupDesc{665, "SQWGP18", 18, 512, 8, 16},
-        GpaCounterGroupDesc{666, "SQWGP19", 19, 512, 8, 16},
-        GpaCounterGroupDesc{667, "SQWGP20", 20, 512, 8, 16},
-        GpaCounterGroupDesc{668, "SQWGP21", 21, 512, 8, 16},
-        GpaCounterGroupDesc{669, "SQWGP22", 22, 512, 8, 16},
-        GpaCounterGroupDesc{670, "SQWGP23", 23, 512, 8, 16},
-        GpaCounterGroupDesc{671, "SQWGP24", 24, 512, 8, 16},
-        GpaCounterGroupDesc{672, "SQWGP25", 25, 512, 8, 16},
-        GpaCounterGroupDesc{673, "SQWGP26", 26, 512, 8, 16},
-        GpaCounterGroupDesc{674, "SQWGP27", 27, 512, 8, 16},
-        GpaCounterGroupDesc{675, "SQWGP28", 28, 512, 8, 16},
-        GpaCounterGroupDesc{676, "SQWGP29", 29, 512, 8, 16},
-        GpaCounterGroupDesc{677, "SQWGP30", 30, 512, 8, 16},
-        GpaCounterGroupDesc{678, "SQWGP31", 31, 512, 8, 16},
-        GpaCounterGroupDesc{679, "SQWGP32", 32, 512, 8, 16},
-        GpaCounterGroupDesc{680, "SQWGP33", 33, 512, 8, 16},
-        GpaCounterGroupDesc{681, "SQWGP34", 34, 512, 8, 16},
-        GpaCounterGroupDesc{682, "SQWGP35", 35, 512, 8, 16},
-        GpaCounterGroupDesc{683, "SQWGP36", 36, 512, 8, 16},
-        GpaCounterGroupDesc{684, "SQWGP37", 37, 512, 8, 16},
-        GpaCounterGroupDesc{685, "SQWGP38", 38, 512, 8, 16},
-        GpaCounterGroupDesc{686, "SQWGP39", 39, 512, 8, 16},
-        GpaCounterGroupDesc{687, "SQWGP40", 40, 512, 8, 16},
-        GpaCounterGroupDesc{688, "SQWGP41", 41, 512, 8, 16},
-        GpaCounterGroupDesc{689, "SQWGP42", 42, 512, 8, 16},
-        GpaCounterGroupDesc{690, "SQWGP43", 43, 512, 8, 16},
-        GpaCounterGroupDesc{691, "SQWGP44", 44, 512, 8, 16},
-        GpaCounterGroupDesc{692, "SQWGP45", 45, 512, 8, 16},
-        GpaCounterGroupDesc{693, "SQWGP46", 46, 512, 8, 16},
-        GpaCounterGroupDesc{694, "SQWGP47", 47, 512, 8, 16},
-        GpaCounterGroupDesc{695, "SQWGP_GS0", 0, 512, 8, 16},
-        GpaCounterGroupDesc{696, "SQWGP_GS1", 1, 512, 8, 16},
-        GpaCounterGroupDesc{697, "SQWGP_GS2", 2, 512, 8, 16},
-        GpaCounterGroupDesc{698, "SQWGP_GS3", 3, 512, 8, 16},
-        GpaCounterGroupDesc{699, "SQWGP_GS4", 4, 512, 8, 16},
-        GpaCounterGroupDesc{700, "SQWGP_GS5", 5, 512, 8, 16},
-        GpaCounterGroupDesc{701, "SQWGP_GS6", 6, 512, 8, 16},
-        GpaCounterGroupDesc{702, "SQWGP_GS7", 7, 512, 8, 16},
-        GpaCounterGroupDesc{703, "SQWGP_GS8", 8, 512, 8, 16},
-        GpaCounterGroupDesc{704, "SQWGP_GS9", 9, 512, 8, 16},
-        GpaCounterGroupDesc{705, "SQWGP_GS10", 10, 512, 8, 16},
-        GpaCounterGroupDesc{706, "SQWGP_GS11", 11, 512, 8, 16},
-        GpaCounterGroupDesc{707, "SQWGP_GS12", 12, 512, 8, 16},
-        GpaCounterGroupDesc{708, "SQWGP_GS13", 13, 512, 8, 16},
-        GpaCounterGroupDesc{709, "SQWGP_GS14", 14, 512, 8, 16},
-        GpaCounterGroupDesc{710, "SQWGP_GS15", 15, 512, 8, 16},
-        GpaCounterGroupDesc{711, "SQWGP_GS16", 16, 512, 8, 16},
-        GpaCounterGroupDesc{712, "SQWGP_GS17", 17, 512, 8, 16},
-        GpaCounterGroupDesc{713, "SQWGP_GS18", 18, 512, 8, 16},
-        GpaCounterGroupDesc{714, "SQWGP_GS19", 19, 512, 8, 16},
-        GpaCounterGroupDesc{715, "SQWGP_GS20", 20, 512, 8, 16},
-        GpaCounterGroupDesc{716, "SQWGP_GS21", 21, 512, 8, 16},
-        GpaCounterGroupDesc{717, "SQWGP_GS22", 22, 512, 8, 16},
-        GpaCounterGroupDesc{718, "SQWGP_GS23", 23, 512, 8, 16},
-        GpaCounterGroupDesc{719, "SQWGP_GS24", 24, 512, 8, 16},
-        GpaCounterGroupDesc{720, "SQWGP_GS25", 25, 512, 8, 16},
-        GpaCounterGroupDesc{721, "SQWGP_GS26", 26, 512, 8, 16},
-        GpaCounterGroupDesc{722, "SQWGP_GS27", 27, 512, 8, 16},
-        GpaCounterGroupDesc{723, "SQWGP_GS28", 28, 512, 8, 16},
-        GpaCounterGroupDesc{724, "SQWGP_GS29", 29, 512, 8, 16},
-        GpaCounterGroupDesc{725, "SQWGP_GS30", 30, 512, 8, 16},
-        GpaCounterGroupDesc{726, "SQWGP_GS31", 31, 512, 8, 16},
-        GpaCounterGroupDesc{727, "SQWGP_GS32", 32, 512, 8, 16},
-        GpaCounterGroupDesc{728, "SQWGP_GS33", 33, 512, 8, 16},
-        GpaCounterGroupDesc{729, "SQWGP_GS34", 34, 512, 8, 16},
-        GpaCounterGroupDesc{730, "SQWGP_GS35", 35, 512, 8, 16},
-        GpaCounterGroupDesc{731, "SQWGP_GS36", 36, 512, 8, 16},
-        GpaCounterGroupDesc{732, "SQWGP_GS37", 37, 512, 8, 16},
-        GpaCounterGroupDesc{733, "SQWGP_GS38", 38, 512, 8, 16},
-        GpaCounterGroupDesc{734, "SQWGP_GS39", 39, 512, 8, 16},
-        GpaCounterGroupDesc{735, "SQWGP_GS40", 40, 512, 8, 16},
-        GpaCounterGroupDesc{736, "SQWGP_GS41", 41, 512, 8, 16},
-        GpaCounterGroupDesc{737, "SQWGP_GS42", 42, 512, 8, 16},
-        GpaCounterGroupDesc{738, "SQWGP_GS43", 43, 512, 8, 16},
-        GpaCounterGroupDesc{739, "SQWGP_GS44", 44, 512, 8, 16},
-        GpaCounterGroupDesc{740, "SQWGP_GS45", 45, 512, 8, 16},
-        GpaCounterGroupDesc{741, "SQWGP_GS46", 46, 512, 8, 16},
-        GpaCounterGroupDesc{742, "SQWGP_GS47", 47, 512, 8, 16},
-        GpaCounterGroupDesc{743, "SQWGP_PS0", 0, 512, 8, 16},
-        GpaCounterGroupDesc{744, "SQWGP_PS1", 1, 512, 8, 16},
-        GpaCounterGroupDesc{745, "SQWGP_PS2", 2, 512, 8, 16},
-        GpaCounterGroupDesc{746, "SQWGP_PS3", 3, 512, 8, 16},
-        GpaCounterGroupDesc{747, "SQWGP_PS4", 4, 512, 8, 16},
-        GpaCounterGroupDesc{748, "SQWGP_PS5", 5, 512, 8, 16},
-        GpaCounterGroupDesc{749, "SQWGP_PS6", 6, 512, 8, 16},
-        GpaCounterGroupDesc{750, "SQWGP_PS7", 7, 512, 8, 16},
-        GpaCounterGroupDesc{751, "SQWGP_PS8", 8, 512, 8, 16},
-        GpaCounterGroupDesc{752, "SQWGP_PS9", 9, 512, 8, 16},
-        GpaCounterGroupDesc{753, "SQWGP_PS10", 10, 512, 8, 16},
-        GpaCounterGroupDesc{754, "SQWGP_PS11", 11, 512, 8, 16},
-        GpaCounterGroupDesc{755, "SQWGP_PS12", 12, 512, 8, 16},
-        GpaCounterGroupDesc{756, "SQWGP_PS13", 13, 512, 8, 16},
-        GpaCounterGroupDesc{757, "SQWGP_PS14", 14, 512, 8, 16},
-        GpaCounterGroupDesc{758, "SQWGP_PS15", 15, 512, 8, 16},
-        GpaCounterGroupDesc{759, "SQWGP_PS16", 16, 512, 8, 16},
-        GpaCounterGroupDesc{760, "SQWGP_PS17", 17, 512, 8, 16},
-        GpaCounterGroupDesc{761, "SQWGP_PS18", 18, 512, 8, 16},
-        GpaCounterGroupDesc{762, "SQWGP_PS19", 19, 512, 8, 16},
-        GpaCounterGroupDesc{763, "SQWGP_PS20", 20, 512, 8, 16},
-        GpaCounterGroupDesc{764, "SQWGP_PS21", 21, 512, 8, 16},
-        GpaCounterGroupDesc{765, "SQWGP_PS22", 22, 512, 8, 16},
-        GpaCounterGroupDesc{766, "SQWGP_PS23", 23, 512, 8, 16},
-        GpaCounterGroupDesc{767, "SQWGP_PS24", 24, 512, 8, 16},
-        GpaCounterGroupDesc{768, "SQWGP_PS25", 25, 512, 8, 16},
-        GpaCounterGroupDesc{769, "SQWGP_PS26", 26, 512, 8, 16},
-        GpaCounterGroupDesc{770, "SQWGP_PS27", 27, 512, 8, 16},
-        GpaCounterGroupDesc{771, "SQWGP_PS28", 28, 512, 8, 16},
-        GpaCounterGroupDesc{772, "SQWGP_PS29", 29, 512, 8, 16},
-        GpaCounterGroupDesc{773, "SQWGP_PS30", 30, 512, 8, 16},
-        GpaCounterGroupDesc{774, "SQWGP_PS31", 31, 512, 8, 16},
-        GpaCounterGroupDesc{775, "SQWGP_PS32", 32, 512, 8, 16},
-        GpaCounterGroupDesc{776, "SQWGP_PS33", 33, 512, 8, 16},
-        GpaCounterGroupDesc{777, "SQWGP_PS34", 34, 512, 8, 16},
-        GpaCounterGroupDesc{778, "SQWGP_PS35", 35, 512, 8, 16},
-        GpaCounterGroupDesc{779, "SQWGP_PS36", 36, 512, 8, 16},
-        GpaCounterGroupDesc{780, "SQWGP_PS37", 37, 512, 8, 16},
-        GpaCounterGroupDesc{781, "SQWGP_PS38", 38, 512, 8, 16},
-        GpaCounterGroupDesc{782, "SQWGP_PS39", 39, 512, 8, 16},
-        GpaCounterGroupDesc{783, "SQWGP_PS40", 40, 512, 8, 16},
-        GpaCounterGroupDesc{784, "SQWGP_PS41", 41, 512, 8, 16},
-        GpaCounterGroupDesc{785, "SQWGP_PS42", 42, 512, 8, 16},
-        GpaCounterGroupDesc{786, "SQWGP_PS43", 43, 512, 8, 16},
-        GpaCounterGroupDesc{787, "SQWGP_PS44", 44, 512, 8, 16},
-        GpaCounterGroupDesc{788, "SQWGP_PS45", 45, 512, 8, 16},
-        GpaCounterGroupDesc{789, "SQWGP_PS46", 46, 512, 8, 16},
-        GpaCounterGroupDesc{790, "SQWGP_PS47", 47, 512, 8, 16},
-        GpaCounterGroupDesc{791, "SQWGP_HS0", 0, 512, 8, 16},
-        GpaCounterGroupDesc{792, "SQWGP_HS1", 1, 512, 8, 16},
-        GpaCounterGroupDesc{793, "SQWGP_HS2", 2, 512, 8, 16},
-        GpaCounterGroupDesc{794, "SQWGP_HS3", 3, 512, 8, 16},
-        GpaCounterGroupDesc{795, "SQWGP_HS4", 4, 512, 8, 16},
-        GpaCounterGroupDesc{796, "SQWGP_HS5", 5, 512, 8, 16},
-        GpaCounterGroupDesc{797, "SQWGP_HS6", 6, 512, 8, 16},
-        GpaCounterGroupDesc{798, "SQWGP_HS7", 7, 512, 8, 16},
-        GpaCounterGroupDesc{799, "SQWGP_HS8", 8, 512, 8, 16},
-        GpaCounterGroupDesc{800, "SQWGP_HS9", 9, 512, 8, 16},
-        GpaCounterGroupDesc{801, "SQWGP_HS10", 10, 512, 8, 16},
-        GpaCounterGroupDesc{802, "SQWGP_HS11", 11, 512, 8, 16},
-        GpaCounterGroupDesc{803, "SQWGP_HS12", 12, 512, 8, 16},
-        GpaCounterGroupDesc{804, "SQWGP_HS13", 13, 512, 8, 16},
-        GpaCounterGroupDesc{805, "SQWGP_HS14", 14, 512, 8, 16},
-        GpaCounterGroupDesc{806, "SQWGP_HS15", 15, 512, 8, 16},
-        GpaCounterGroupDesc{807, "SQWGP_HS16", 16, 512, 8, 16},
-        GpaCounterGroupDesc{808, "SQWGP_HS17", 17, 512, 8, 16},
-        GpaCounterGroupDesc{809, "SQWGP_HS18", 18, 512, 8, 16},
-        GpaCounterGroupDesc{810, "SQWGP_HS19", 19, 512, 8, 16},
-        GpaCounterGroupDesc{811, "SQWGP_HS20", 20, 512, 8, 16},
-        GpaCounterGroupDesc{812, "SQWGP_HS21", 21, 512, 8, 16},
-        GpaCounterGroupDesc{813, "SQWGP_HS22", 22, 512, 8, 16},
-        GpaCounterGroupDesc{814, "SQWGP_HS23", 23, 512, 8, 16},
-        GpaCounterGroupDesc{815, "SQWGP_HS24", 24, 512, 8, 16},
-        GpaCounterGroupDesc{816, "SQWGP_HS25", 25, 512, 8, 16},
-        GpaCounterGroupDesc{817, "SQWGP_HS26", 26, 512, 8, 16},
-        GpaCounterGroupDesc{818, "SQWGP_HS27", 27, 512, 8, 16},
-        GpaCounterGroupDesc{819, "SQWGP_HS28", 28, 512, 8, 16},
-        GpaCounterGroupDesc{820, "SQWGP_HS29", 29, 512, 8, 16},
-        GpaCounterGroupDesc{821, "SQWGP_HS30", 30, 512, 8, 16},
-        GpaCounterGroupDesc{822, "SQWGP_HS31", 31, 512, 8, 16},
-        GpaCounterGroupDesc{823, "SQWGP_HS32", 32, 512, 8, 16},
-        GpaCounterGroupDesc{824, "SQWGP_HS33", 33, 512, 8, 16},
-        GpaCounterGroupDesc{825, "SQWGP_HS34", 34, 512, 8, 16},
-        GpaCounterGroupDesc{826, "SQWGP_HS35", 35, 512, 8, 16},
-        GpaCounterGroupDesc{827, "SQWGP_HS36", 36, 512, 8, 16},
-        GpaCounterGroupDesc{828, "SQWGP_HS37", 37, 512, 8, 16},
-        GpaCounterGroupDesc{829, "SQWGP_HS38", 38, 512, 8, 16},
-        GpaCounterGroupDesc{830, "SQWGP_HS39", 39, 512, 8, 16},
-        GpaCounterGroupDesc{831, "SQWGP_HS40", 40, 512, 8, 16},
-        GpaCounterGroupDesc{832, "SQWGP_HS41", 41, 512, 8, 16},
-        GpaCounterGroupDesc{833, "SQWGP_HS42", 42, 512, 8, 16},
-        GpaCounterGroupDesc{834, "SQWGP_HS43", 43, 512, 8, 16},
-        GpaCounterGroupDesc{835, "SQWGP_HS44", 44, 512, 8, 16},
-        GpaCounterGroupDesc{836, "SQWGP_HS45", 45, 512, 8, 16},
-        GpaCounterGroupDesc{837, "SQWGP_HS46", 46, 512, 8, 16},
-        GpaCounterGroupDesc{838, "SQWGP_HS47", 47, 512, 8, 16},
-        GpaCounterGroupDesc{839, "SQWGP_CS0", 0, 512, 8, 16},
-        GpaCounterGroupDesc{840, "SQWGP_CS1", 1, 512, 8, 16},
-        GpaCounterGroupDesc{841, "SQWGP_CS2", 2, 512, 8, 16},
-        GpaCounterGroupDesc{842, "SQWGP_CS3", 3, 512, 8, 16},
-        GpaCounterGroupDesc{843, "SQWGP_CS4", 4, 512, 8, 16},
-        GpaCounterGroupDesc{844, "SQWGP_CS5", 5, 512, 8, 16},
-        GpaCounterGroupDesc{845, "SQWGP_CS6", 6, 512, 8, 16},
-        GpaCounterGroupDesc{846, "SQWGP_CS7", 7, 512, 8, 16},
-        GpaCounterGroupDesc{847, "SQWGP_CS8", 8, 512, 8, 16},
-        GpaCounterGroupDesc{848, "SQWGP_CS9", 9, 512, 8, 16},
-        GpaCounterGroupDesc{849, "SQWGP_CS10", 10, 512, 8, 16},
-        GpaCounterGroupDesc{850, "SQWGP_CS11", 11, 512, 8, 16},
-        GpaCounterGroupDesc{851, "SQWGP_CS12", 12, 512, 8, 16},
-        GpaCounterGroupDesc{852, "SQWGP_CS13", 13, 512, 8, 16},
-        GpaCounterGroupDesc{853, "SQWGP_CS14", 14, 512, 8, 16},
-        GpaCounterGroupDesc{854, "SQWGP_CS15", 15, 512, 8, 16},
-        GpaCounterGroupDesc{855, "SQWGP_CS16", 16, 512, 8, 16},
-        GpaCounterGroupDesc{856, "SQWGP_CS17", 17, 512, 8, 16},
-        GpaCounterGroupDesc{857, "SQWGP_CS18", 18, 512, 8, 16},
-        GpaCounterGroupDesc{858, "SQWGP_CS19", 19, 512, 8, 16},
-        GpaCounterGroupDesc{859, "SQWGP_CS20", 20, 512, 8, 16},
-        GpaCounterGroupDesc{860, "SQWGP_CS21", 21, 512, 8, 16},
-        GpaCounterGroupDesc{861, "SQWGP_CS22", 22, 512, 8, 16},
-        GpaCounterGroupDesc{862, "SQWGP_CS23", 23, 512, 8, 16},
-        GpaCounterGroupDesc{863, "SQWGP_CS24", 24, 512, 8, 16},
-        GpaCounterGroupDesc{864, "SQWGP_CS25", 25, 512, 8, 16},
-        GpaCounterGroupDesc{865, "SQWGP_CS26", 26, 512, 8, 16},
-        GpaCounterGroupDesc{866, "SQWGP_CS27", 27, 512, 8, 16},
-        GpaCounterGroupDesc{867, "SQWGP_CS28", 28, 512, 8, 16},
-        GpaCounterGroupDesc{868, "SQWGP_CS29", 29, 512, 8, 16},
-        GpaCounterGroupDesc{869, "SQWGP_CS30", 30, 512, 8, 16},
-        GpaCounterGroupDesc{870, "SQWGP_CS31", 31, 512, 8, 16},
-        GpaCounterGroupDesc{871, "SQWGP_CS32", 32, 512, 8, 16},
-        GpaCounterGroupDesc{872, "SQWGP_CS33", 33, 512, 8, 16},
-        GpaCounterGroupDesc{873, "SQWGP_CS34", 34, 512, 8, 16},
-        GpaCounterGroupDesc{874, "SQWGP_CS35", 35, 512, 8, 16},
-        GpaCounterGroupDesc{875, "SQWGP_CS36", 36, 512, 8, 16},
-        GpaCounterGroupDesc{876, "SQWGP_CS37", 37, 512, 8, 16},
-        GpaCounterGroupDesc{877, "SQWGP_CS38", 38, 512, 8, 16},
-        GpaCounterGroupDesc{878, "SQWGP_CS39", 39, 512, 8, 16},
-        GpaCounterGroupDesc{879, "SQWGP_CS40", 40, 512, 8, 16},
-        GpaCounterGroupDesc{880, "SQWGP_CS41", 41, 512, 8, 16},
-        GpaCounterGroupDesc{881, "SQWGP_CS42", 42, 512, 8, 16},
-        GpaCounterGroupDesc{882, "SQWGP_CS43", 43, 512, 8, 16},
-        GpaCounterGroupDesc{883, "SQWGP_CS44", 44, 512, 8, 16},
-        GpaCounterGroupDesc{884, "SQWGP_CS45", 45, 512, 8, 16},
-        GpaCounterGroupDesc{885, "SQWGP_CS46", 46, 512, 8, 16},
-        GpaCounterGroupDesc{886, "SQWGP_CS47", 47, 512, 8, 16},
-        GpaCounterGroupDesc{887, "PC0", 0, 47, 4, 0},
-        GpaCounterGroupDesc{888, "PC1", 1, 47, 4, 0},
-        GpaCounterGroupDesc{889, "PC2", 2, 47, 4, 0},
-        GpaCounterGroupDesc{890, "PC3", 3, 47, 4, 0},
-        GpaCounterGroupDesc{891, "PC4", 4, 47, 4, 0},
-        GpaCounterGroupDesc{892, "PC5", 5, 47, 4, 0},
-        GpaCounterGroupDesc{893, "GPUTime", 0, 6, 3, 0},
+        GpaCounterGroupDesc{432, "ATC", 0, 24, 0, 0},
+        GpaCounterGroupDesc{433, "ATCL2", 0, 18, 0, 0},
+        GpaCounterGroupDesc{434, "GCVML2", 0, 91, 8, 8},
+        GpaCounterGroupDesc{435, "GCEA0", 0, 98, 2, 4},
+        GpaCounterGroupDesc{436, "GCEA1", 1, 98, 2, 4},
+        GpaCounterGroupDesc{437, "GCEA2", 2, 98, 2, 4},
+        GpaCounterGroupDesc{438, "GCEA3", 3, 98, 2, 4},
+        GpaCounterGroupDesc{439, "GCEA4", 4, 98, 2, 4},
+        GpaCounterGroupDesc{440, "GCEA5", 5, 98, 2, 4},
+        GpaCounterGroupDesc{441, "GCEA6", 6, 98, 2, 4},
+        GpaCounterGroupDesc{442, "GCEA7", 7, 98, 2, 4},
+        GpaCounterGroupDesc{443, "GCEA8", 8, 98, 2, 4},
+        GpaCounterGroupDesc{444, "GCEA9", 9, 98, 2, 4},
+        GpaCounterGroupDesc{445, "GCEA10", 10, 98, 2, 4},
+        GpaCounterGroupDesc{446, "GCEA11", 11, 98, 2, 4},
+        GpaCounterGroupDesc{447, "GCEA12", 12, 98, 2, 4},
+        GpaCounterGroupDesc{448, "GCEA13", 13, 98, 2, 4},
+        GpaCounterGroupDesc{449, "GCEA14", 14, 98, 2, 4},
+        GpaCounterGroupDesc{450, "GCEA15", 15, 98, 2, 4},
+        GpaCounterGroupDesc{451, "GCEA16", 16, 98, 2, 4},
+        GpaCounterGroupDesc{452, "GCEA17", 17, 98, 2, 4},
+        GpaCounterGroupDesc{453, "GCEA18", 18, 98, 2, 4},
+        GpaCounterGroupDesc{454, "GCEA19", 19, 98, 2, 4},
+        GpaCounterGroupDesc{455, "GCEA20", 20, 98, 2, 4},
+        GpaCounterGroupDesc{456, "GCEA21", 21, 98, 2, 4},
+        GpaCounterGroupDesc{457, "GCEA22", 22, 98, 2, 4},
+        GpaCounterGroupDesc{458, "GCEA23", 23, 98, 2, 4},
+        GpaCounterGroupDesc{459, "RPB", 0, 64, 4, 0},
+        GpaCounterGroupDesc{460, "RMI0", 0, 139, 2, 4},
+        GpaCounterGroupDesc{461, "RMI1", 1, 139, 2, 4},
+        GpaCounterGroupDesc{462, "RMI2", 2, 139, 2, 4},
+        GpaCounterGroupDesc{463, "RMI3", 3, 139, 2, 4},
+        GpaCounterGroupDesc{464, "RMI4", 4, 139, 2, 4},
+        GpaCounterGroupDesc{465, "RMI5", 5, 139, 2, 4},
+        GpaCounterGroupDesc{466, "RMI6", 6, 139, 2, 4},
+        GpaCounterGroupDesc{467, "RMI7", 7, 139, 2, 4},
+        GpaCounterGroupDesc{468, "RMI8", 8, 139, 2, 4},
+        GpaCounterGroupDesc{469, "RMI9", 9, 139, 2, 4},
+        GpaCounterGroupDesc{470, "RMI10", 10, 139, 2, 4},
+        GpaCounterGroupDesc{471, "RMI11", 11, 139, 2, 4},
+        GpaCounterGroupDesc{472, "RMI12", 12, 139, 2, 4},
+        GpaCounterGroupDesc{473, "RMI13", 13, 139, 2, 4},
+        GpaCounterGroupDesc{474, "RMI14", 14, 139, 2, 4},
+        GpaCounterGroupDesc{475, "RMI15", 15, 139, 2, 4},
+        GpaCounterGroupDesc{476, "RMI16", 16, 139, 2, 4},
+        GpaCounterGroupDesc{477, "RMI17", 17, 139, 2, 4},
+        GpaCounterGroupDesc{478, "RMI18", 18, 139, 2, 4},
+        GpaCounterGroupDesc{479, "RMI19", 19, 139, 2, 4},
+        GpaCounterGroupDesc{480, "RMI20", 20, 139, 2, 4},
+        GpaCounterGroupDesc{481, "RMI21", 21, 139, 2, 4},
+        GpaCounterGroupDesc{482, "RMI22", 22, 139, 2, 4},
+        GpaCounterGroupDesc{483, "RMI23", 23, 139, 2, 4},
+        GpaCounterGroupDesc{484, "RMI24", 24, 139, 2, 4},
+        GpaCounterGroupDesc{485, "RMI25", 25, 139, 2, 4},
+        GpaCounterGroupDesc{486, "RMI26", 26, 139, 2, 4},
+        GpaCounterGroupDesc{487, "RMI27", 27, 139, 2, 4},
+        GpaCounterGroupDesc{488, "RMI28", 28, 139, 2, 4},
+        GpaCounterGroupDesc{489, "RMI29", 29, 139, 2, 4},
+        GpaCounterGroupDesc{490, "RMI30", 30, 139, 2, 4},
+        GpaCounterGroupDesc{491, "RMI31", 31, 139, 2, 4},
+        GpaCounterGroupDesc{492, "RMI32", 32, 139, 2, 4},
+        GpaCounterGroupDesc{493, "RMI33", 33, 139, 2, 4},
+        GpaCounterGroupDesc{494, "RMI34", 34, 139, 2, 4},
+        GpaCounterGroupDesc{495, "RMI35", 35, 139, 2, 4},
+        GpaCounterGroupDesc{496, "RMI36", 36, 139, 2, 4},
+        GpaCounterGroupDesc{497, "RMI37", 37, 139, 2, 4},
+        GpaCounterGroupDesc{498, "RMI38", 38, 139, 2, 4},
+        GpaCounterGroupDesc{499, "RMI39", 39, 139, 2, 4},
+        GpaCounterGroupDesc{500, "RMI40", 40, 139, 2, 4},
+        GpaCounterGroupDesc{501, "RMI41", 41, 139, 2, 4},
+        GpaCounterGroupDesc{502, "RMI42", 42, 139, 2, 4},
+        GpaCounterGroupDesc{503, "RMI43", 43, 139, 2, 4},
+        GpaCounterGroupDesc{504, "RMI44", 44, 139, 2, 4},
+        GpaCounterGroupDesc{505, "RMI45", 45, 139, 2, 4},
+        GpaCounterGroupDesc{506, "RMI46", 46, 139, 2, 4},
+        GpaCounterGroupDesc{507, "RMI47", 47, 139, 2, 4},
+        GpaCounterGroupDesc{508, "UMC0", 0, 40, 5, 0},
+        GpaCounterGroupDesc{509, "UMC1", 1, 40, 5, 0},
+        GpaCounterGroupDesc{510, "UMC2", 2, 40, 5, 0},
+        GpaCounterGroupDesc{511, "UMC3", 3, 40, 5, 0},
+        GpaCounterGroupDesc{512, "UMC4", 4, 40, 5, 0},
+        GpaCounterGroupDesc{513, "UMC5", 5, 40, 5, 0},
+        GpaCounterGroupDesc{514, "UMC6", 6, 40, 5, 0},
+        GpaCounterGroupDesc{515, "UMC7", 7, 40, 5, 0},
+        GpaCounterGroupDesc{516, "UMC8", 8, 40, 5, 0},
+        GpaCounterGroupDesc{517, "UMC9", 9, 40, 5, 0},
+        GpaCounterGroupDesc{518, "UMC10", 10, 40, 5, 0},
+        GpaCounterGroupDesc{519, "UMC11", 11, 40, 5, 0},
+        GpaCounterGroupDesc{520, "UMC12", 12, 40, 5, 0},
+        GpaCounterGroupDesc{521, "UMC13", 13, 40, 5, 0},
+        GpaCounterGroupDesc{522, "UMC14", 14, 40, 5, 0},
+        GpaCounterGroupDesc{523, "UMC15", 15, 40, 5, 0},
+        GpaCounterGroupDesc{524, "UMC16", 16, 40, 5, 0},
+        GpaCounterGroupDesc{525, "UMC17", 17, 40, 5, 0},
+        GpaCounterGroupDesc{526, "UMC18", 18, 40, 5, 0},
+        GpaCounterGroupDesc{527, "UMC19", 19, 40, 5, 0},
+        GpaCounterGroupDesc{528, "UMC20", 20, 40, 5, 0},
+        GpaCounterGroupDesc{529, "UMC21", 21, 40, 5, 0},
+        GpaCounterGroupDesc{530, "UMC22", 22, 40, 5, 0},
+        GpaCounterGroupDesc{531, "UMC23", 23, 40, 5, 0},
+        GpaCounterGroupDesc{532, "GE", 0, 40, 4, 16},
+        GpaCounterGroupDesc{533, "GL1A0", 0, 24, 4, 4},
+        GpaCounterGroupDesc{534, "GL1A1", 1, 24, 4, 4},
+        GpaCounterGroupDesc{535, "GL1A2", 2, 24, 4, 4},
+        GpaCounterGroupDesc{536, "GL1A3", 3, 24, 4, 4},
+        GpaCounterGroupDesc{537, "GL1A4", 4, 24, 4, 4},
+        GpaCounterGroupDesc{538, "GL1A5", 5, 24, 4, 4},
+        GpaCounterGroupDesc{539, "GL1A6", 6, 24, 4, 4},
+        GpaCounterGroupDesc{540, "GL1A7", 7, 24, 4, 4},
+        GpaCounterGroupDesc{541, "GL1A8", 8, 24, 4, 4},
+        GpaCounterGroupDesc{542, "GL1A9", 9, 24, 4, 4},
+        GpaCounterGroupDesc{543, "GL1A10", 10, 24, 4, 4},
+        GpaCounterGroupDesc{544, "GL1A11", 11, 24, 4, 4},
+        GpaCounterGroupDesc{545, "GL1C0", 0, 84, 4, 4},
+        GpaCounterGroupDesc{546, "GL1C1", 1, 84, 4, 4},
+        GpaCounterGroupDesc{547, "GL1C2", 2, 84, 4, 4},
+        GpaCounterGroupDesc{548, "GL1C3", 3, 84, 4, 4},
+        GpaCounterGroupDesc{549, "GL1C4", 4, 84, 4, 4},
+        GpaCounterGroupDesc{550, "GL1C5", 5, 84, 4, 4},
+        GpaCounterGroupDesc{551, "GL1C6", 6, 84, 4, 4},
+        GpaCounterGroupDesc{552, "GL1C7", 7, 84, 4, 4},
+        GpaCounterGroupDesc{553, "GL1C8", 8, 84, 4, 4},
+        GpaCounterGroupDesc{554, "GL1C9", 9, 84, 4, 4},
+        GpaCounterGroupDesc{555, "GL1C10", 10, 84, 4, 4},
+        GpaCounterGroupDesc{556, "GL1C11", 11, 84, 4, 4},
+        GpaCounterGroupDesc{557, "GL1C12", 12, 84, 4, 4},
+        GpaCounterGroupDesc{558, "GL1C13", 13, 84, 4, 4},
+        GpaCounterGroupDesc{559, "GL1C14", 14, 84, 4, 4},
+        GpaCounterGroupDesc{560, "GL1C15", 15, 84, 4, 4},
+        GpaCounterGroupDesc{561, "GL1C16", 16, 84, 4, 4},
+        GpaCounterGroupDesc{562, "GL1C17", 17, 84, 4, 4},
+        GpaCounterGroupDesc{563, "GL1C18", 18, 84, 4, 4},
+        GpaCounterGroupDesc{564, "GL1C19", 19, 84, 4, 4},
+        GpaCounterGroupDesc{565, "GL1C20", 20, 84, 4, 4},
+        GpaCounterGroupDesc{566, "GL1C21", 21, 84, 4, 4},
+        GpaCounterGroupDesc{567, "GL1C22", 22, 84, 4, 4},
+        GpaCounterGroupDesc{568, "GL1C23", 23, 84, 4, 4},
+        GpaCounterGroupDesc{569, "GL1C24", 24, 84, 4, 4},
+        GpaCounterGroupDesc{570, "GL1C25", 25, 84, 4, 4},
+        GpaCounterGroupDesc{571, "GL1C26", 26, 84, 4, 4},
+        GpaCounterGroupDesc{572, "GL1C27", 27, 84, 4, 4},
+        GpaCounterGroupDesc{573, "GL1C28", 28, 84, 4, 4},
+        GpaCounterGroupDesc{574, "GL1C29", 29, 84, 4, 4},
+        GpaCounterGroupDesc{575, "GL1C30", 30, 84, 4, 4},
+        GpaCounterGroupDesc{576, "GL1C31", 31, 84, 4, 4},
+        GpaCounterGroupDesc{577, "GL1C32", 32, 84, 4, 4},
+        GpaCounterGroupDesc{578, "GL1C33", 33, 84, 4, 4},
+        GpaCounterGroupDesc{579, "GL1C34", 34, 84, 4, 4},
+        GpaCounterGroupDesc{580, "GL1C35", 35, 84, 4, 4},
+        GpaCounterGroupDesc{581, "GL1C36", 36, 84, 4, 4},
+        GpaCounterGroupDesc{582, "GL1C37", 37, 84, 4, 4},
+        GpaCounterGroupDesc{583, "GL1C38", 38, 84, 4, 4},
+        GpaCounterGroupDesc{584, "GL1C39", 39, 84, 4, 4},
+        GpaCounterGroupDesc{585, "GL1C40", 40, 84, 4, 4},
+        GpaCounterGroupDesc{586, "GL1C41", 41, 84, 4, 4},
+        GpaCounterGroupDesc{587, "GL1C42", 42, 84, 4, 4},
+        GpaCounterGroupDesc{588, "GL1C43", 43, 84, 4, 4},
+        GpaCounterGroupDesc{589, "GL1C44", 44, 84, 4, 4},
+        GpaCounterGroupDesc{590, "GL1C45", 45, 84, 4, 4},
+        GpaCounterGroupDesc{591, "GL1C46", 46, 84, 4, 4},
+        GpaCounterGroupDesc{592, "GL1C47", 47, 84, 4, 4},
+        GpaCounterGroupDesc{593, "GL2A0", 0, 108, 4, 8},
+        GpaCounterGroupDesc{594, "GL2A1", 1, 108, 4, 8},
+        GpaCounterGroupDesc{595, "GL2A2", 2, 108, 4, 8},
+        GpaCounterGroupDesc{596, "GL2A3", 3, 108, 4, 8},
+        GpaCounterGroupDesc{597, "GL2C0", 0, 259, 4, 8},
+        GpaCounterGroupDesc{598, "GL2C1", 1, 259, 4, 8},
+        GpaCounterGroupDesc{599, "GL2C2", 2, 259, 4, 8},
+        GpaCounterGroupDesc{600, "GL2C3", 3, 259, 4, 8},
+        GpaCounterGroupDesc{601, "GL2C4", 4, 259, 4, 8},
+        GpaCounterGroupDesc{602, "GL2C5", 5, 259, 4, 8},
+        GpaCounterGroupDesc{603, "GL2C6", 6, 259, 4, 8},
+        GpaCounterGroupDesc{604, "GL2C7", 7, 259, 4, 8},
+        GpaCounterGroupDesc{605, "GL2C8", 8, 259, 4, 8},
+        GpaCounterGroupDesc{606, "GL2C9", 9, 259, 4, 8},
+        GpaCounterGroupDesc{607, "GL2C10", 10, 259, 4, 8},
+        GpaCounterGroupDesc{608, "GL2C11", 11, 259, 4, 8},
+        GpaCounterGroupDesc{609, "GL2C12", 12, 259, 4, 8},
+        GpaCounterGroupDesc{610, "GL2C13", 13, 259, 4, 8},
+        GpaCounterGroupDesc{611, "GL2C14", 14, 259, 4, 8},
+        GpaCounterGroupDesc{612, "GL2C15", 15, 259, 4, 8},
+        GpaCounterGroupDesc{613, "GL2C16", 16, 259, 4, 8},
+        GpaCounterGroupDesc{614, "GL2C17", 17, 259, 4, 8},
+        GpaCounterGroupDesc{615, "GL2C18", 18, 259, 4, 8},
+        GpaCounterGroupDesc{616, "GL2C19", 19, 259, 4, 8},
+        GpaCounterGroupDesc{617, "GL2C20", 20, 259, 4, 8},
+        GpaCounterGroupDesc{618, "GL2C21", 21, 259, 4, 8},
+        GpaCounterGroupDesc{619, "GL2C22", 22, 259, 4, 8},
+        GpaCounterGroupDesc{620, "GL2C23", 23, 259, 4, 8},
+        GpaCounterGroupDesc{621, "CHA", 0, 40, 4, 4},
+        GpaCounterGroupDesc{622, "CHC0", 0, 44, 4, 4},
+        GpaCounterGroupDesc{623, "CHC1", 1, 44, 4, 4},
+        GpaCounterGroupDesc{624, "CHC2", 2, 44, 4, 4},
+        GpaCounterGroupDesc{625, "CHC3", 3, 44, 4, 4},
+        GpaCounterGroupDesc{626, "CHCG", 0, 44, 4, 4},
+        GpaCounterGroupDesc{627, "GUS", 0, 176, 2, 4},
+        GpaCounterGroupDesc{628, "GCR", 0, 155, 2, 4},
+        GpaCounterGroupDesc{629, "PA_PH", 0, 1024, 8, 16},
+        GpaCounterGroupDesc{630, "UTCL10", 0, 66, 2, 0},
+        GpaCounterGroupDesc{631, "UTCL11", 1, 66, 2, 0},
+        GpaCounterGroupDesc{632, "UTCL12", 2, 66, 2, 0},
+        GpaCounterGroupDesc{633, "UTCL13", 3, 66, 2, 0},
+        GpaCounterGroupDesc{634, "UTCL14", 4, 66, 2, 0},
+        GpaCounterGroupDesc{635, "UTCL15", 5, 66, 2, 0},
+        GpaCounterGroupDesc{636, "UTCL16", 6, 66, 2, 0},
+        GpaCounterGroupDesc{637, "UTCL17", 7, 66, 2, 0},
+        GpaCounterGroupDesc{638, "UTCL18", 8, 66, 2, 0},
+        GpaCounterGroupDesc{639, "UTCL19", 9, 66, 2, 0},
+        GpaCounterGroupDesc{640, "UTCL110", 10, 66, 2, 0},
+        GpaCounterGroupDesc{641, "UTCL111", 11, 66, 2, 0},
+        GpaCounterGroupDesc{642, "GE2_DIST", 0, 101, 4, 16},
+        GpaCounterGroupDesc{643, "GE_SE0", 0, 88, 4, 16},
+        GpaCounterGroupDesc{644, "GE_SE1", 1, 88, 4, 16},
+        GpaCounterGroupDesc{645, "GE_SE2", 2, 88, 4, 16},
+        GpaCounterGroupDesc{646, "GE_SE3", 3, 88, 4, 16},
+        GpaCounterGroupDesc{647, "GE_SE4", 4, 88, 4, 16},
+        GpaCounterGroupDesc{648, "GE_SE5", 5, 88, 4, 16},
+        GpaCounterGroupDesc{649, "SQWGP0", 0, 512, 8, 16},
+        GpaCounterGroupDesc{650, "SQWGP1", 1, 512, 8, 16},
+        GpaCounterGroupDesc{651, "SQWGP2", 2, 512, 8, 16},
+        GpaCounterGroupDesc{652, "SQWGP3", 3, 512, 8, 16},
+        GpaCounterGroupDesc{653, "SQWGP4", 4, 512, 8, 16},
+        GpaCounterGroupDesc{654, "SQWGP5", 5, 512, 8, 16},
+        GpaCounterGroupDesc{655, "SQWGP6", 6, 512, 8, 16},
+        GpaCounterGroupDesc{656, "SQWGP7", 7, 512, 8, 16},
+        GpaCounterGroupDesc{657, "SQWGP8", 8, 512, 8, 16},
+        GpaCounterGroupDesc{658, "SQWGP9", 9, 512, 8, 16},
+        GpaCounterGroupDesc{659, "SQWGP10", 10, 512, 8, 16},
+        GpaCounterGroupDesc{660, "SQWGP11", 11, 512, 8, 16},
+        GpaCounterGroupDesc{661, "SQWGP12", 12, 512, 8, 16},
+        GpaCounterGroupDesc{662, "SQWGP13", 13, 512, 8, 16},
+        GpaCounterGroupDesc{663, "SQWGP14", 14, 512, 8, 16},
+        GpaCounterGroupDesc{664, "SQWGP15", 15, 512, 8, 16},
+        GpaCounterGroupDesc{665, "SQWGP16", 16, 512, 8, 16},
+        GpaCounterGroupDesc{666, "SQWGP17", 17, 512, 8, 16},
+        GpaCounterGroupDesc{667, "SQWGP18", 18, 512, 8, 16},
+        GpaCounterGroupDesc{668, "SQWGP19", 19, 512, 8, 16},
+        GpaCounterGroupDesc{669, "SQWGP20", 20, 512, 8, 16},
+        GpaCounterGroupDesc{670, "SQWGP21", 21, 512, 8, 16},
+        GpaCounterGroupDesc{671, "SQWGP22", 22, 512, 8, 16},
+        GpaCounterGroupDesc{672, "SQWGP23", 23, 512, 8, 16},
+        GpaCounterGroupDesc{673, "SQWGP24", 24, 512, 8, 16},
+        GpaCounterGroupDesc{674, "SQWGP25", 25, 512, 8, 16},
+        GpaCounterGroupDesc{675, "SQWGP26", 26, 512, 8, 16},
+        GpaCounterGroupDesc{676, "SQWGP27", 27, 512, 8, 16},
+        GpaCounterGroupDesc{677, "SQWGP28", 28, 512, 8, 16},
+        GpaCounterGroupDesc{678, "SQWGP29", 29, 512, 8, 16},
+        GpaCounterGroupDesc{679, "SQWGP30", 30, 512, 8, 16},
+        GpaCounterGroupDesc{680, "SQWGP31", 31, 512, 8, 16},
+        GpaCounterGroupDesc{681, "SQWGP32", 32, 512, 8, 16},
+        GpaCounterGroupDesc{682, "SQWGP33", 33, 512, 8, 16},
+        GpaCounterGroupDesc{683, "SQWGP34", 34, 512, 8, 16},
+        GpaCounterGroupDesc{684, "SQWGP35", 35, 512, 8, 16},
+        GpaCounterGroupDesc{685, "SQWGP36", 36, 512, 8, 16},
+        GpaCounterGroupDesc{686, "SQWGP37", 37, 512, 8, 16},
+        GpaCounterGroupDesc{687, "SQWGP38", 38, 512, 8, 16},
+        GpaCounterGroupDesc{688, "SQWGP39", 39, 512, 8, 16},
+        GpaCounterGroupDesc{689, "SQWGP40", 40, 512, 8, 16},
+        GpaCounterGroupDesc{690, "SQWGP41", 41, 512, 8, 16},
+        GpaCounterGroupDesc{691, "SQWGP42", 42, 512, 8, 16},
+        GpaCounterGroupDesc{692, "SQWGP43", 43, 512, 8, 16},
+        GpaCounterGroupDesc{693, "SQWGP44", 44, 512, 8, 16},
+        GpaCounterGroupDesc{694, "SQWGP45", 45, 512, 8, 16},
+        GpaCounterGroupDesc{695, "SQWGP46", 46, 512, 8, 16},
+        GpaCounterGroupDesc{696, "SQWGP47", 47, 512, 8, 16},
+        GpaCounterGroupDesc{697, "SQWGP_GS0", 0, 512, 8, 16},
+        GpaCounterGroupDesc{698, "SQWGP_GS1", 1, 512, 8, 16},
+        GpaCounterGroupDesc{699, "SQWGP_GS2", 2, 512, 8, 16},
+        GpaCounterGroupDesc{700, "SQWGP_GS3", 3, 512, 8, 16},
+        GpaCounterGroupDesc{701, "SQWGP_GS4", 4, 512, 8, 16},
+        GpaCounterGroupDesc{702, "SQWGP_GS5", 5, 512, 8, 16},
+        GpaCounterGroupDesc{703, "SQWGP_GS6", 6, 512, 8, 16},
+        GpaCounterGroupDesc{704, "SQWGP_GS7", 7, 512, 8, 16},
+        GpaCounterGroupDesc{705, "SQWGP_GS8", 8, 512, 8, 16},
+        GpaCounterGroupDesc{706, "SQWGP_GS9", 9, 512, 8, 16},
+        GpaCounterGroupDesc{707, "SQWGP_GS10", 10, 512, 8, 16},
+        GpaCounterGroupDesc{708, "SQWGP_GS11", 11, 512, 8, 16},
+        GpaCounterGroupDesc{709, "SQWGP_GS12", 12, 512, 8, 16},
+        GpaCounterGroupDesc{710, "SQWGP_GS13", 13, 512, 8, 16},
+        GpaCounterGroupDesc{711, "SQWGP_GS14", 14, 512, 8, 16},
+        GpaCounterGroupDesc{712, "SQWGP_GS15", 15, 512, 8, 16},
+        GpaCounterGroupDesc{713, "SQWGP_GS16", 16, 512, 8, 16},
+        GpaCounterGroupDesc{714, "SQWGP_GS17", 17, 512, 8, 16},
+        GpaCounterGroupDesc{715, "SQWGP_GS18", 18, 512, 8, 16},
+        GpaCounterGroupDesc{716, "SQWGP_GS19", 19, 512, 8, 16},
+        GpaCounterGroupDesc{717, "SQWGP_GS20", 20, 512, 8, 16},
+        GpaCounterGroupDesc{718, "SQWGP_GS21", 21, 512, 8, 16},
+        GpaCounterGroupDesc{719, "SQWGP_GS22", 22, 512, 8, 16},
+        GpaCounterGroupDesc{720, "SQWGP_GS23", 23, 512, 8, 16},
+        GpaCounterGroupDesc{721, "SQWGP_GS24", 24, 512, 8, 16},
+        GpaCounterGroupDesc{722, "SQWGP_GS25", 25, 512, 8, 16},
+        GpaCounterGroupDesc{723, "SQWGP_GS26", 26, 512, 8, 16},
+        GpaCounterGroupDesc{724, "SQWGP_GS27", 27, 512, 8, 16},
+        GpaCounterGroupDesc{725, "SQWGP_GS28", 28, 512, 8, 16},
+        GpaCounterGroupDesc{726, "SQWGP_GS29", 29, 512, 8, 16},
+        GpaCounterGroupDesc{727, "SQWGP_GS30", 30, 512, 8, 16},
+        GpaCounterGroupDesc{728, "SQWGP_GS31", 31, 512, 8, 16},
+        GpaCounterGroupDesc{729, "SQWGP_GS32", 32, 512, 8, 16},
+        GpaCounterGroupDesc{730, "SQWGP_GS33", 33, 512, 8, 16},
+        GpaCounterGroupDesc{731, "SQWGP_GS34", 34, 512, 8, 16},
+        GpaCounterGroupDesc{732, "SQWGP_GS35", 35, 512, 8, 16},
+        GpaCounterGroupDesc{733, "SQWGP_GS36", 36, 512, 8, 16},
+        GpaCounterGroupDesc{734, "SQWGP_GS37", 37, 512, 8, 16},
+        GpaCounterGroupDesc{735, "SQWGP_GS38", 38, 512, 8, 16},
+        GpaCounterGroupDesc{736, "SQWGP_GS39", 39, 512, 8, 16},
+        GpaCounterGroupDesc{737, "SQWGP_GS40", 40, 512, 8, 16},
+        GpaCounterGroupDesc{738, "SQWGP_GS41", 41, 512, 8, 16},
+        GpaCounterGroupDesc{739, "SQWGP_GS42", 42, 512, 8, 16},
+        GpaCounterGroupDesc{740, "SQWGP_GS43", 43, 512, 8, 16},
+        GpaCounterGroupDesc{741, "SQWGP_GS44", 44, 512, 8, 16},
+        GpaCounterGroupDesc{742, "SQWGP_GS45", 45, 512, 8, 16},
+        GpaCounterGroupDesc{743, "SQWGP_GS46", 46, 512, 8, 16},
+        GpaCounterGroupDesc{744, "SQWGP_GS47", 47, 512, 8, 16},
+        GpaCounterGroupDesc{745, "SQWGP_PS0", 0, 512, 8, 16},
+        GpaCounterGroupDesc{746, "SQWGP_PS1", 1, 512, 8, 16},
+        GpaCounterGroupDesc{747, "SQWGP_PS2", 2, 512, 8, 16},
+        GpaCounterGroupDesc{748, "SQWGP_PS3", 3, 512, 8, 16},
+        GpaCounterGroupDesc{749, "SQWGP_PS4", 4, 512, 8, 16},
+        GpaCounterGroupDesc{750, "SQWGP_PS5", 5, 512, 8, 16},
+        GpaCounterGroupDesc{751, "SQWGP_PS6", 6, 512, 8, 16},
+        GpaCounterGroupDesc{752, "SQWGP_PS7", 7, 512, 8, 16},
+        GpaCounterGroupDesc{753, "SQWGP_PS8", 8, 512, 8, 16},
+        GpaCounterGroupDesc{754, "SQWGP_PS9", 9, 512, 8, 16},
+        GpaCounterGroupDesc{755, "SQWGP_PS10", 10, 512, 8, 16},
+        GpaCounterGroupDesc{756, "SQWGP_PS11", 11, 512, 8, 16},
+        GpaCounterGroupDesc{757, "SQWGP_PS12", 12, 512, 8, 16},
+        GpaCounterGroupDesc{758, "SQWGP_PS13", 13, 512, 8, 16},
+        GpaCounterGroupDesc{759, "SQWGP_PS14", 14, 512, 8, 16},
+        GpaCounterGroupDesc{760, "SQWGP_PS15", 15, 512, 8, 16},
+        GpaCounterGroupDesc{761, "SQWGP_PS16", 16, 512, 8, 16},
+        GpaCounterGroupDesc{762, "SQWGP_PS17", 17, 512, 8, 16},
+        GpaCounterGroupDesc{763, "SQWGP_PS18", 18, 512, 8, 16},
+        GpaCounterGroupDesc{764, "SQWGP_PS19", 19, 512, 8, 16},
+        GpaCounterGroupDesc{765, "SQWGP_PS20", 20, 512, 8, 16},
+        GpaCounterGroupDesc{766, "SQWGP_PS21", 21, 512, 8, 16},
+        GpaCounterGroupDesc{767, "SQWGP_PS22", 22, 512, 8, 16},
+        GpaCounterGroupDesc{768, "SQWGP_PS23", 23, 512, 8, 16},
+        GpaCounterGroupDesc{769, "SQWGP_PS24", 24, 512, 8, 16},
+        GpaCounterGroupDesc{770, "SQWGP_PS25", 25, 512, 8, 16},
+        GpaCounterGroupDesc{771, "SQWGP_PS26", 26, 512, 8, 16},
+        GpaCounterGroupDesc{772, "SQWGP_PS27", 27, 512, 8, 16},
+        GpaCounterGroupDesc{773, "SQWGP_PS28", 28, 512, 8, 16},
+        GpaCounterGroupDesc{774, "SQWGP_PS29", 29, 512, 8, 16},
+        GpaCounterGroupDesc{775, "SQWGP_PS30", 30, 512, 8, 16},
+        GpaCounterGroupDesc{776, "SQWGP_PS31", 31, 512, 8, 16},
+        GpaCounterGroupDesc{777, "SQWGP_PS32", 32, 512, 8, 16},
+        GpaCounterGroupDesc{778, "SQWGP_PS33", 33, 512, 8, 16},
+        GpaCounterGroupDesc{779, "SQWGP_PS34", 34, 512, 8, 16},
+        GpaCounterGroupDesc{780, "SQWGP_PS35", 35, 512, 8, 16},
+        GpaCounterGroupDesc{781, "SQWGP_PS36", 36, 512, 8, 16},
+        GpaCounterGroupDesc{782, "SQWGP_PS37", 37, 512, 8, 16},
+        GpaCounterGroupDesc{783, "SQWGP_PS38", 38, 512, 8, 16},
+        GpaCounterGroupDesc{784, "SQWGP_PS39", 39, 512, 8, 16},
+        GpaCounterGroupDesc{785, "SQWGP_PS40", 40, 512, 8, 16},
+        GpaCounterGroupDesc{786, "SQWGP_PS41", 41, 512, 8, 16},
+        GpaCounterGroupDesc{787, "SQWGP_PS42", 42, 512, 8, 16},
+        GpaCounterGroupDesc{788, "SQWGP_PS43", 43, 512, 8, 16},
+        GpaCounterGroupDesc{789, "SQWGP_PS44", 44, 512, 8, 16},
+        GpaCounterGroupDesc{790, "SQWGP_PS45", 45, 512, 8, 16},
+        GpaCounterGroupDesc{791, "SQWGP_PS46", 46, 512, 8, 16},
+        GpaCounterGroupDesc{792, "SQWGP_PS47", 47, 512, 8, 16},
+        GpaCounterGroupDesc{793, "SQWGP_HS0", 0, 512, 8, 16},
+        GpaCounterGroupDesc{794, "SQWGP_HS1", 1, 512, 8, 16},
+        GpaCounterGroupDesc{795, "SQWGP_HS2", 2, 512, 8, 16},
+        GpaCounterGroupDesc{796, "SQWGP_HS3", 3, 512, 8, 16},
+        GpaCounterGroupDesc{797, "SQWGP_HS4", 4, 512, 8, 16},
+        GpaCounterGroupDesc{798, "SQWGP_HS5", 5, 512, 8, 16},
+        GpaCounterGroupDesc{799, "SQWGP_HS6", 6, 512, 8, 16},
+        GpaCounterGroupDesc{800, "SQWGP_HS7", 7, 512, 8, 16},
+        GpaCounterGroupDesc{801, "SQWGP_HS8", 8, 512, 8, 16},
+        GpaCounterGroupDesc{802, "SQWGP_HS9", 9, 512, 8, 16},
+        GpaCounterGroupDesc{803, "SQWGP_HS10", 10, 512, 8, 16},
+        GpaCounterGroupDesc{804, "SQWGP_HS11", 11, 512, 8, 16},
+        GpaCounterGroupDesc{805, "SQWGP_HS12", 12, 512, 8, 16},
+        GpaCounterGroupDesc{806, "SQWGP_HS13", 13, 512, 8, 16},
+        GpaCounterGroupDesc{807, "SQWGP_HS14", 14, 512, 8, 16},
+        GpaCounterGroupDesc{808, "SQWGP_HS15", 15, 512, 8, 16},
+        GpaCounterGroupDesc{809, "SQWGP_HS16", 16, 512, 8, 16},
+        GpaCounterGroupDesc{810, "SQWGP_HS17", 17, 512, 8, 16},
+        GpaCounterGroupDesc{811, "SQWGP_HS18", 18, 512, 8, 16},
+        GpaCounterGroupDesc{812, "SQWGP_HS19", 19, 512, 8, 16},
+        GpaCounterGroupDesc{813, "SQWGP_HS20", 20, 512, 8, 16},
+        GpaCounterGroupDesc{814, "SQWGP_HS21", 21, 512, 8, 16},
+        GpaCounterGroupDesc{815, "SQWGP_HS22", 22, 512, 8, 16},
+        GpaCounterGroupDesc{816, "SQWGP_HS23", 23, 512, 8, 16},
+        GpaCounterGroupDesc{817, "SQWGP_HS24", 24, 512, 8, 16},
+        GpaCounterGroupDesc{818, "SQWGP_HS25", 25, 512, 8, 16},
+        GpaCounterGroupDesc{819, "SQWGP_HS26", 26, 512, 8, 16},
+        GpaCounterGroupDesc{820, "SQWGP_HS27", 27, 512, 8, 16},
+        GpaCounterGroupDesc{821, "SQWGP_HS28", 28, 512, 8, 16},
+        GpaCounterGroupDesc{822, "SQWGP_HS29", 29, 512, 8, 16},
+        GpaCounterGroupDesc{823, "SQWGP_HS30", 30, 512, 8, 16},
+        GpaCounterGroupDesc{824, "SQWGP_HS31", 31, 512, 8, 16},
+        GpaCounterGroupDesc{825, "SQWGP_HS32", 32, 512, 8, 16},
+        GpaCounterGroupDesc{826, "SQWGP_HS33", 33, 512, 8, 16},
+        GpaCounterGroupDesc{827, "SQWGP_HS34", 34, 512, 8, 16},
+        GpaCounterGroupDesc{828, "SQWGP_HS35", 35, 512, 8, 16},
+        GpaCounterGroupDesc{829, "SQWGP_HS36", 36, 512, 8, 16},
+        GpaCounterGroupDesc{830, "SQWGP_HS37", 37, 512, 8, 16},
+        GpaCounterGroupDesc{831, "SQWGP_HS38", 38, 512, 8, 16},
+        GpaCounterGroupDesc{832, "SQWGP_HS39", 39, 512, 8, 16},
+        GpaCounterGroupDesc{833, "SQWGP_HS40", 40, 512, 8, 16},
+        GpaCounterGroupDesc{834, "SQWGP_HS41", 41, 512, 8, 16},
+        GpaCounterGroupDesc{835, "SQWGP_HS42", 42, 512, 8, 16},
+        GpaCounterGroupDesc{836, "SQWGP_HS43", 43, 512, 8, 16},
+        GpaCounterGroupDesc{837, "SQWGP_HS44", 44, 512, 8, 16},
+        GpaCounterGroupDesc{838, "SQWGP_HS45", 45, 512, 8, 16},
+        GpaCounterGroupDesc{839, "SQWGP_HS46", 46, 512, 8, 16},
+        GpaCounterGroupDesc{840, "SQWGP_HS47", 47, 512, 8, 16},
+        GpaCounterGroupDesc{841, "SQWGP_CS0", 0, 512, 8, 16},
+        GpaCounterGroupDesc{842, "SQWGP_CS1", 1, 512, 8, 16},
+        GpaCounterGroupDesc{843, "SQWGP_CS2", 2, 512, 8, 16},
+        GpaCounterGroupDesc{844, "SQWGP_CS3", 3, 512, 8, 16},
+        GpaCounterGroupDesc{845, "SQWGP_CS4", 4, 512, 8, 16},
+        GpaCounterGroupDesc{846, "SQWGP_CS5", 5, 512, 8, 16},
+        GpaCounterGroupDesc{847, "SQWGP_CS6", 6, 512, 8, 16},
+        GpaCounterGroupDesc{848, "SQWGP_CS7", 7, 512, 8, 16},
+        GpaCounterGroupDesc{849, "SQWGP_CS8", 8, 512, 8, 16},
+        GpaCounterGroupDesc{850, "SQWGP_CS9", 9, 512, 8, 16},
+        GpaCounterGroupDesc{851, "SQWGP_CS10", 10, 512, 8, 16},
+        GpaCounterGroupDesc{852, "SQWGP_CS11", 11, 512, 8, 16},
+        GpaCounterGroupDesc{853, "SQWGP_CS12", 12, 512, 8, 16},
+        GpaCounterGroupDesc{854, "SQWGP_CS13", 13, 512, 8, 16},
+        GpaCounterGroupDesc{855, "SQWGP_CS14", 14, 512, 8, 16},
+        GpaCounterGroupDesc{856, "SQWGP_CS15", 15, 512, 8, 16},
+        GpaCounterGroupDesc{857, "SQWGP_CS16", 16, 512, 8, 16},
+        GpaCounterGroupDesc{858, "SQWGP_CS17", 17, 512, 8, 16},
+        GpaCounterGroupDesc{859, "SQWGP_CS18", 18, 512, 8, 16},
+        GpaCounterGroupDesc{860, "SQWGP_CS19", 19, 512, 8, 16},
+        GpaCounterGroupDesc{861, "SQWGP_CS20", 20, 512, 8, 16},
+        GpaCounterGroupDesc{862, "SQWGP_CS21", 21, 512, 8, 16},
+        GpaCounterGroupDesc{863, "SQWGP_CS22", 22, 512, 8, 16},
+        GpaCounterGroupDesc{864, "SQWGP_CS23", 23, 512, 8, 16},
+        GpaCounterGroupDesc{865, "SQWGP_CS24", 24, 512, 8, 16},
+        GpaCounterGroupDesc{866, "SQWGP_CS25", 25, 512, 8, 16},
+        GpaCounterGroupDesc{867, "SQWGP_CS26", 26, 512, 8, 16},
+        GpaCounterGroupDesc{868, "SQWGP_CS27", 27, 512, 8, 16},
+        GpaCounterGroupDesc{869, "SQWGP_CS28", 28, 512, 8, 16},
+        GpaCounterGroupDesc{870, "SQWGP_CS29", 29, 512, 8, 16},
+        GpaCounterGroupDesc{871, "SQWGP_CS30", 30, 512, 8, 16},
+        GpaCounterGroupDesc{872, "SQWGP_CS31", 31, 512, 8, 16},
+        GpaCounterGroupDesc{873, "SQWGP_CS32", 32, 512, 8, 16},
+        GpaCounterGroupDesc{874, "SQWGP_CS33", 33, 512, 8, 16},
+        GpaCounterGroupDesc{875, "SQWGP_CS34", 34, 512, 8, 16},
+        GpaCounterGroupDesc{876, "SQWGP_CS35", 35, 512, 8, 16},
+        GpaCounterGroupDesc{877, "SQWGP_CS36", 36, 512, 8, 16},
+        GpaCounterGroupDesc{878, "SQWGP_CS37", 37, 512, 8, 16},
+        GpaCounterGroupDesc{879, "SQWGP_CS38", 38, 512, 8, 16},
+        GpaCounterGroupDesc{880, "SQWGP_CS39", 39, 512, 8, 16},
+        GpaCounterGroupDesc{881, "SQWGP_CS40", 40, 512, 8, 16},
+        GpaCounterGroupDesc{882, "SQWGP_CS41", 41, 512, 8, 16},
+        GpaCounterGroupDesc{883, "SQWGP_CS42", 42, 512, 8, 16},
+        GpaCounterGroupDesc{884, "SQWGP_CS43", 43, 512, 8, 16},
+        GpaCounterGroupDesc{885, "SQWGP_CS44", 44, 512, 8, 16},
+        GpaCounterGroupDesc{886, "SQWGP_CS45", 45, 512, 8, 16},
+        GpaCounterGroupDesc{887, "SQWGP_CS46", 46, 512, 8, 16},
+        GpaCounterGroupDesc{888, "SQWGP_CS47", 47, 512, 8, 16},
+        GpaCounterGroupDesc{889, "PC0", 0, 47, 4, 0},
+        GpaCounterGroupDesc{890, "PC1", 1, 47, 4, 0},
+        GpaCounterGroupDesc{891, "PC2", 2, 47, 4, 0},
+        GpaCounterGroupDesc{892, "PC3", 3, 47, 4, 0},
+        GpaCounterGroupDesc{893, "PC4", 4, 47, 4, 0},
+        GpaCounterGroupDesc{894, "PC5", 5, 47, 4, 0},
+        GpaCounterGroupDesc{895, "GPUTime", 0, 6, 3, 0},
     };
     const gpa_array_view<GpaCounterGroupDesc> kHwVkGroupsGfx11 = kHwVkGroupsGfx11Data;
 
@@ -2098,350 +2100,350 @@ namespace counter_vk_gfx11
         {413, 88456},
         {414, 88770},
         {416, 89232},
-        {433, 90260},
-        {434, 90358},
-        {435, 90456},
-        {436, 90554},
-        {437, 90652},
-        {438, 90750},
-        {439, 90848},
-        {440, 90946},
-        {441, 91044},
-        {442, 91142},
-        {443, 91240},
-        {444, 91338},
-        {445, 91436},
-        {446, 91534},
-        {447, 91632},
-        {448, 91730},
-        {449, 91828},
-        {450, 91926},
-        {451, 92024},
-        {452, 92122},
-        {453, 92220},
-        {454, 92318},
-        {455, 92416},
-        {456, 92514},
-        {543, 100636},
-        {544, 100720},
-        {545, 100804},
-        {546, 100888},
-        {547, 100972},
-        {548, 101056},
-        {549, 101140},
-        {550, 101224},
-        {551, 101308},
-        {552, 101392},
-        {553, 101476},
-        {554, 101560},
-        {555, 101644},
-        {556, 101728},
-        {557, 101812},
-        {558, 101896},
-        {559, 101980},
-        {560, 102064},
-        {561, 102148},
-        {562, 102232},
-        {563, 102316},
-        {564, 102400},
-        {565, 102484},
-        {566, 102568},
-        {567, 102652},
-        {568, 102736},
-        {569, 102820},
-        {570, 102904},
-        {571, 102988},
-        {572, 103072},
-        {573, 103156},
-        {574, 103240},
-        {575, 103324},
-        {576, 103408},
-        {577, 103492},
-        {578, 103576},
-        {579, 103660},
-        {580, 103744},
-        {581, 103828},
-        {582, 103912},
-        {583, 103996},
-        {584, 104080},
-        {585, 104164},
-        {586, 104248},
-        {587, 104332},
-        {588, 104416},
-        {589, 104500},
-        {590, 104584},
-        {595, 105100},
-        {596, 105359},
-        {597, 105618},
-        {598, 105877},
-        {599, 106136},
-        {600, 106395},
-        {601, 106654},
-        {602, 106913},
-        {603, 107172},
-        {604, 107431},
-        {605, 107690},
-        {606, 107949},
-        {607, 108208},
-        {608, 108467},
-        {609, 108726},
-        {610, 108985},
-        {611, 109244},
-        {612, 109503},
-        {613, 109762},
-        {614, 110021},
-        {615, 110280},
-        {616, 110539},
-        {617, 110798},
-        {618, 111057},
-        {640, 113723},
-        {641, 113824},
-        {642, 113912},
-        {643, 114000},
-        {644, 114088},
-        {645, 114176},
-        {646, 114264},
-        {647, 114352},
-        {648, 114864},
-        {649, 115376},
-        {650, 115888},
-        {651, 116400},
-        {652, 116912},
-        {653, 117424},
-        {654, 117936},
-        {655, 118448},
-        {656, 118960},
-        {657, 119472},
-        {658, 119984},
-        {659, 120496},
-        {660, 121008},
-        {661, 121520},
-        {662, 122032},
-        {663, 122544},
-        {664, 123056},
-        {665, 123568},
-        {666, 124080},
-        {667, 124592},
-        {668, 125104},
-        {669, 125616},
-        {670, 126128},
-        {671, 126640},
-        {672, 127152},
-        {673, 127664},
-        {674, 128176},
-        {675, 128688},
-        {676, 129200},
-        {677, 129712},
-        {678, 130224},
-        {679, 130736},
-        {680, 131248},
-        {681, 131760},
-        {682, 132272},
-        {683, 132784},
-        {684, 133296},
-        {685, 133808},
-        {686, 134320},
-        {687, 134832},
-        {688, 135344},
-        {689, 135856},
-        {690, 136368},
-        {691, 136880},
-        {692, 137392},
-        {693, 137904},
-        {694, 138416},
-        {695, 138928},
-        {696, 139440},
-        {697, 139952},
-        {698, 140464},
-        {699, 140976},
-        {700, 141488},
-        {701, 142000},
-        {702, 142512},
-        {703, 143024},
-        {704, 143536},
-        {705, 144048},
-        {706, 144560},
-        {707, 145072},
-        {708, 145584},
-        {709, 146096},
-        {710, 146608},
-        {711, 147120},
-        {712, 147632},
-        {713, 148144},
-        {714, 148656},
-        {715, 149168},
-        {716, 149680},
-        {717, 150192},
-        {718, 150704},
-        {719, 151216},
-        {720, 151728},
-        {721, 152240},
-        {722, 152752},
-        {723, 153264},
-        {724, 153776},
-        {725, 154288},
-        {726, 154800},
-        {727, 155312},
-        {728, 155824},
-        {729, 156336},
-        {730, 156848},
-        {731, 157360},
-        {732, 157872},
-        {733, 158384},
-        {734, 158896},
-        {735, 159408},
-        {736, 159920},
-        {737, 160432},
-        {738, 160944},
-        {739, 161456},
-        {740, 161968},
-        {741, 162480},
-        {742, 162992},
-        {743, 163504},
-        {744, 164016},
-        {745, 164528},
-        {746, 165040},
-        {747, 165552},
-        {748, 166064},
-        {749, 166576},
-        {750, 167088},
-        {751, 167600},
-        {752, 168112},
-        {753, 168624},
-        {754, 169136},
-        {755, 169648},
-        {756, 170160},
-        {757, 170672},
-        {758, 171184},
-        {759, 171696},
-        {760, 172208},
-        {761, 172720},
-        {762, 173232},
-        {763, 173744},
-        {764, 174256},
-        {765, 174768},
-        {766, 175280},
-        {767, 175792},
-        {768, 176304},
-        {769, 176816},
-        {770, 177328},
-        {771, 177840},
-        {772, 178352},
-        {773, 178864},
-        {774, 179376},
-        {775, 179888},
-        {776, 180400},
-        {777, 180912},
-        {778, 181424},
-        {779, 181936},
-        {780, 182448},
-        {781, 182960},
-        {782, 183472},
-        {783, 183984},
-        {784, 184496},
-        {785, 185008},
-        {786, 185520},
-        {787, 186032},
-        {788, 186544},
-        {789, 187056},
-        {790, 187568},
-        {791, 188080},
-        {792, 188592},
-        {793, 189104},
-        {794, 189616},
-        {795, 190128},
-        {796, 190640},
-        {797, 191152},
-        {798, 191664},
-        {799, 192176},
-        {800, 192688},
-        {801, 193200},
-        {802, 193712},
-        {803, 194224},
-        {804, 194736},
-        {805, 195248},
-        {806, 195760},
-        {807, 196272},
-        {808, 196784},
-        {809, 197296},
-        {810, 197808},
-        {811, 198320},
-        {812, 198832},
-        {813, 199344},
-        {814, 199856},
-        {815, 200368},
-        {816, 200880},
-        {817, 201392},
-        {818, 201904},
-        {819, 202416},
-        {820, 202928},
-        {821, 203440},
-        {822, 203952},
-        {823, 204464},
-        {824, 204976},
-        {825, 205488},
-        {826, 206000},
-        {827, 206512},
-        {828, 207024},
-        {829, 207536},
-        {830, 208048},
-        {831, 208560},
-        {832, 209072},
-        {833, 209584},
-        {834, 210096},
-        {835, 210608},
-        {836, 211120},
-        {837, 211632},
-        {838, 212144},
-        {839, 212656},
-        {840, 213168},
-        {841, 213680},
-        {842, 214192},
-        {843, 214704},
-        {844, 215216},
-        {845, 215728},
-        {846, 216240},
-        {847, 216752},
-        {848, 217264},
-        {849, 217776},
-        {850, 218288},
-        {851, 218800},
-        {852, 219312},
-        {853, 219824},
-        {854, 220336},
-        {855, 220848},
-        {856, 221360},
-        {857, 221872},
-        {858, 222384},
-        {859, 222896},
-        {860, 223408},
-        {861, 223920},
-        {862, 224432},
-        {863, 224944},
-        {864, 225456},
-        {865, 225968},
-        {866, 226480},
-        {867, 226992},
-        {868, 227504},
-        {869, 228016},
-        {870, 228528},
-        {871, 229040},
-        {872, 229552},
-        {873, 230064},
-        {874, 230576},
-        {875, 231088},
-        {876, 231600},
-        {877, 232112},
-        {878, 232624},
-        {879, 233136},
-        {880, 233648},
-        {881, 234160},
-        {882, 234672},
-        {883, 235184},
-        {884, 235696},
-        {885, 236208},
-        {886, 236720},
-        {893, 237514},
+        {435, 90302},
+        {436, 90400},
+        {437, 90498},
+        {438, 90596},
+        {439, 90694},
+        {440, 90792},
+        {441, 90890},
+        {442, 90988},
+        {443, 91086},
+        {444, 91184},
+        {445, 91282},
+        {446, 91380},
+        {447, 91478},
+        {448, 91576},
+        {449, 91674},
+        {450, 91772},
+        {451, 91870},
+        {452, 91968},
+        {453, 92066},
+        {454, 92164},
+        {455, 92262},
+        {456, 92360},
+        {457, 92458},
+        {458, 92556},
+        {545, 100678},
+        {546, 100762},
+        {547, 100846},
+        {548, 100930},
+        {549, 101014},
+        {550, 101098},
+        {551, 101182},
+        {552, 101266},
+        {553, 101350},
+        {554, 101434},
+        {555, 101518},
+        {556, 101602},
+        {557, 101686},
+        {558, 101770},
+        {559, 101854},
+        {560, 101938},
+        {561, 102022},
+        {562, 102106},
+        {563, 102190},
+        {564, 102274},
+        {565, 102358},
+        {566, 102442},
+        {567, 102526},
+        {568, 102610},
+        {569, 102694},
+        {570, 102778},
+        {571, 102862},
+        {572, 102946},
+        {573, 103030},
+        {574, 103114},
+        {575, 103198},
+        {576, 103282},
+        {577, 103366},
+        {578, 103450},
+        {579, 103534},
+        {580, 103618},
+        {581, 103702},
+        {582, 103786},
+        {583, 103870},
+        {584, 103954},
+        {585, 104038},
+        {586, 104122},
+        {587, 104206},
+        {588, 104290},
+        {589, 104374},
+        {590, 104458},
+        {591, 104542},
+        {592, 104626},
+        {597, 105142},
+        {598, 105401},
+        {599, 105660},
+        {600, 105919},
+        {601, 106178},
+        {602, 106437},
+        {603, 106696},
+        {604, 106955},
+        {605, 107214},
+        {606, 107473},
+        {607, 107732},
+        {608, 107991},
+        {609, 108250},
+        {610, 108509},
+        {611, 108768},
+        {612, 109027},
+        {613, 109286},
+        {614, 109545},
+        {615, 109804},
+        {616, 110063},
+        {617, 110322},
+        {618, 110581},
+        {619, 110840},
+        {620, 111099},
+        {642, 113765},
+        {643, 113866},
+        {644, 113954},
+        {645, 114042},
+        {646, 114130},
+        {647, 114218},
+        {648, 114306},
+        {649, 114394},
+        {650, 114906},
+        {651, 115418},
+        {652, 115930},
+        {653, 116442},
+        {654, 116954},
+        {655, 117466},
+        {656, 117978},
+        {657, 118490},
+        {658, 119002},
+        {659, 119514},
+        {660, 120026},
+        {661, 120538},
+        {662, 121050},
+        {663, 121562},
+        {664, 122074},
+        {665, 122586},
+        {666, 123098},
+        {667, 123610},
+        {668, 124122},
+        {669, 124634},
+        {670, 125146},
+        {671, 125658},
+        {672, 126170},
+        {673, 126682},
+        {674, 127194},
+        {675, 127706},
+        {676, 128218},
+        {677, 128730},
+        {678, 129242},
+        {679, 129754},
+        {680, 130266},
+        {681, 130778},
+        {682, 131290},
+        {683, 131802},
+        {684, 132314},
+        {685, 132826},
+        {686, 133338},
+        {687, 133850},
+        {688, 134362},
+        {689, 134874},
+        {690, 135386},
+        {691, 135898},
+        {692, 136410},
+        {693, 136922},
+        {694, 137434},
+        {695, 137946},
+        {696, 138458},
+        {697, 138970},
+        {698, 139482},
+        {699, 139994},
+        {700, 140506},
+        {701, 141018},
+        {702, 141530},
+        {703, 142042},
+        {704, 142554},
+        {705, 143066},
+        {706, 143578},
+        {707, 144090},
+        {708, 144602},
+        {709, 145114},
+        {710, 145626},
+        {711, 146138},
+        {712, 146650},
+        {713, 147162},
+        {714, 147674},
+        {715, 148186},
+        {716, 148698},
+        {717, 149210},
+        {718, 149722},
+        {719, 150234},
+        {720, 150746},
+        {721, 151258},
+        {722, 151770},
+        {723, 152282},
+        {724, 152794},
+        {725, 153306},
+        {726, 153818},
+        {727, 154330},
+        {728, 154842},
+        {729, 155354},
+        {730, 155866},
+        {731, 156378},
+        {732, 156890},
+        {733, 157402},
+        {734, 157914},
+        {735, 158426},
+        {736, 158938},
+        {737, 159450},
+        {738, 159962},
+        {739, 160474},
+        {740, 160986},
+        {741, 161498},
+        {742, 162010},
+        {743, 162522},
+        {744, 163034},
+        {745, 163546},
+        {746, 164058},
+        {747, 164570},
+        {748, 165082},
+        {749, 165594},
+        {750, 166106},
+        {751, 166618},
+        {752, 167130},
+        {753, 167642},
+        {754, 168154},
+        {755, 168666},
+        {756, 169178},
+        {757, 169690},
+        {758, 170202},
+        {759, 170714},
+        {760, 171226},
+        {761, 171738},
+        {762, 172250},
+        {763, 172762},
+        {764, 173274},
+        {765, 173786},
+        {766, 174298},
+        {767, 174810},
+        {768, 175322},
+        {769, 175834},
+        {770, 176346},
+        {771, 176858},
+        {772, 177370},
+        {773, 177882},
+        {774, 178394},
+        {775, 178906},
+        {776, 179418},
+        {777, 179930},
+        {778, 180442},
+        {779, 180954},
+        {780, 181466},
+        {781, 181978},
+        {782, 182490},
+        {783, 183002},
+        {784, 183514},
+        {785, 184026},
+        {786, 184538},
+        {787, 185050},
+        {788, 185562},
+        {789, 186074},
+        {790, 186586},
+        {791, 187098},
+        {792, 187610},
+        {793, 188122},
+        {794, 188634},
+        {795, 189146},
+        {796, 189658},
+        {797, 190170},
+        {798, 190682},
+        {799, 191194},
+        {800, 191706},
+        {801, 192218},
+        {802, 192730},
+        {803, 193242},
+        {804, 193754},
+        {805, 194266},
+        {806, 194778},
+        {807, 195290},
+        {808, 195802},
+        {809, 196314},
+        {810, 196826},
+        {811, 197338},
+        {812, 197850},
+        {813, 198362},
+        {814, 198874},
+        {815, 199386},
+        {816, 199898},
+        {817, 200410},
+        {818, 200922},
+        {819, 201434},
+        {820, 201946},
+        {821, 202458},
+        {822, 202970},
+        {823, 203482},
+        {824, 203994},
+        {825, 204506},
+        {826, 205018},
+        {827, 205530},
+        {828, 206042},
+        {829, 206554},
+        {830, 207066},
+        {831, 207578},
+        {832, 208090},
+        {833, 208602},
+        {834, 209114},
+        {835, 209626},
+        {836, 210138},
+        {837, 210650},
+        {838, 211162},
+        {839, 211674},
+        {840, 212186},
+        {841, 212698},
+        {842, 213210},
+        {843, 213722},
+        {844, 214234},
+        {845, 214746},
+        {846, 215258},
+        {847, 215770},
+        {848, 216282},
+        {849, 216794},
+        {850, 217306},
+        {851, 217818},
+        {852, 218330},
+        {853, 218842},
+        {854, 219354},
+        {855, 219866},
+        {856, 220378},
+        {857, 220890},
+        {858, 221402},
+        {859, 221914},
+        {860, 222426},
+        {861, 222938},
+        {862, 223450},
+        {863, 223962},
+        {864, 224474},
+        {865, 224986},
+        {866, 225498},
+        {867, 226010},
+        {868, 226522},
+        {869, 227034},
+        {870, 227546},
+        {871, 228058},
+        {872, 228570},
+        {873, 229082},
+        {874, 229594},
+        {875, 230106},
+        {876, 230618},
+        {877, 231130},
+        {878, 231642},
+        {879, 232154},
+        {880, 232666},
+        {881, 233178},
+        {882, 233690},
+        {883, 234202},
+        {884, 234714},
+        {885, 235226},
+        {886, 235738},
+        {887, 236250},
+        {888, 236762},
+        {895, 237556},
     };
 
     GpaPaddedCounterDesc kVkPaddedCounterByGroupGfx11[] = {
@@ -2685,12 +2687,14 @@ namespace counter_vk_gfx11
         {388, 80369, {388, 389, 390, 391, 392}},
         {389, 80762, {388, 389, 390, 391, 392}},
         {390, 81155, {388, 389, 390, 391, 392}},
-        {887, 237232, {45, 46}},
-        {888, 237279, {45, 46}},
-        {889, 237326, {45, 46}},
-        {890, 237373, {45, 46}},
-        {891, 237420, {45, 46}},
-        {892, 237467, {45, 46}},
+        {432, 90169, {0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23}},
+        {433, 90193, {0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17}},
+        {889, 237274, {45, 46}},
+        {890, 237321, {45, 46}},
+        {891, 237368, {45, 46}},
+        {892, 237415, {45, 46}},
+        {893, 237462, {45, 46}},
+        {894, 237509, {45, 46}},
     };
 
     GpaSqCounterGroupDesc kHwVkSqGroupsGfx11[] = {
@@ -2724,246 +2728,246 @@ namespace counter_vk_gfx11
         {64, 3, kSqCs},
         {65, 4, kSqCs},
         {66, 5, kSqCs},
-        {647, 0, kSqAll},
-        {648, 1, kSqAll},
-        {649, 2, kSqAll},
-        {650, 3, kSqAll},
-        {651, 4, kSqAll},
-        {652, 5, kSqAll},
-        {653, 6, kSqAll},
-        {654, 7, kSqAll},
-        {655, 8, kSqAll},
-        {656, 9, kSqAll},
-        {657, 10, kSqAll},
-        {658, 11, kSqAll},
-        {659, 12, kSqAll},
-        {660, 13, kSqAll},
-        {661, 14, kSqAll},
-        {662, 15, kSqAll},
-        {663, 16, kSqAll},
-        {664, 17, kSqAll},
-        {665, 18, kSqAll},
-        {666, 19, kSqAll},
-        {667, 20, kSqAll},
-        {668, 21, kSqAll},
-        {669, 22, kSqAll},
-        {670, 23, kSqAll},
-        {671, 24, kSqAll},
-        {672, 25, kSqAll},
-        {673, 26, kSqAll},
-        {674, 27, kSqAll},
-        {675, 28, kSqAll},
-        {676, 29, kSqAll},
-        {677, 30, kSqAll},
-        {678, 31, kSqAll},
-        {679, 32, kSqAll},
-        {680, 33, kSqAll},
-        {681, 34, kSqAll},
-        {682, 35, kSqAll},
-        {683, 36, kSqAll},
-        {684, 37, kSqAll},
-        {685, 38, kSqAll},
-        {686, 39, kSqAll},
-        {687, 40, kSqAll},
-        {688, 41, kSqAll},
-        {689, 42, kSqAll},
-        {690, 43, kSqAll},
-        {691, 44, kSqAll},
-        {692, 45, kSqAll},
-        {693, 46, kSqAll},
-        {694, 47, kSqAll},
-        {695, 0, kSqGs},
-        {696, 1, kSqGs},
-        {697, 2, kSqGs},
-        {698, 3, kSqGs},
-        {699, 4, kSqGs},
-        {700, 5, kSqGs},
-        {701, 6, kSqGs},
-        {702, 7, kSqGs},
-        {703, 8, kSqGs},
-        {704, 9, kSqGs},
-        {705, 10, kSqGs},
-        {706, 11, kSqGs},
-        {707, 12, kSqGs},
-        {708, 13, kSqGs},
-        {709, 14, kSqGs},
-        {710, 15, kSqGs},
-        {711, 16, kSqGs},
-        {712, 17, kSqGs},
-        {713, 18, kSqGs},
-        {714, 19, kSqGs},
-        {715, 20, kSqGs},
-        {716, 21, kSqGs},
-        {717, 22, kSqGs},
-        {718, 23, kSqGs},
-        {719, 24, kSqGs},
-        {720, 25, kSqGs},
-        {721, 26, kSqGs},
-        {722, 27, kSqGs},
-        {723, 28, kSqGs},
-        {724, 29, kSqGs},
-        {725, 30, kSqGs},
-        {726, 31, kSqGs},
-        {727, 32, kSqGs},
-        {728, 33, kSqGs},
-        {729, 34, kSqGs},
-        {730, 35, kSqGs},
-        {731, 36, kSqGs},
-        {732, 37, kSqGs},
-        {733, 38, kSqGs},
-        {734, 39, kSqGs},
-        {735, 40, kSqGs},
-        {736, 41, kSqGs},
-        {737, 42, kSqGs},
-        {738, 43, kSqGs},
-        {739, 44, kSqGs},
-        {740, 45, kSqGs},
-        {741, 46, kSqGs},
-        {742, 47, kSqGs},
-        {743, 0, kSqPs},
-        {744, 1, kSqPs},
-        {745, 2, kSqPs},
-        {746, 3, kSqPs},
-        {747, 4, kSqPs},
-        {748, 5, kSqPs},
-        {749, 6, kSqPs},
-        {750, 7, kSqPs},
-        {751, 8, kSqPs},
-        {752, 9, kSqPs},
-        {753, 10, kSqPs},
-        {754, 11, kSqPs},
-        {755, 12, kSqPs},
-        {756, 13, kSqPs},
-        {757, 14, kSqPs},
-        {758, 15, kSqPs},
-        {759, 16, kSqPs},
-        {760, 17, kSqPs},
-        {761, 18, kSqPs},
-        {762, 19, kSqPs},
-        {763, 20, kSqPs},
-        {764, 21, kSqPs},
-        {765, 22, kSqPs},
-        {766, 23, kSqPs},
-        {767, 24, kSqPs},
-        {768, 25, kSqPs},
-        {769, 26, kSqPs},
-        {770, 27, kSqPs},
-        {771, 28, kSqPs},
-        {772, 29, kSqPs},
-        {773, 30, kSqPs},
-        {774, 31, kSqPs},
-        {775, 32, kSqPs},
-        {776, 33, kSqPs},
-        {777, 34, kSqPs},
-        {778, 35, kSqPs},
-        {779, 36, kSqPs},
-        {780, 37, kSqPs},
-        {781, 38, kSqPs},
-        {782, 39, kSqPs},
-        {783, 40, kSqPs},
-        {784, 41, kSqPs},
-        {785, 42, kSqPs},
-        {786, 43, kSqPs},
-        {787, 44, kSqPs},
-        {788, 45, kSqPs},
-        {789, 46, kSqPs},
-        {790, 47, kSqPs},
-        {791, 0, kSqHs},
-        {792, 1, kSqHs},
-        {793, 2, kSqHs},
-        {794, 3, kSqHs},
-        {795, 4, kSqHs},
-        {796, 5, kSqHs},
-        {797, 6, kSqHs},
-        {798, 7, kSqHs},
-        {799, 8, kSqHs},
-        {800, 9, kSqHs},
-        {801, 10, kSqHs},
-        {802, 11, kSqHs},
-        {803, 12, kSqHs},
-        {804, 13, kSqHs},
-        {805, 14, kSqHs},
-        {806, 15, kSqHs},
-        {807, 16, kSqHs},
-        {808, 17, kSqHs},
-        {809, 18, kSqHs},
-        {810, 19, kSqHs},
-        {811, 20, kSqHs},
-        {812, 21, kSqHs},
-        {813, 22, kSqHs},
-        {814, 23, kSqHs},
-        {815, 24, kSqHs},
-        {816, 25, kSqHs},
-        {817, 26, kSqHs},
-        {818, 27, kSqHs},
-        {819, 28, kSqHs},
-        {820, 29, kSqHs},
-        {821, 30, kSqHs},
-        {822, 31, kSqHs},
-        {823, 32, kSqHs},
-        {824, 33, kSqHs},
-        {825, 34, kSqHs},
-        {826, 35, kSqHs},
-        {827, 36, kSqHs},
-        {828, 37, kSqHs},
-        {829, 38, kSqHs},
-        {830, 39, kSqHs},
-        {831, 40, kSqHs},
-        {832, 41, kSqHs},
-        {833, 42, kSqHs},
-        {834, 43, kSqHs},
-        {835, 44, kSqHs},
-        {836, 45, kSqHs},
-        {837, 46, kSqHs},
-        {838, 47, kSqHs},
-        {839, 0, kSqCs},
-        {840, 1, kSqCs},
-        {841, 2, kSqCs},
-        {842, 3, kSqCs},
-        {843, 4, kSqCs},
-        {844, 5, kSqCs},
-        {845, 6, kSqCs},
-        {846, 7, kSqCs},
-        {847, 8, kSqCs},
-        {848, 9, kSqCs},
-        {849, 10, kSqCs},
-        {850, 11, kSqCs},
-        {851, 12, kSqCs},
-        {852, 13, kSqCs},
-        {853, 14, kSqCs},
-        {854, 15, kSqCs},
-        {855, 16, kSqCs},
-        {856, 17, kSqCs},
-        {857, 18, kSqCs},
-        {858, 19, kSqCs},
-        {859, 20, kSqCs},
-        {860, 21, kSqCs},
-        {861, 22, kSqCs},
-        {862, 23, kSqCs},
-        {863, 24, kSqCs},
-        {864, 25, kSqCs},
-        {865, 26, kSqCs},
-        {866, 27, kSqCs},
-        {867, 28, kSqCs},
-        {868, 29, kSqCs},
-        {869, 30, kSqCs},
-        {870, 31, kSqCs},
-        {871, 32, kSqCs},
-        {872, 33, kSqCs},
-        {873, 34, kSqCs},
-        {874, 35, kSqCs},
-        {875, 36, kSqCs},
-        {876, 37, kSqCs},
-        {877, 38, kSqCs},
-        {878, 39, kSqCs},
-        {879, 40, kSqCs},
-        {880, 41, kSqCs},
-        {881, 42, kSqCs},
-        {882, 43, kSqCs},
-        {883, 44, kSqCs},
-        {884, 45, kSqCs},
-        {885, 46, kSqCs},
-        {886, 47, kSqCs},
+        {649, 0, kSqAll},
+        {650, 1, kSqAll},
+        {651, 2, kSqAll},
+        {652, 3, kSqAll},
+        {653, 4, kSqAll},
+        {654, 5, kSqAll},
+        {655, 6, kSqAll},
+        {656, 7, kSqAll},
+        {657, 8, kSqAll},
+        {658, 9, kSqAll},
+        {659, 10, kSqAll},
+        {660, 11, kSqAll},
+        {661, 12, kSqAll},
+        {662, 13, kSqAll},
+        {663, 14, kSqAll},
+        {664, 15, kSqAll},
+        {665, 16, kSqAll},
+        {666, 17, kSqAll},
+        {667, 18, kSqAll},
+        {668, 19, kSqAll},
+        {669, 20, kSqAll},
+        {670, 21, kSqAll},
+        {671, 22, kSqAll},
+        {672, 23, kSqAll},
+        {673, 24, kSqAll},
+        {674, 25, kSqAll},
+        {675, 26, kSqAll},
+        {676, 27, kSqAll},
+        {677, 28, kSqAll},
+        {678, 29, kSqAll},
+        {679, 30, kSqAll},
+        {680, 31, kSqAll},
+        {681, 32, kSqAll},
+        {682, 33, kSqAll},
+        {683, 34, kSqAll},
+        {684, 35, kSqAll},
+        {685, 36, kSqAll},
+        {686, 37, kSqAll},
+        {687, 38, kSqAll},
+        {688, 39, kSqAll},
+        {689, 40, kSqAll},
+        {690, 41, kSqAll},
+        {691, 42, kSqAll},
+        {692, 43, kSqAll},
+        {693, 44, kSqAll},
+        {694, 45, kSqAll},
+        {695, 46, kSqAll},
+        {696, 47, kSqAll},
+        {697, 0, kSqGs},
+        {698, 1, kSqGs},
+        {699, 2, kSqGs},
+        {700, 3, kSqGs},
+        {701, 4, kSqGs},
+        {702, 5, kSqGs},
+        {703, 6, kSqGs},
+        {704, 7, kSqGs},
+        {705, 8, kSqGs},
+        {706, 9, kSqGs},
+        {707, 10, kSqGs},
+        {708, 11, kSqGs},
+        {709, 12, kSqGs},
+        {710, 13, kSqGs},
+        {711, 14, kSqGs},
+        {712, 15, kSqGs},
+        {713, 16, kSqGs},
+        {714, 17, kSqGs},
+        {715, 18, kSqGs},
+        {716, 19, kSqGs},
+        {717, 20, kSqGs},
+        {718, 21, kSqGs},
+        {719, 22, kSqGs},
+        {720, 23, kSqGs},
+        {721, 24, kSqGs},
+        {722, 25, kSqGs},
+        {723, 26, kSqGs},
+        {724, 27, kSqGs},
+        {725, 28, kSqGs},
+        {726, 29, kSqGs},
+        {727, 30, kSqGs},
+        {728, 31, kSqGs},
+        {729, 32, kSqGs},
+        {730, 33, kSqGs},
+        {731, 34, kSqGs},
+        {732, 35, kSqGs},
+        {733, 36, kSqGs},
+        {734, 37, kSqGs},
+        {735, 38, kSqGs},
+        {736, 39, kSqGs},
+        {737, 40, kSqGs},
+        {738, 41, kSqGs},
+        {739, 42, kSqGs},
+        {740, 43, kSqGs},
+        {741, 44, kSqGs},
+        {742, 45, kSqGs},
+        {743, 46, kSqGs},
+        {744, 47, kSqGs},
+        {745, 0, kSqPs},
+        {746, 1, kSqPs},
+        {747, 2, kSqPs},
+        {748, 3, kSqPs},
+        {749, 4, kSqPs},
+        {750, 5, kSqPs},
+        {751, 6, kSqPs},
+        {752, 7, kSqPs},
+        {753, 8, kSqPs},
+        {754, 9, kSqPs},
+        {755, 10, kSqPs},
+        {756, 11, kSqPs},
+        {757, 12, kSqPs},
+        {758, 13, kSqPs},
+        {759, 14, kSqPs},
+        {760, 15, kSqPs},
+        {761, 16, kSqPs},
+        {762, 17, kSqPs},
+        {763, 18, kSqPs},
+        {764, 19, kSqPs},
+        {765, 20, kSqPs},
+        {766, 21, kSqPs},
+        {767, 22, kSqPs},
+        {768, 23, kSqPs},
+        {769, 24, kSqPs},
+        {770, 25, kSqPs},
+        {771, 26, kSqPs},
+        {772, 27, kSqPs},
+        {773, 28, kSqPs},
+        {774, 29, kSqPs},
+        {775, 30, kSqPs},
+        {776, 31, kSqPs},
+        {777, 32, kSqPs},
+        {778, 33, kSqPs},
+        {779, 34, kSqPs},
+        {780, 35, kSqPs},
+        {781, 36, kSqPs},
+        {782, 37, kSqPs},
+        {783, 38, kSqPs},
+        {784, 39, kSqPs},
+        {785, 40, kSqPs},
+        {786, 41, kSqPs},
+        {787, 42, kSqPs},
+        {788, 43, kSqPs},
+        {789, 44, kSqPs},
+        {790, 45, kSqPs},
+        {791, 46, kSqPs},
+        {792, 47, kSqPs},
+        {793, 0, kSqHs},
+        {794, 1, kSqHs},
+        {795, 2, kSqHs},
+        {796, 3, kSqHs},
+        {797, 4, kSqHs},
+        {798, 5, kSqHs},
+        {799, 6, kSqHs},
+        {800, 7, kSqHs},
+        {801, 8, kSqHs},
+        {802, 9, kSqHs},
+        {803, 10, kSqHs},
+        {804, 11, kSqHs},
+        {805, 12, kSqHs},
+        {806, 13, kSqHs},
+        {807, 14, kSqHs},
+        {808, 15, kSqHs},
+        {809, 16, kSqHs},
+        {810, 17, kSqHs},
+        {811, 18, kSqHs},
+        {812, 19, kSqHs},
+        {813, 20, kSqHs},
+        {814, 21, kSqHs},
+        {815, 22, kSqHs},
+        {816, 23, kSqHs},
+        {817, 24, kSqHs},
+        {818, 25, kSqHs},
+        {819, 26, kSqHs},
+        {820, 27, kSqHs},
+        {821, 28, kSqHs},
+        {822, 29, kSqHs},
+        {823, 30, kSqHs},
+        {824, 31, kSqHs},
+        {825, 32, kSqHs},
+        {826, 33, kSqHs},
+        {827, 34, kSqHs},
+        {828, 35, kSqHs},
+        {829, 36, kSqHs},
+        {830, 37, kSqHs},
+        {831, 38, kSqHs},
+        {832, 39, kSqHs},
+        {833, 40, kSqHs},
+        {834, 41, kSqHs},
+        {835, 42, kSqHs},
+        {836, 43, kSqHs},
+        {837, 44, kSqHs},
+        {838, 45, kSqHs},
+        {839, 46, kSqHs},
+        {840, 47, kSqHs},
+        {841, 0, kSqCs},
+        {842, 1, kSqCs},
+        {843, 2, kSqCs},
+        {844, 3, kSqCs},
+        {845, 4, kSqCs},
+        {846, 5, kSqCs},
+        {847, 6, kSqCs},
+        {848, 7, kSqCs},
+        {849, 8, kSqCs},
+        {850, 9, kSqCs},
+        {851, 10, kSqCs},
+        {852, 11, kSqCs},
+        {853, 12, kSqCs},
+        {854, 13, kSqCs},
+        {855, 14, kSqCs},
+        {856, 15, kSqCs},
+        {857, 16, kSqCs},
+        {858, 17, kSqCs},
+        {859, 18, kSqCs},
+        {860, 19, kSqCs},
+        {861, 20, kSqCs},
+        {862, 21, kSqCs},
+        {863, 22, kSqCs},
+        {864, 23, kSqCs},
+        {865, 24, kSqCs},
+        {866, 25, kSqCs},
+        {867, 26, kSqCs},
+        {868, 27, kSqCs},
+        {869, 28, kSqCs},
+        {870, 29, kSqCs},
+        {871, 30, kSqCs},
+        {872, 31, kSqCs},
+        {873, 32, kSqCs},
+        {874, 33, kSqCs},
+        {875, 34, kSqCs},
+        {876, 35, kSqCs},
+        {877, 36, kSqCs},
+        {878, 37, kSqCs},
+        {879, 38, kSqCs},
+        {880, 39, kSqCs},
+        {881, 40, kSqCs},
+        {882, 41, kSqCs},
+        {883, 42, kSqCs},
+        {884, 43, kSqCs},
+        {885, 44, kSqCs},
+        {886, 45, kSqCs},
+        {887, 46, kSqCs},
+        {888, 47, kSqCs},
     };
 
     unsigned int kHwVkSqIsolatedGroupsGfx11[] = {
@@ -3690,6 +3694,8 @@ namespace counter_vk_gfx11
     VK_GPA_PERF_BLOCK_DMA_AMD,
     VK_GPA_PERF_BLOCK_CPG_AMD,
     VK_GPA_PERF_BLOCK_CPC_AMD,
+    VK_GPA_PERF_BLOCK_ATC_AMD,
+    VK_GPA_PERF_BLOCK_ATC_L2_AMD,
     VK_GPA_PERF_BLOCK_MC_VM_L2_AMD,
     VK_GPA_PERF_BLOCK_EA_AMD,
     VK_GPA_PERF_BLOCK_EA_AMD,
@@ -4154,16 +4160,16 @@ namespace counter_vk_gfx11
     VK_GPA_PERF_BLOCK_RANGE_SIZE_AMD,
     };
 
-    const unsigned int kHwVkGpuTimeBottomToBottomDurationIndexGfx11 = 237514; ///< Index of delta between the previous command reaching bottom of pipe and the current command reaching bottom of pipe, will not include latency of first data to travel through pipeline, best for large data sets. for Vk for Gfx11 family
-    const unsigned int kHwVkGpuTimeBottomToBottomStartIndexGfx11 = 237515; ///< Index of time of the previous command reaching bottom of pipe for Vk for Gfx11 family
-    const unsigned int kHwVkGpuTimeBottomToBottomEndIndexGfx11 = 237516; ///< Index of time of the current command reaching bottom of pipe for Vk for Gfx11 family
-    const unsigned int kHwVkGpuTimeTopToBottomDurationIndexGfx11 = 237517; ///< Index of execution duration of the current command from top of pipe to bottom of pipe, may include overhead of time in queue for Vk for Gfx11 family
-    const unsigned int kHwVkGpuTimeTopToBottomStartIndexGfx11 = 237518; ///< Index of time that the current command reaches the top of pipe for Vk for Gfx11 family
-    const unsigned int kHwVkGpuTimeTopToBottomEndIndexGfx11 = 237519; ///< Index of time that the current command reaches the bottom of pipe for Vk for Gfx11 family
-    const std::set<unsigned int> kHwVkTimestampBlockIdsGfx11 =       {893}; ///< Timestamp block id's for VK for gfx11 family
-    const std::set<unsigned int> kHwVkTimeCounterIndicesGfx11 =      {237514, 237515, 237516, 237517, 237518, 237519}; ///< Timestamp counter indices for VK for gfx11 family
+    const unsigned int kHwVkGpuTimeBottomToBottomDurationIndexGfx11 = 237556; ///< Index of delta between the previous command reaching bottom of pipe and the current command reaching bottom of pipe, will not include latency of first data to travel through pipeline, best for large data sets. for Vk for Gfx11 family
+    const unsigned int kHwVkGpuTimeBottomToBottomStartIndexGfx11 = 237557; ///< Index of time of the previous command reaching bottom of pipe for Vk for Gfx11 family
+    const unsigned int kHwVkGpuTimeBottomToBottomEndIndexGfx11 = 237558; ///< Index of time of the current command reaching bottom of pipe for Vk for Gfx11 family
+    const unsigned int kHwVkGpuTimeTopToBottomDurationIndexGfx11 = 237559; ///< Index of execution duration of the current command from top of pipe to bottom of pipe, may include overhead of time in queue for Vk for Gfx11 family
+    const unsigned int kHwVkGpuTimeTopToBottomStartIndexGfx11 = 237560; ///< Index of time that the current command reaches the top of pipe for Vk for Gfx11 family
+    const unsigned int kHwVkGpuTimeTopToBottomEndIndexGfx11 = 237561; ///< Index of time that the current command reaches the bottom of pipe for Vk for Gfx11 family
+    const std::set<unsigned int> kHwVkTimestampBlockIdsGfx11 =       {895}; ///< Timestamp block id's for VK for gfx11 family
+    const std::set<unsigned int> kHwVkTimeCounterIndicesGfx11 =      {237556, 237557, 237558, 237559, 237560, 237561}; ///< Timestamp counter indices for VK for gfx11 family
     const unsigned int           kHwVkGroupCountGfx11                = static_cast<unsigned int>(kHwVkGroupsGfx11.size());
-    const unsigned int           kVkPaddedCounterGroupCountGfx11 = 246;
+    const unsigned int           kVkPaddedCounterGroupCountGfx11 = 248;
     const unsigned int           kHwVkSqGroupCountGfx11              = sizeof(kHwVkSqGroupsGfx11) / sizeof(GpaSqCounterGroupDesc);
     const unsigned int           kHwVkSqIsolatedGroupCountGfx11      = sizeof(kHwVkSqIsolatedGroupsGfx11) / sizeof(unsigned int);
 } //  namespace counter_vk_gfx11
