@@ -17,7 +17,7 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
 {
     { // Index:0
         static constexpr std::array<GpaUInt32, 1> kHardwareCounters = {
-                235949,
+                235991,
         };
         c.DefineDerivedCounter("GPUTime",
                                "Timing",
@@ -63,7 +63,7 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:3
         static constexpr std::array<GpaUInt32, 2> kHardwareCounters = {
-                112196,
+                112238,
                 24,
         };
         c.DefineDerivedCounter("TessellatorBusy",
@@ -79,7 +79,7 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:4
         static constexpr std::array<GpaUInt32, 1> kHardwareCounters = {
-                112196,
+                112238,
         };
         c.DefineDerivedCounter("TessellatorBusyCycles",
                                "Timing",
@@ -147,7 +147,7 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:7
         static constexpr std::array<GpaUInt32, 14> kHardwareCounters = {
-                235949,
+                235991,
                 36247,
                 36531,
                 36815,
@@ -228,7 +228,7 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:10
         static constexpr std::array<GpaUInt32, 14> kHardwareCounters = {
-                235949,
+                235991,
                 36267,
                 36551,
                 36835,
@@ -309,7 +309,7 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:13
         static constexpr std::array<GpaUInt32, 14> kHardwareCounters = {
-                235949,
+                235991,
                 36247,
                 36531,
                 36815,
@@ -462,7 +462,7 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:16
         static constexpr std::array<GpaUInt32, 50> kHardwareCounters = {
-                235949,
+                235991,
                 36302,
                 36586,
                 36870,
@@ -603,7 +603,7 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:19
         static constexpr std::array<GpaUInt32, 26> kHardwareCounters = {
-                235949,
+                235991,
                 36283,
                 36567,
                 36851,
@@ -643,12 +643,12 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:20
         static constexpr std::array<GpaUInt32, 12> kHardwareCounters = {
-                112270,
-                112358,
-                112446,
-                112534,
-                112622,
-                112710,
+                112312,
+                112400,
+                112488,
+                112576,
+                112664,
+                112752,
                 36272,
                 36556,
                 36840,
@@ -669,12 +669,12 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:21
         static constexpr std::array<GpaUInt32, 12> kHardwareCounters = {
-                112275,
-                112363,
-                112451,
-                112539,
-                112627,
-                112715,
+                112317,
+                112405,
+                112493,
+                112581,
+                112669,
+                112757,
                 36272,
                 36556,
                 36840,
@@ -695,12 +695,12 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:22
         static constexpr std::array<GpaUInt32, 12> kHardwareCounters = {
-                112281,
-                112369,
-                112457,
-                112545,
-                112633,
-                112721,
+                112323,
+                112411,
+                112499,
+                112587,
+                112675,
+                112763,
                 36272,
                 36556,
                 36840,
@@ -721,12 +721,12 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:23
         static constexpr std::array<GpaUInt32, 12> kHardwareCounters = {
-                112258,
-                112346,
-                112434,
-                112522,
-                112610,
-                112698,
+                112300,
+                112388,
+                112476,
+                112564,
+                112652,
+                112740,
                 36272,
                 36556,
                 36840,
@@ -1264,102 +1264,102 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:38
         static constexpr std::array<GpaUInt32, 97> kHardwareCounters = {
-                211119,
-                211631,
-                212143,
-                212655,
-                213167,
-                213679,
-                214191,
-                214703,
-                215215,
-                215727,
-                216239,
-                216751,
-                217263,
-                217775,
-                218287,
-                218799,
-                219311,
-                219823,
-                220335,
-                220847,
-                221359,
-                221871,
-                222383,
-                222895,
-                223407,
-                223919,
-                224431,
-                224943,
-                225455,
-                225967,
-                226479,
-                226991,
-                227503,
-                228015,
-                228527,
-                229039,
-                229551,
-                230063,
-                230575,
-                231087,
-                231599,
-                232111,
-                232623,
-                233135,
-                233647,
-                234159,
-                234671,
-                235183,
-                211094,
-                211606,
-                212118,
-                212630,
-                213142,
-                213654,
-                214166,
-                214678,
-                215190,
-                215702,
-                216214,
-                216726,
-                217238,
-                217750,
-                218262,
-                218774,
-                219286,
-                219798,
-                220310,
-                220822,
-                221334,
-                221846,
-                222358,
-                222870,
-                223382,
-                223894,
-                224406,
-                224918,
-                225430,
-                225942,
-                226454,
-                226966,
-                227478,
-                227990,
-                228502,
-                229014,
-                229526,
-                230038,
-                230550,
-                231062,
-                231574,
-                232086,
-                232598,
-                233110,
-                233622,
-                234134,
-                234646,
-                235158,
+                211161,
+                211673,
+                212185,
+                212697,
+                213209,
+                213721,
+                214233,
+                214745,
+                215257,
+                215769,
+                216281,
+                216793,
+                217305,
+                217817,
+                218329,
+                218841,
+                219353,
+                219865,
+                220377,
+                220889,
+                221401,
+                221913,
+                222425,
+                222937,
+                223449,
+                223961,
+                224473,
+                224985,
+                225497,
+                226009,
+                226521,
+                227033,
+                227545,
+                228057,
+                228569,
+                229081,
+                229593,
+                230105,
+                230617,
+                231129,
+                231641,
+                232153,
+                232665,
+                233177,
+                233689,
+                234201,
+                234713,
+                235225,
+                211136,
+                211648,
+                212160,
+                212672,
+                213184,
+                213696,
+                214208,
+                214720,
+                215232,
+                215744,
+                216256,
+                216768,
+                217280,
+                217792,
+                218304,
+                218816,
+                219328,
+                219840,
+                220352,
+                220864,
+                221376,
+                221888,
+                222400,
+                222912,
+                223424,
+                223936,
+                224448,
+                224960,
+                225472,
+                225984,
+                226496,
+                227008,
+                227520,
+                228032,
+                228544,
+                229056,
+                229568,
+                230080,
+                230592,
+                231104,
+                231616,
+                232128,
+                232640,
+                233152,
+                233664,
+                234176,
+                234688,
+                235200,
                 24,
         };
         c.DefineDerivedCounter("CSALUStalledByLDS",
@@ -1375,102 +1375,102 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:39
         static constexpr std::array<GpaUInt32, 96> kHardwareCounters = {
-                211119,
-                211631,
-                212143,
-                212655,
-                213167,
-                213679,
-                214191,
-                214703,
-                215215,
-                215727,
-                216239,
-                216751,
-                217263,
-                217775,
-                218287,
-                218799,
-                219311,
-                219823,
-                220335,
-                220847,
-                221359,
-                221871,
-                222383,
-                222895,
-                223407,
-                223919,
-                224431,
-                224943,
-                225455,
-                225967,
-                226479,
-                226991,
-                227503,
-                228015,
-                228527,
-                229039,
-                229551,
-                230063,
-                230575,
-                231087,
-                231599,
-                232111,
-                232623,
-                233135,
-                233647,
-                234159,
-                234671,
-                235183,
-                211094,
-                211606,
-                212118,
-                212630,
-                213142,
-                213654,
-                214166,
-                214678,
-                215190,
-                215702,
-                216214,
-                216726,
-                217238,
-                217750,
-                218262,
-                218774,
-                219286,
-                219798,
-                220310,
-                220822,
-                221334,
-                221846,
-                222358,
-                222870,
-                223382,
-                223894,
-                224406,
-                224918,
-                225430,
-                225942,
-                226454,
-                226966,
-                227478,
-                227990,
-                228502,
-                229014,
-                229526,
-                230038,
-                230550,
-                231062,
-                231574,
-                232086,
-                232598,
-                233110,
-                233622,
-                234134,
-                234646,
-                235158,
+                211161,
+                211673,
+                212185,
+                212697,
+                213209,
+                213721,
+                214233,
+                214745,
+                215257,
+                215769,
+                216281,
+                216793,
+                217305,
+                217817,
+                218329,
+                218841,
+                219353,
+                219865,
+                220377,
+                220889,
+                221401,
+                221913,
+                222425,
+                222937,
+                223449,
+                223961,
+                224473,
+                224985,
+                225497,
+                226009,
+                226521,
+                227033,
+                227545,
+                228057,
+                228569,
+                229081,
+                229593,
+                230105,
+                230617,
+                231129,
+                231641,
+                232153,
+                232665,
+                233177,
+                233689,
+                234201,
+                234713,
+                235225,
+                211136,
+                211648,
+                212160,
+                212672,
+                213184,
+                213696,
+                214208,
+                214720,
+                215232,
+                215744,
+                216256,
+                216768,
+                217280,
+                217792,
+                218304,
+                218816,
+                219328,
+                219840,
+                220352,
+                220864,
+                221376,
+                221888,
+                222400,
+                222912,
+                223424,
+                223936,
+                224448,
+                224960,
+                225472,
+                225984,
+                226496,
+                227008,
+                227520,
+                228032,
+                228544,
+                229056,
+                229568,
+                230080,
+                230592,
+                231104,
+                231616,
+                232128,
+                232640,
+                233152,
+                233664,
+                234176,
+                234688,
+                235200,
         };
         c.DefineDerivedCounter("CSALUStalledByLDSCycles",
                                "ComputeShader",
@@ -1485,54 +1485,54 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:40
         static constexpr std::array<GpaUInt32, 49> kHardwareCounters = {
-                211346,
-                211858,
-                212370,
-                212882,
-                213394,
-                213906,
-                214418,
-                214930,
-                215442,
-                215954,
-                216466,
-                216978,
-                217490,
-                218002,
-                218514,
-                219026,
-                219538,
-                220050,
-                220562,
-                221074,
-                221586,
-                222098,
-                222610,
-                223122,
-                223634,
-                224146,
-                224658,
-                225170,
-                225682,
-                226194,
-                226706,
-                227218,
-                227730,
-                228242,
-                228754,
-                229266,
-                229778,
-                230290,
-                230802,
-                231314,
-                231826,
-                232338,
-                232850,
-                233362,
-                233874,
-                234386,
-                234898,
-                235410,
+                211388,
+                211900,
+                212412,
+                212924,
+                213436,
+                213948,
+                214460,
+                214972,
+                215484,
+                215996,
+                216508,
+                217020,
+                217532,
+                218044,
+                218556,
+                219068,
+                219580,
+                220092,
+                220604,
+                221116,
+                221628,
+                222140,
+                222652,
+                223164,
+                223676,
+                224188,
+                224700,
+                225212,
+                225724,
+                226236,
+                226748,
+                227260,
+                227772,
+                228284,
+                228796,
+                229308,
+                229820,
+                230332,
+                230844,
+                231356,
+                231868,
+                232380,
+                232892,
+                233404,
+                233916,
+                234428,
+                234940,
+                235452,
                 24,
         };
         c.DefineDerivedCounter("CSLDSBankConflict",
@@ -1548,54 +1548,54 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:41
         static constexpr std::array<GpaUInt32, 48> kHardwareCounters = {
-                211346,
-                211858,
-                212370,
-                212882,
-                213394,
-                213906,
-                214418,
-                214930,
-                215442,
-                215954,
-                216466,
-                216978,
-                217490,
-                218002,
-                218514,
-                219026,
-                219538,
-                220050,
-                220562,
-                221074,
-                221586,
-                222098,
-                222610,
-                223122,
-                223634,
-                224146,
-                224658,
-                225170,
-                225682,
-                226194,
-                226706,
-                227218,
-                227730,
-                228242,
-                228754,
-                229266,
-                229778,
-                230290,
-                230802,
-                231314,
-                231826,
-                232338,
-                232850,
-                233362,
-                233874,
-                234386,
-                234898,
-                235410,
+                211388,
+                211900,
+                212412,
+                212924,
+                213436,
+                213948,
+                214460,
+                214972,
+                215484,
+                215996,
+                216508,
+                217020,
+                217532,
+                218044,
+                218556,
+                219068,
+                219580,
+                220092,
+                220604,
+                221116,
+                221628,
+                222140,
+                222652,
+                223164,
+                223676,
+                224188,
+                224700,
+                225212,
+                225724,
+                226236,
+                226748,
+                227260,
+                227772,
+                228284,
+                228796,
+                229308,
+                229820,
+                230332,
+                230844,
+                231356,
+                231868,
+                232380,
+                232892,
+                233404,
+                233916,
+                234428,
+                234940,
+                235452,
         };
         c.DefineDerivedCounter("CSLDSBankConflictCycles",
                                "ComputeShader",
@@ -5447,102 +5447,102 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:81
         static constexpr std::array<GpaUInt32, 96> kHardwareCounters = {
-                99499,
-                99583,
-                99667,
-                99751,
-                99835,
-                99919,
-                100003,
-                100087,
-                100171,
-                100255,
-                100339,
-                100423,
-                100507,
-                100591,
-                100675,
-                100759,
-                100843,
-                100927,
-                101011,
-                101095,
-                101179,
-                101263,
-                101347,
-                101431,
-                101515,
-                101599,
-                101683,
-                101767,
-                101851,
-                101935,
-                102019,
-                102103,
-                102187,
-                102271,
-                102355,
-                102439,
-                102523,
-                102607,
-                102691,
-                102775,
-                102859,
-                102943,
-                103027,
-                103111,
-                103195,
-                103279,
-                103363,
-                103447,
-                99503,
-                99587,
-                99671,
-                99755,
-                99839,
-                99923,
-                100007,
-                100091,
-                100175,
-                100259,
-                100343,
-                100427,
-                100511,
-                100595,
-                100679,
-                100763,
-                100847,
-                100931,
-                101015,
-                101099,
-                101183,
-                101267,
-                101351,
-                101435,
-                101519,
-                101603,
-                101687,
-                101771,
-                101855,
-                101939,
-                102023,
-                102107,
-                102191,
-                102275,
-                102359,
-                102443,
-                102527,
-                102611,
-                102695,
-                102779,
-                102863,
-                102947,
-                103031,
-                103115,
-                103199,
-                103283,
-                103367,
-                103451,
+                99541,
+                99625,
+                99709,
+                99793,
+                99877,
+                99961,
+                100045,
+                100129,
+                100213,
+                100297,
+                100381,
+                100465,
+                100549,
+                100633,
+                100717,
+                100801,
+                100885,
+                100969,
+                101053,
+                101137,
+                101221,
+                101305,
+                101389,
+                101473,
+                101557,
+                101641,
+                101725,
+                101809,
+                101893,
+                101977,
+                102061,
+                102145,
+                102229,
+                102313,
+                102397,
+                102481,
+                102565,
+                102649,
+                102733,
+                102817,
+                102901,
+                102985,
+                103069,
+                103153,
+                103237,
+                103321,
+                103405,
+                103489,
+                99545,
+                99629,
+                99713,
+                99797,
+                99881,
+                99965,
+                100049,
+                100133,
+                100217,
+                100301,
+                100385,
+                100469,
+                100553,
+                100637,
+                100721,
+                100805,
+                100889,
+                100973,
+                101057,
+                101141,
+                101225,
+                101309,
+                101393,
+                101477,
+                101561,
+                101645,
+                101729,
+                101813,
+                101897,
+                101981,
+                102065,
+                102149,
+                102233,
+                102317,
+                102401,
+                102485,
+                102569,
+                102653,
+                102737,
+                102821,
+                102905,
+                102989,
+                103073,
+                103157,
+                103241,
+                103325,
+                103409,
+                103493,
         };
         c.DefineDerivedCounter("L1CacheHit",
                                "MemoryCache",
@@ -5557,54 +5557,54 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:82
         static constexpr std::array<GpaUInt32, 48> kHardwareCounters = {
-                99499,
-                99583,
-                99667,
-                99751,
-                99835,
-                99919,
-                100003,
-                100087,
-                100171,
-                100255,
-                100339,
-                100423,
-                100507,
-                100591,
-                100675,
-                100759,
-                100843,
-                100927,
-                101011,
-                101095,
-                101179,
-                101263,
-                101347,
-                101431,
-                101515,
-                101599,
-                101683,
-                101767,
-                101851,
-                101935,
-                102019,
-                102103,
-                102187,
-                102271,
-                102355,
-                102439,
-                102523,
-                102607,
-                102691,
-                102775,
-                102859,
-                102943,
-                103027,
-                103111,
-                103195,
-                103279,
-                103363,
-                103447,
+                99541,
+                99625,
+                99709,
+                99793,
+                99877,
+                99961,
+                100045,
+                100129,
+                100213,
+                100297,
+                100381,
+                100465,
+                100549,
+                100633,
+                100717,
+                100801,
+                100885,
+                100969,
+                101053,
+                101137,
+                101221,
+                101305,
+                101389,
+                101473,
+                101557,
+                101641,
+                101725,
+                101809,
+                101893,
+                101977,
+                102061,
+                102145,
+                102229,
+                102313,
+                102397,
+                102481,
+                102565,
+                102649,
+                102733,
+                102817,
+                102901,
+                102985,
+                103069,
+                103153,
+                103237,
+                103321,
+                103405,
+                103489,
         };
         c.DefineDerivedCounter("L1CacheRequestCount",
                                "MemoryCache",
@@ -5619,102 +5619,102 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:83
         static constexpr std::array<GpaUInt32, 96> kHardwareCounters = {
-                99499,
-                99583,
-                99667,
-                99751,
-                99835,
-                99919,
-                100003,
-                100087,
-                100171,
-                100255,
-                100339,
-                100423,
-                100507,
-                100591,
-                100675,
-                100759,
-                100843,
-                100927,
-                101011,
-                101095,
-                101179,
-                101263,
-                101347,
-                101431,
-                101515,
-                101599,
-                101683,
-                101767,
-                101851,
-                101935,
-                102019,
-                102103,
-                102187,
-                102271,
-                102355,
-                102439,
-                102523,
-                102607,
-                102691,
-                102775,
-                102859,
-                102943,
-                103027,
-                103111,
-                103195,
-                103279,
-                103363,
-                103447,
-                99503,
-                99587,
-                99671,
-                99755,
-                99839,
-                99923,
-                100007,
-                100091,
-                100175,
-                100259,
-                100343,
-                100427,
-                100511,
-                100595,
-                100679,
-                100763,
-                100847,
-                100931,
-                101015,
-                101099,
-                101183,
-                101267,
-                101351,
-                101435,
-                101519,
-                101603,
-                101687,
-                101771,
-                101855,
-                101939,
-                102023,
-                102107,
-                102191,
-                102275,
-                102359,
-                102443,
-                102527,
-                102611,
-                102695,
-                102779,
-                102863,
-                102947,
-                103031,
-                103115,
-                103199,
-                103283,
-                103367,
-                103451,
+                99541,
+                99625,
+                99709,
+                99793,
+                99877,
+                99961,
+                100045,
+                100129,
+                100213,
+                100297,
+                100381,
+                100465,
+                100549,
+                100633,
+                100717,
+                100801,
+                100885,
+                100969,
+                101053,
+                101137,
+                101221,
+                101305,
+                101389,
+                101473,
+                101557,
+                101641,
+                101725,
+                101809,
+                101893,
+                101977,
+                102061,
+                102145,
+                102229,
+                102313,
+                102397,
+                102481,
+                102565,
+                102649,
+                102733,
+                102817,
+                102901,
+                102985,
+                103069,
+                103153,
+                103237,
+                103321,
+                103405,
+                103489,
+                99545,
+                99629,
+                99713,
+                99797,
+                99881,
+                99965,
+                100049,
+                100133,
+                100217,
+                100301,
+                100385,
+                100469,
+                100553,
+                100637,
+                100721,
+                100805,
+                100889,
+                100973,
+                101057,
+                101141,
+                101225,
+                101309,
+                101393,
+                101477,
+                101561,
+                101645,
+                101729,
+                101813,
+                101897,
+                101981,
+                102065,
+                102149,
+                102233,
+                102317,
+                102401,
+                102485,
+                102569,
+                102653,
+                102737,
+                102821,
+                102905,
+                102989,
+                103073,
+                103157,
+                103241,
+                103325,
+                103409,
+                103493,
         };
         c.DefineDerivedCounter("L1CacheHitCount",
                                "MemoryCache",
@@ -5729,54 +5729,54 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:84
         static constexpr std::array<GpaUInt32, 48> kHardwareCounters = {
-                99503,
-                99587,
-                99671,
-                99755,
-                99839,
-                99923,
-                100007,
-                100091,
-                100175,
-                100259,
-                100343,
-                100427,
-                100511,
-                100595,
-                100679,
-                100763,
-                100847,
-                100931,
-                101015,
-                101099,
-                101183,
-                101267,
-                101351,
-                101435,
-                101519,
-                101603,
-                101687,
-                101771,
-                101855,
-                101939,
-                102023,
-                102107,
-                102191,
-                102275,
-                102359,
-                102443,
-                102527,
-                102611,
-                102695,
-                102779,
-                102863,
-                102947,
-                103031,
-                103115,
-                103199,
-                103283,
-                103367,
-                103451,
+                99545,
+                99629,
+                99713,
+                99797,
+                99881,
+                99965,
+                100049,
+                100133,
+                100217,
+                100301,
+                100385,
+                100469,
+                100553,
+                100637,
+                100721,
+                100805,
+                100889,
+                100973,
+                101057,
+                101141,
+                101225,
+                101309,
+                101393,
+                101477,
+                101561,
+                101645,
+                101729,
+                101813,
+                101897,
+                101981,
+                102065,
+                102149,
+                102233,
+                102317,
+                102401,
+                102485,
+                102569,
+                102653,
+                102737,
+                102821,
+                102905,
+                102989,
+                103073,
+                103157,
+                103241,
+                103325,
+                103409,
+                103493,
         };
         c.DefineDerivedCounter("L1CacheMissCount",
                                "MemoryCache",
@@ -5791,54 +5791,54 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:85
         static constexpr std::array<GpaUInt32, 48> kHardwareCounters = {
-                103952,
-                104211,
-                104470,
-                104729,
-                104988,
-                105247,
-                105506,
-                105765,
-                106024,
-                106283,
-                106542,
-                106801,
-                107060,
-                107319,
-                107578,
-                107837,
-                108096,
-                108355,
-                108614,
-                108873,
-                109132,
-                109391,
-                109650,
-                109909,
-                103992,
-                104251,
-                104510,
-                104769,
-                105028,
-                105287,
-                105546,
-                105805,
-                106064,
-                106323,
-                106582,
-                106841,
-                107100,
-                107359,
-                107618,
-                107877,
-                108136,
-                108395,
-                108654,
-                108913,
-                109172,
-                109431,
-                109690,
-                109949,
+                103994,
+                104253,
+                104512,
+                104771,
+                105030,
+                105289,
+                105548,
+                105807,
+                106066,
+                106325,
+                106584,
+                106843,
+                107102,
+                107361,
+                107620,
+                107879,
+                108138,
+                108397,
+                108656,
+                108915,
+                109174,
+                109433,
+                109692,
+                109951,
+                104034,
+                104293,
+                104552,
+                104811,
+                105070,
+                105329,
+                105588,
+                105847,
+                106106,
+                106365,
+                106624,
+                106883,
+                107142,
+                107401,
+                107660,
+                107919,
+                108178,
+                108437,
+                108696,
+                108955,
+                109214,
+                109473,
+                109732,
+                109991,
         };
         c.DefineDerivedCounter("L2CacheHit",
                                "MemoryCache",
@@ -5853,54 +5853,54 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:86
         static constexpr std::array<GpaUInt32, 48> kHardwareCounters = {
-                103992,
-                104251,
-                104510,
-                104769,
-                105028,
-                105287,
-                105546,
-                105805,
-                106064,
-                106323,
-                106582,
-                106841,
-                107100,
-                107359,
-                107618,
-                107877,
-                108136,
-                108395,
-                108654,
-                108913,
-                109172,
-                109431,
-                109690,
-                109949,
-                103952,
-                104211,
-                104470,
-                104729,
-                104988,
-                105247,
-                105506,
-                105765,
-                106024,
-                106283,
-                106542,
-                106801,
-                107060,
-                107319,
-                107578,
-                107837,
-                108096,
-                108355,
-                108614,
-                108873,
-                109132,
-                109391,
-                109650,
-                109909,
+                104034,
+                104293,
+                104552,
+                104811,
+                105070,
+                105329,
+                105588,
+                105847,
+                106106,
+                106365,
+                106624,
+                106883,
+                107142,
+                107401,
+                107660,
+                107919,
+                108178,
+                108437,
+                108696,
+                108955,
+                109214,
+                109473,
+                109732,
+                109991,
+                103994,
+                104253,
+                104512,
+                104771,
+                105030,
+                105289,
+                105548,
+                105807,
+                106066,
+                106325,
+                106584,
+                106843,
+                107102,
+                107361,
+                107620,
+                107879,
+                108138,
+                108397,
+                108656,
+                108915,
+                109174,
+                109433,
+                109692,
+                109951,
         };
         c.DefineDerivedCounter("L2CacheMiss",
                                "MemoryCache",
@@ -5915,30 +5915,30 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:87
         static constexpr std::array<GpaUInt32, 24> kHardwareCounters = {
-                103952,
-                104211,
-                104470,
-                104729,
-                104988,
-                105247,
-                105506,
-                105765,
-                106024,
-                106283,
-                106542,
-                106801,
-                107060,
-                107319,
-                107578,
-                107837,
-                108096,
-                108355,
-                108614,
-                108873,
-                109132,
-                109391,
-                109650,
-                109909,
+                103994,
+                104253,
+                104512,
+                104771,
+                105030,
+                105289,
+                105548,
+                105807,
+                106066,
+                106325,
+                106584,
+                106843,
+                107102,
+                107361,
+                107620,
+                107879,
+                108138,
+                108397,
+                108656,
+                108915,
+                109174,
+                109433,
+                109692,
+                109951,
         };
         c.DefineDerivedCounter("L2CacheRequestCount",
                                "MemoryCache",
@@ -5953,54 +5953,54 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:88
         static constexpr std::array<GpaUInt32, 48> kHardwareCounters = {
-                103952,
-                104211,
-                104470,
-                104729,
-                104988,
-                105247,
-                105506,
-                105765,
-                106024,
-                106283,
-                106542,
-                106801,
-                107060,
-                107319,
-                107578,
-                107837,
-                108096,
-                108355,
-                108614,
-                108873,
-                109132,
-                109391,
-                109650,
-                109909,
-                103992,
-                104251,
-                104510,
-                104769,
-                105028,
-                105287,
-                105546,
-                105805,
-                106064,
-                106323,
-                106582,
-                106841,
-                107100,
-                107359,
-                107618,
-                107877,
-                108136,
-                108395,
-                108654,
-                108913,
-                109172,
-                109431,
-                109690,
-                109949,
+                103994,
+                104253,
+                104512,
+                104771,
+                105030,
+                105289,
+                105548,
+                105807,
+                106066,
+                106325,
+                106584,
+                106843,
+                107102,
+                107361,
+                107620,
+                107879,
+                108138,
+                108397,
+                108656,
+                108915,
+                109174,
+                109433,
+                109692,
+                109951,
+                104034,
+                104293,
+                104552,
+                104811,
+                105070,
+                105329,
+                105588,
+                105847,
+                106106,
+                106365,
+                106624,
+                106883,
+                107142,
+                107401,
+                107660,
+                107919,
+                108178,
+                108437,
+                108696,
+                108955,
+                109214,
+                109473,
+                109732,
+                109991,
         };
         c.DefineDerivedCounter("L2CacheHitCount",
                                "MemoryCache",
@@ -6015,30 +6015,30 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:89
         static constexpr std::array<GpaUInt32, 24> kHardwareCounters = {
-                103992,
-                104251,
-                104510,
-                104769,
-                105028,
-                105287,
-                105546,
-                105805,
-                106064,
-                106323,
-                106582,
-                106841,
-                107100,
-                107359,
-                107618,
-                107877,
-                108136,
-                108395,
-                108654,
-                108913,
-                109172,
-                109431,
-                109690,
-                109949,
+                104034,
+                104293,
+                104552,
+                104811,
+                105070,
+                105329,
+                105588,
+                105847,
+                106106,
+                106365,
+                106624,
+                106883,
+                107142,
+                107401,
+                107660,
+                107919,
+                108178,
+                108437,
+                108696,
+                108955,
+                109214,
+                109473,
+                109732,
+                109991,
         };
         c.DefineDerivedCounter("L2CacheMissCount",
                                "MemoryCache",
@@ -6383,102 +6383,102 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:93
         static constexpr std::array<GpaUInt32, 96> kHardwareCounters = {
-                104048,
-                104307,
-                104566,
-                104825,
-                105084,
-                105343,
-                105602,
-                105861,
-                106120,
-                106379,
-                106638,
-                106897,
-                107156,
-                107415,
-                107674,
-                107933,
-                108192,
-                108451,
-                108710,
-                108969,
-                109228,
-                109487,
-                109746,
-                110005,
-                104049,
-                104308,
-                104567,
-                104826,
-                105085,
-                105344,
-                105603,
-                105862,
-                106121,
-                106380,
-                106639,
-                106898,
-                107157,
-                107416,
-                107675,
-                107934,
-                108193,
-                108452,
-                108711,
-                108970,
-                109229,
-                109488,
-                109747,
-                110006,
-                104050,
-                104309,
-                104568,
-                104827,
-                105086,
-                105345,
-                105604,
-                105863,
-                106122,
-                106381,
-                106640,
-                106899,
-                107158,
-                107417,
-                107676,
-                107935,
-                108194,
-                108453,
-                108712,
-                108971,
-                109230,
-                109489,
-                109748,
-                110007,
-                104051,
-                104310,
-                104569,
-                104828,
-                105087,
-                105346,
-                105605,
-                105864,
-                106123,
-                106382,
-                106641,
-                106900,
-                107159,
-                107418,
-                107677,
-                107936,
-                108195,
-                108454,
-                108713,
-                108972,
-                109231,
-                109490,
-                109749,
-                110008,
+                104090,
+                104349,
+                104608,
+                104867,
+                105126,
+                105385,
+                105644,
+                105903,
+                106162,
+                106421,
+                106680,
+                106939,
+                107198,
+                107457,
+                107716,
+                107975,
+                108234,
+                108493,
+                108752,
+                109011,
+                109270,
+                109529,
+                109788,
+                110047,
+                104091,
+                104350,
+                104609,
+                104868,
+                105127,
+                105386,
+                105645,
+                105904,
+                106163,
+                106422,
+                106681,
+                106940,
+                107199,
+                107458,
+                107717,
+                107976,
+                108235,
+                108494,
+                108753,
+                109012,
+                109271,
+                109530,
+                109789,
+                110048,
+                104092,
+                104351,
+                104610,
+                104869,
+                105128,
+                105387,
+                105646,
+                105905,
+                106164,
+                106423,
+                106682,
+                106941,
+                107200,
+                107459,
+                107718,
+                107977,
+                108236,
+                108495,
+                108754,
+                109013,
+                109272,
+                109531,
+                109790,
+                110049,
+                104093,
+                104352,
+                104611,
+                104870,
+                105129,
+                105388,
+                105647,
+                105906,
+                106165,
+                106424,
+                106683,
+                106942,
+                107201,
+                107460,
+                107719,
+                107978,
+                108237,
+                108496,
+                108755,
+                109014,
+                109273,
+                109532,
+                109791,
+                110050,
         };
         c.DefineDerivedCounter("FetchSize",
                                "GlobalMemory",
@@ -6493,54 +6493,54 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:94
         static constexpr std::array<GpaUInt32, 48> kHardwareCounters = {
-                104032,
-                104291,
-                104550,
-                104809,
-                105068,
-                105327,
-                105586,
-                105845,
-                106104,
-                106363,
-                106622,
-                106881,
-                107140,
-                107399,
-                107658,
-                107917,
-                108176,
-                108435,
-                108694,
-                108953,
-                109212,
-                109471,
-                109730,
-                109989,
-                104034,
-                104293,
-                104552,
-                104811,
-                105070,
-                105329,
-                105588,
-                105847,
-                106106,
-                106365,
-                106624,
-                106883,
-                107142,
-                107401,
-                107660,
-                107919,
-                108178,
-                108437,
-                108696,
-                108955,
-                109214,
-                109473,
-                109732,
-                109991,
+                104074,
+                104333,
+                104592,
+                104851,
+                105110,
+                105369,
+                105628,
+                105887,
+                106146,
+                106405,
+                106664,
+                106923,
+                107182,
+                107441,
+                107700,
+                107959,
+                108218,
+                108477,
+                108736,
+                108995,
+                109254,
+                109513,
+                109772,
+                110031,
+                104076,
+                104335,
+                104594,
+                104853,
+                105112,
+                105371,
+                105630,
+                105889,
+                106148,
+                106407,
+                106666,
+                106925,
+                107184,
+                107443,
+                107702,
+                107961,
+                108220,
+                108479,
+                108738,
+                108997,
+                109256,
+                109515,
+                109774,
+                110033,
         };
         c.DefineDerivedCounter("WriteSize",
                                "GlobalMemory",
@@ -6997,30 +6997,30 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:99
         static constexpr std::array<GpaUInt32, 25> kHardwareCounters = {
-                104037,
-                104296,
-                104555,
-                104814,
-                105073,
-                105332,
-                105591,
-                105850,
-                106109,
-                106368,
-                106627,
-                106886,
-                107145,
-                107404,
-                107663,
-                107922,
-                108181,
-                108440,
-                108699,
-                108958,
-                109217,
-                109476,
-                109735,
-                109994,
+                104079,
+                104338,
+                104597,
+                104856,
+                105115,
+                105374,
+                105633,
+                105892,
+                106151,
+                106410,
+                106669,
+                106928,
+                107187,
+                107446,
+                107705,
+                107964,
+                108223,
+                108482,
+                108741,
+                109000,
+                109259,
+                109518,
+                109777,
+                110036,
                 24,
         };
         c.DefineDerivedCounter("WriteUnitStalled",
@@ -7036,30 +7036,30 @@ void AutoDefinePublicDerivedCountersDx11Gfx11(GpaDerivedCounters& c)
     }
     { // Index:100
         static constexpr std::array<GpaUInt32, 24> kHardwareCounters = {
-                104037,
-                104296,
-                104555,
-                104814,
-                105073,
-                105332,
-                105591,
-                105850,
-                106109,
-                106368,
-                106627,
-                106886,
-                107145,
-                107404,
-                107663,
-                107922,
-                108181,
-                108440,
-                108699,
-                108958,
-                109217,
-                109476,
-                109735,
-                109994,
+                104079,
+                104338,
+                104597,
+                104856,
+                105115,
+                105374,
+                105633,
+                105892,
+                106151,
+                106410,
+                106669,
+                106928,
+                107187,
+                107446,
+                107705,
+                107964,
+                108223,
+                108482,
+                108741,
+                109000,
+                109259,
+                109518,
+                109777,
+                110036,
         };
         c.DefineDerivedCounter("WriteUnitStalledCycles",
                                "GlobalMemory",
