@@ -7,14 +7,6 @@
 
 #include "gpu_perf_api_common/gpa_implementor.h"
 
-#if defined(WIN32)
-#include "gpu_perf_api_common/adapter.h"
-#endif
-
-#ifdef _WIN32
-#include "ADLUtil.h"
-#endif
-
 #include "gpu_perf_api_common/gpa_context_counter_mediator.h"
 #include "gpu_perf_api_common/gpa_unique_object.h"
 #include "gpu_perf_api_common/logging.h"
@@ -41,17 +33,8 @@ GpaStatus GpaImplementor::Initialize(GpaInitializeFlags flags)
     {
         if (!is_initialized_)
         {
-            is_initialized_                 = true;
-            GpaContextCounterMediator* temp = GpaContextCounterMediator::Instance();
-
-            if (nullptr == temp)
-            {
-                gpa_status = kGpaStatusErrorFailed;
-            }
-            else
-            {
-                gpa_status = kGpaStatusOk;
-            }
+            is_initialized_ = true;
+            gpa_status      = GpaContextCounterMediator::Initialize(&session_info_map_);
         }
     }
 
@@ -65,8 +48,9 @@ GpaStatus GpaImplementor::Destroy()
     if (is_initialized_)
     {
         is_initialized_ = false;
-        GpaContextCounterMediator::DeleteInstance();
         gpa_status = kGpaStatusOk;
+        GpaContextCounterMediator::Clear();
+        session_info_map_.clear();
         app_context_info_gpa_context_map_.clear();
     }
 
@@ -112,7 +96,7 @@ GpaStatus GpaImplementor::OpenContext(void* context, GpaOpenContextFlags flags, 
     {
         GpaHwInfo hw_info;
 
-        const GpaDriverInfo driver_info = QueryDriverInfo();
+        const GpaDriverInfo driver_info = GpaQueryDriverInfo();
         if (!IsDriverSupported(context))
         {
             // Driver not supported, logging error.
@@ -132,7 +116,7 @@ GpaStatus GpaImplementor::OpenContext(void* context, GpaOpenContextFlags flags, 
 
             if (nullptr != new_gpa_context)
             {
-                *gpa_context_id = reinterpret_cast<GpaContextId>(GpaUniqueObjectManager::Instance()->CreateObject(new_gpa_context));
+                *gpa_context_id = reinterpret_cast<GpaContextId>(GpaUniqueObjectManager::Instance().CreateObject(new_gpa_context));
                 app_context_info_gpa_context_map_.insert(GpaDeviceIdentifierGpaContextPair(GetDeviceIdentifierFromContextInfo(context), new_gpa_context));
             }
             else
@@ -179,7 +163,7 @@ GpaStatus GpaImplementor::CloseContext(GpaContextId gpa_context_id)
             if (CloseApiContext(gpa_context))
             {
                 app_context_info_gpa_context_map_.erase(found_iter);
-                GpaUniqueObjectManager::Instance()->DeleteObject(gpa_context_id);
+                GpaUniqueObjectManager::Instance().DeleteObject(gpa_context_id);
             }
             else
             {
@@ -213,7 +197,7 @@ bool GpaImplementor::DoesContextExist(GpaContextId gpa_context_id) const
 
     if (nullptr != gpa_context_id)
     {
-        context_found = GpaUniqueObjectManager::Instance()->DoesExist(gpa_context_id);
+        context_found = GpaUniqueObjectManager::Instance().DoesExist(gpa_context_id);
 
         if (context_found && GpaObjectType::kGpaObjectTypeContext == gpa_context_id->ObjectType() && GetApiType() == gpa_context_id->Object()->GetApiType())
         {
@@ -230,7 +214,7 @@ bool GpaImplementor::DoesSessionExist(GpaSessionId gpa_session_id) const
 
     if (nullptr != gpa_session_id)
     {
-        session_found = GpaUniqueObjectManager::Instance()->DoesExist(gpa_session_id);
+        session_found = GpaUniqueObjectManager::Instance().DoesExist(gpa_session_id);
 
         if (session_found && GpaObjectType::kGpaObjectTypeSession == gpa_session_id->ObjectType())
         {
@@ -247,7 +231,7 @@ bool GpaImplementor::DoesCommandListExist(GpaCommandListId command_list_id) cons
 
     if (nullptr != command_list_id)
     {
-        command_list_found = GpaUniqueObjectManager::Instance()->DoesExist(command_list_id);
+        command_list_found = GpaUniqueObjectManager::Instance().DoesExist(command_list_id);
 
         if (command_list_found && GpaObjectType::kGpaObjectTypeCommandList == command_list_id->ObjectType())
         {
@@ -283,29 +267,11 @@ bool GpaImplementor::DoesContextInfoExist(GpaContextInfoPtr context_info) const
     return app_context_info_gpa_context_map_.find(GetDeviceIdentifierFromContextInfo(context_info)) != app_context_info_gpa_context_map_.cend();
 }
 
-bool GpaImplementor::IsDeviceGenerationSupported(const GpaHwInfo& hw_info) const
-{
-    GDT_HW_GENERATION device_generation = GDT_HW_GENERATION_NONE;
-    if (hw_info.GetHwGeneration(device_generation))
-    {
-        if (device_generation >= GDT_HW_GENERATION_GFX10)
-        {
-            return true;
-        }
-    }
-
-    return false;
-}
-
 GpaStatus GpaImplementor::IsDeviceSupported(GpaContextInfoPtr    context_info,
                                             GpaOpenContextFlags  flags,
                                             GpaDriverInfo const& driver_info,
                                             GpaHwInfo*           hw_info) const
 {
-    bool         found_matching_hw_info = false;
-#ifdef _WIN32
-    AsicInfoList asic_info_list;
-#endif
     GpaHwInfo    api_hw_info;
 
     if (!GetHwInfoFromApi(context_info, flags, api_hw_info))
@@ -325,114 +291,6 @@ GpaStatus GpaImplementor::IsDeviceSupported(GpaContextInfoPtr    context_info,
             return kGpaStatusErrorHardwareNotSupported;
         }
 
-#ifdef _WIN32
-        AMDTADLUtils::Instance()->GetAsicInfoList(asic_info_list);
-        AMDTADLUtils::DeleteInstance();
-        GpaHwInfo asic_hw_info;
-
-        // Make sure there are available asics for AMD card.
-        // In case there is no AMD driver, we output a message.
-        if (asic_info_list.empty())
-        {
-            GPA_LOG_MESSAGE("Cannot get asicInfoList from ADL.");
-        }
-
-        for (const auto& asic_info : asic_info_list)
-        {
-            if (IsDeviceUnprofilable(asic_info.deviceID, asic_info.revID, api, driver_info))
-            {
-                continue;
-            }
-
-            asic_hw_info.SetVendorId(asic_info.vendorID);
-            asic_hw_info.SetDeviceName(asic_info.adapterName.c_str());
-            asic_hw_info.SetDeviceId(asic_info.deviceID);
-            asic_hw_info.SetRevisionId(asic_info.revID);
-            asic_hw_info.SetGpuIndex(asic_info.gpuIndex);
-            asic_hw_info.UpdateDeviceInfoBasedOnDeviceId(api, driver_info);
-
-            if (CompareHwInfo(api_hw_info, asic_hw_info))
-            {
-                GpaUInt32 api_rev_id = 0;
-
-                // Make sure revision id is set correctly based on the actual hardware.
-                if (api_hw_info.GetRevisionId(api_rev_id) && REVISION_ID_ANY == api_rev_id)
-                {
-                    api_hw_info.SetRevisionId(asic_info.revID);
-                }
-
-                api_hw_info.UpdateDeviceInfoBasedOnDeviceId(api, driver_info);
-
-                // This device matches what the application is running on, so break from the loop.
-                found_matching_hw_info = true;
-                break;
-            }
-        }
-#endif
-    }
-
-#if defined(WIN32)
-
-    if (!found_matching_hw_info)  // ADL will not be available on a clean system that has never had the AMD driver installed.
-    {
-        Adapter adapter;
-        asic_info_list.clear();
-
-        if (adapter.GetAsicInfoList(asic_info_list))
-        {
-            for (const auto& asic_info : asic_info_list)
-            {
-                GpaHwInfo asic_hw_info;
-                asic_hw_info.SetVendorId(asic_info.vendorID);
-                asic_hw_info.SetDeviceName(asic_info.adapterName.c_str());
-                asic_hw_info.SetDeviceId(asic_info.deviceID);
-                asic_hw_info.SetRevisionId(asic_info.revID);
-
-                if (kNvidiaVendorId == asic_info.vendorID)
-                {
-                    asic_hw_info.SetHwGeneration(GDT_HW_GENERATION_NVIDIA);
-                }
-                else if (kIntelVendorId == asic_info.vendorID)
-                {
-                    asic_hw_info.SetHwGeneration(GDT_HW_GENERATION_INTEL);
-                }
-                else
-                {
-                    // This call makes sure the hw generation is set correctly.
-                    asic_hw_info.UpdateDeviceInfoBasedOnDeviceId(api, driver_info);
-                }
-
-                if (CompareHwInfo(api_hw_info, asic_hw_info))
-                {
-                    GpaUInt32 api_rev_id = 0;
-
-                    // Make sure revision id is set correctly based on the actual hardware.
-                    if (api_hw_info.GetRevisionId(api_rev_id) && REVISION_ID_ANY == api_rev_id)
-                    {
-                        api_hw_info.SetRevisionId(asic_info.revID);
-                    }
-
-                    api_hw_info.UpdateDeviceInfoBasedOnDeviceId(api, driver_info);
-
-                    // This device matches what the application is running on, so break from the loop.
-                    found_matching_hw_info = true;
-                    break;
-                }
-            }
-        }
-        else
-        {
-            GPA_LOG_MESSAGE("Unable to get asic info from the Adapter.");
-        }
-    }
-
-#endif
-
-    if (!found_matching_hw_info)
-    {
-        // This code path is for systems where ADL is not available. ADL is not available on ROCm systems as well as on amdgpu systems.
-        // API specific hardware information mostly gets basic information (namely just needs to get VendorID and DeviceID), so we need to update
-        // the device info with additional information that we store per-deviceID.
         const bool device_info_ok = api_hw_info.UpdateDeviceInfoBasedOnDeviceId(api, driver_info);
 
         if (!device_info_ok)
@@ -443,14 +301,8 @@ GpaStatus GpaImplementor::IsDeviceSupported(GpaContextInfoPtr    context_info,
         }
     }
 
-    // Check we support the device generation.
-    if (!IsDeviceGenerationSupported(api_hw_info))
-    {
-        return kGpaStatusErrorHardwareNotSupported;
-    }
-
     // Give the API-specific implementation a chance to verify that the hardware is supported.
-    GpaStatus status = VerifyApiHwSupport(context_info, flags, api_hw_info) ? kGpaStatusOk : kGpaStatusErrorFailed;
+    const GpaStatus status = VerifyApiHwSupport(context_info, flags, api_hw_info) ? kGpaStatusOk : kGpaStatusErrorFailed;
 
     if (kGpaStatusOk == status)
     {
