@@ -1,5 +1,5 @@
 //==============================================================================
-// Copyright (c) 2016-2025 Advanced Micro Devices, Inc. All rights reserved.
+// Copyright (c) 2016-2026 Advanced Micro Devices, Inc. All rights reserved.
 /// @author AMD Developer Tools Team
 /// @file
 /// @brief  Macros used for defining GPUPerfAPI version.
@@ -13,16 +13,16 @@
 #define GPU_PERF_API_COMMMON_GPA_VERSION_H_
 
 #define GPA_MAJOR_VERSION  4 ///< Macro for GPA major version.
-#define GPA_MINOR_VERSION  2 ///< Macro for GPA minor version.
+#define GPA_MINOR_VERSION  3 ///< Macro for GPA minor version.
 #define GPA_UPDATE_VERSION 0 ///< Macro for GPA update version.
 #define GPA_BUILD_NUMBER   0 /// Macro for GPA build number
 
 #define GPA_MAJOR_VERSION_STR  "4" ///< Macro for major version string.
-#define GPA_MINOR_VERSION_STR  "2" ///< Macro for minor version string.
+#define GPA_MINOR_VERSION_STR  "3" ///< Macro for minor version string.
 #define GPA_BUILD_NUMBER_STR   "0" ///< Macro for build number string.
 #define GPA_UPDATE_VERSION_STR "0" ///< Macro for update version string.
 
-#define GPA_COPYRIGHT_STR "Copyright (c) 2010-2025 Advanced Micro Devices, Inc. All rights reserved." ///< Macro for copyright string.
+#define GPA_COPYRIGHT_STR "Copyright (c) 2010-2026 Advanced Micro Devices, Inc. All rights reserved." ///< Macro for copyright string.
 
 // clang-format on
 
