@@ -18,27 +18,14 @@ GpaApiType Dx11GpaSession::GetApiType() const
     return kGpaApiDirectx11;
 }
 
-GpaPass* Dx11GpaSession::CreateApiPass(PassIndex pass_index)
+std::unique_ptr<GpaPass> Dx11GpaSession::CreateApiPass(PassIndex pass_index)
 {
-    GpaPass* ret_pass = nullptr;
-
     CounterList* pass_counters = GetCountersForPass(pass_index);
     assert(pass_counters != nullptr);
     if (pass_counters != nullptr && pass_counters->size() > 0)
     {
         GpaCounterSource counter_source = GetCounterSource((*pass_counters)[0]);
-
-        Dx11GpaPass* dx11_pass = new (std::nothrow) Dx11GpaPass(this, pass_index, counter_source, pass_counters);
-
-        if (nullptr == dx11_pass)
-        {
-            GPA_LOG_ERROR("Unable to allocate memory for the pass.");
-        }
-        else
-        {
-            ret_pass = dx11_pass;
-        }
+        return std::make_unique<Dx11GpaPass>(this, pass_index, counter_source, pass_counters);
     }
-
-    return ret_pass;
+    return nullptr;
 }
