@@ -31,6 +31,6 @@ public:
 
 private:
     /// @copydoc GpaSession::CreateApiPass()
-    GpaPass* CreateApiPass(PassIndex pass_index) override;
+    std::unique_ptr<GpaPass> CreateApiPass(PassIndex pass_index) override;
 };
 #endif
